% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{add_star}
\alias{add_star}
\title{add_star}
\usage{
add_star(.mod)
}
\arguments{
\item{.mod}{list object of model data}
}
\description{
add_star
}
\section{Functions}{
\itemize{
\item \code{add_star}: tar on a model object and corresponding YAML.
}}

