% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-log.R
\name{add_aic_bic}
\alias{add_aic_bic}
\title{Calculate Akaike Information Criterion (AIC) and Bayesian Information
Criterion (BIC), and append to summary log}
\usage{
add_aic_bic(.log_df)
}
\arguments{
\item{.log_df}{The output of \code{\link[=summary_log]{summary_log()}}) or \code{run_log() \%>\% add_summary()}}
}
\description{
Calculate Akaike Information Criterion (AIC) and Bayesian Information
Criterion (BIC), and append to summary log
}
\keyword{internal}
