% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{add_bbi_args}
\alias{add_bbi_args}
\title{Add new babylon args to model}
\usage{
add_bbi_args(.mod, .bbi_args)
}
\arguments{
\item{.mod}{A `bbi_{.model_type}_model` object}

\item{.bbi_args}{named list of arguments to add to the model}
}
\description{
Modifies model object and corresponding YAML by adding new .bbi_args,
overwriting any args that are already present with the new values.
Use `print_nonmem_args()` to see a list list of valid babylon arguments.
}
