% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{add_bbi_to_path}
\alias{add_bbi_to_path}
\title{Private helper to add bbi to $PATH}
\usage{
add_bbi_to_path(.dest_bbi_path)
}
\arguments{
\item{.dest_bbi_path}{absolute path to bbi (\code{dirname(.dest_bbi_path)} will be added to the \verb{$PATH})}
}
\description{
Private helper to add bbi to $PATH
}
\keyword{internal}
