% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-log.R
\name{add_config}
\alias{add_config}
\title{Joins config log tibble onto a matching run log tibble}
\usage{
add_config(.log_df, ...)
}
\arguments{
\item{.log_df}{the output tibble from `run_log()`}

\item{...}{arguments passed through to `config_log()`}
}
\description{
Joins config log tibble onto a matching run log tibble
}
