% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-log.R
\name{add_dofv}
\alias{add_dofv}
\title{Calculate the change objective function value models in a run log}
\usage{
add_dofv(.log_df, .mod_ref = NULL)
}
\arguments{
\item{.log_df}{A \code{bbi_run_log_df} tibble (as output by \code{\link[=run_log]{run_log()}} or
\code{\link[=summary_log]{summary_log()}}).}

\item{.mod_ref}{The reference model to use when calculating the change in
objective function value for each model in the run log. By default, the
reference model is each model's parent (as identified by the first value
returned by \code{\link[=get_based_on]{get_based_on()}}). Pass a \code{bbi_nonmem_model} object to use its
objective function value as the reference value for all calculations.}
}
\value{
The \code{.log_df} data frame with a new "dofv" column. If \code{.log_df} is
not a summary log, a "ofv" column is also added for context.
}
\description{
Extend a run log with a "dofv" column that reports the difference in
objective function value between row's model and a reference model.
}
\examples{
\dontrun{

# Append to run logs:
run_log(MODEL_DIR) \%>\% add_summary() \%>\% add_dofv()
summary_log(MODEL_DIR) \%>\% add_dofv()

# Automatically adds `ofv` column in absence of `add_summary()` call:
run_log(MODEL_DIR) \%>\% add_dofv()
}


}
\seealso{
\code{\link[=add_summary]{add_summary()}}
}
