% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_model_status}
\alias{add_model_status}
\title{Add model status to a run log}
\usage{
add_model_status(.log_df)
}
\arguments{
\item{.log_df}{a \code{bbi_run_log_df} tibble (the output of \code{\link[=run_log]{run_log()}})}
}
\description{
Add model status to a run log
}
\keyword{internal}
