% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-nm-data.R
\name{add_na_filter}
\alias{add_na_filter}
\title{Adjust filters to retain NA values for relevant columns}
\usage{
add_na_filter(r_expr)
}
\arguments{
\item{r_expr}{An \code{R} filter expression. e.g., \code{'BLQ!=1'}, \code{C!='C'}.}
}
\description{
This function serves to add a  \verb{| is.na(col)} filter to an \code{R} filter to
avoid filtering out \code{NA} rows. It is \emph{not} needed when using \code{grepl} for
string detection (e.g., \verb{IGNORE=#}, \verb{IGNORE=@}).
}
\seealso{
\code{\link[=translate_nm_operator]{translate_nm_operator()}}, \code{\link[=invert_operator]{invert_operator()}}
}
\keyword{internal}
