% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimates.R
\name{add_param_shrinkage}
\alias{add_param_shrinkage}
\title{Add shrinkage details to param table}
\usage{
add_param_shrinkage(.param_df, .summary)
}
\arguments{
\item{.param_df}{The parameter estimates table containing, at minimum \code{parameter_names} and \code{diag} columns.}

\item{.summary}{A \code{bbi_nonmem_summary} object.}
}
\description{
Private helper to unpack shrinkage details and create a column on the input tibble
with the shrinkage assigned to the relevant parameters.
}
\keyword{internal}
