% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{add_simulation}
\alias{add_simulation}
\title{Add a simulation to a \code{bbi_nonmem_model} object}
\usage{
add_simulation(
  .mod,
  n = 200,
  seed = 1234,
  sim_cols = "DV",
  .join_col = "NUM",
  .suffix = "sim",
  .inherit_tags = TRUE,
  .overwrite = FALSE,
  .mode = getOption("bbr.bbi_exe_mode")
)
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} object.}

\item{n}{number of simulations/subproblems. Adds \code{SUBPROBLEMS=n} to a
\verb{$SIMULATION} record.}

\item{seed}{a seed for simulation. Appended to \verb{$SIMULATION} record.}

\item{sim_cols}{Character column name(s) defining the simulated values to
table out.}

\item{.join_col}{Character column name(s) used to join table files post
execution. Gets appended to the generated \verb{$TABLE} record. See
\code{\link[=nm_join_sim]{nm_join_sim()}} documentation for details. Defaults to \code{'NUM'}.}

\item{.inherit_tags}{If \code{TRUE}, the default, inherit any tags from \code{.mod}.}

\item{.overwrite}{If \code{FALSE}, the default,  function will error if a model
file already exists at specified \code{.new_model} path. If \code{TRUE} any existing
file at \code{.new_model} will be overwritten silently.}

\item{.mode}{Either \code{"sge"}, the default, to submit model(s) to the grid or
\code{"local"} for local execution. This can be passed directly to this argument
or set globally with \code{options("bbr.bbi_exe_mode")}.}
}
\description{
Add a simulation to a \code{bbi_nonmem_model} object
}
\details{
\code{add_simulation} does the following things:
\itemize{
\item Checks that \code{.mod} was previously executed and tabled out an \code{.MSF} file
(i.e. \verb{$EST MSFO=1.MSF}).
\itemize{
\item \strong{Note:} The \code{.MSF} file must have an upper case extension, otherwise
it will be cleaned up after submission
}
\item Performs various checks to confirm the status of \code{.mod} and contents of its
control stream.
\item Creates a new \code{bbi_nmsim_model} object with the following differences from
the original control stream:
\itemize{
\item Removes the following record types for simulation: \verb{$EST}, \verb{$COV},
\verb{$TABLE}, \verb{$SIMULATION}
\item Removes PK and prior records: \verb{$PRIOR}, \verb{$THETA/$THETAP/$THETAPV},
\verb{$OMEGA/$OMEGAP/$OMEGAPD}, \verb{$SIGMA/$SIGMAP/$SIGMAPD}
\item Adds a new custom \verb{$SIMULATION} record using user specified values (e.g.
\code{seed} and \code{n}). \code{TRUE=FINAL} is appended to ensure the final values are
used rather than the initial estimates.
\item Adds a new \verb{$TABLE} record tabling out simulated values and \code{.join_col}
column(s)
\item Adds a new \verb{$MSFI} record (run with \code{NOMSFTEST}) pointing to the \code{.MSF}
file of \code{.mod}
}
\item Creates a specification file, storing \code{seeds}, \code{n}, and other various items
helpful for traceability purposes.
\item Submits the model for execution.
}
}
\seealso{
nm_join_sim
}
