% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{add_star}
\alias{add_star}
\title{Star a model}
\value{
The modified \verb{bbi_\{.model_type\}_model} object
}
\description{
Add or remove star on a model object and corresponding YAML.
Tags can be modified at any time (i.e. before \emph{or} after a model is submitted).
}
\details{
The \code{starred} fields on a \verb{bbi_\{.model_type\}_model} object contains a boolean variable if model is starred
}
\seealso{
\code{\link[=run_log]{run_log()}} \code{\link[=collapse_to_string]{collapse_to_string()}} \code{\link[=modify_notes]{modify_notes()}} \code{\link[=modify_based_on]{modify_based_on()}} \code{\link[=modify_description]{modify_description()}} \code{\link[=modify_bbi_args]{modify_bbi_args()}}
}
