% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{add_to_path_message}
\alias{add_to_path_message}
\title{Helper to message user about adding the bbi directory to $PATH}
\usage{
add_to_path_message(.bbi_path)
}
\arguments{
\item{.bbi_path}{absolute path to the new bbi}
}
\description{
Will return invisibly if .bbi_path is the same as \code{getOption('bbr.bbi_exe_path')}
or if \code{dirname(.bbi_path)} is already in \verb{$PATH}
}
\keyword{internal}
