% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-threads.R
\name{adjust_estimation_options}
\alias{adjust_estimation_options}
\title{Parse model files and overwrite estimation iterations}
\usage{
adjust_estimation_options(
  .mods,
  .max_eval,
  .detect = c("MAXEVAL", "NITER", "NBURN", "MAX")
)
}
\arguments{
\item{.mods}{a bbi model object or list of model objects. Generally created by \code{test_threads()}.}

\item{.max_eval}{Max number of iterations for NONMEM to run.}

\item{.detect}{Estimation options that will be set to \code{.max_eval} if found.}
}
\description{
Parse model files and overwrite estimation iterations
}
\keyword{internal}
