% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak-initial-estimates.R
\name{adjust_tweaked_theta}
\alias{adjust_tweaked_theta}
\title{Adjust initial theta estimates to be within bounds}
\usage{
adjust_tweaked_theta(init_thetas, digits)
}
\arguments{
\item{init_thetas}{table of theta estimates and bounds}

\item{digits}{number of significant figures to round to}
}
\description{
Adjust initial theta estimates to be within bounds
}
\keyword{internal}
