% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-run-utils.R
\name{analysis_can_be_summarized}
\alias{analysis_can_be_summarized}
\title{Check if the analysis run (Bootstrap or SSE) can be summarized}
\usage{
analysis_can_be_summarized(.run)
}
\arguments{
\item{.run}{A \code{bbi_nmboot_model} or \code{bbi_nmsse_model} model object}
}
\description{
Check if the analysis run (Bootstrap or SSE) can be summarized
}
\keyword{internal}
