% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-run-utils.R
\name{analysis_estimates}
\alias{analysis_estimates}
\title{Tabulate parameter estimates for each model submission in an analysis run}
\usage{
analysis_estimates(.run, format_long = FALSE, force_resummarize = FALSE)
}
\arguments{
\item{.run}{A \code{bbi_nmboot_model} or \code{bbi_nmsse_model} object.}

\item{format_long}{Logical (T/F). If \code{TRUE}, format data as a long table,
making the data more portable for plotting.}

\item{force_resummarize}{Logical (T/F). If \code{TRUE}, force re-summarization.
Will \emph{only} update the saved out \code{RDS} file when specified via
\code{summarize_bootstrap_run()}. See details for more information.}
}
\description{
Tabulate parameter estimates for each model submission in an analysis run
}
\keyword{internal}
