% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-labels.R
\name{apply_indices}
\alias{apply_indices}
\title{Add parameter indices to a label tibble}
\usage{
apply_indices(.label_df)
}
\arguments{
\item{.label_df}{A tibble like the output of `param_labels()`, containing columns `names, label, unit, type`}
}
\description{
Add parameter indices to a label tibble
}
