% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{as_model}
\alias{as_model}
\alias{as_model.bbi_nonmem_model}
\alias{as_model.babylon_process}
\title{Convert object to \verb{bbi_\{.model_type\}_model}}
\usage{
as_model(.obj)

\method{as_model}{bbi_nonmem_model}(.obj)

\method{as_model}{babylon_process}(.obj)
}
\arguments{
\item{.obj}{Object to convert to \verb{bbi_\{.model_type\}_model}}
}
\description{
Convert object to \verb{bbi_\{.model_type\}_model}
}
\section{Methods (by class)}{
\itemize{
\item \code{bbi_nonmem_model}: Simply passes through the \code{bbi_nonmem_model} object.

\item \code{babylon_process}: Takes a \code{babylon_process} object and converts it to the corresponding \code{bbi_nonmem_model} object.
Only works if YAML and model file are in the same directory with the same name and different file extensions.
}}

