% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbr.R
\name{assert_bbi_version}
\alias{assert_bbi_version}
\alias{test_bbi_version}
\title{Check that bbi satisfies a version constraint}
\usage{
assert_bbi_version(
  .bbi_exe_path = getOption("bbr.bbi_exe_path"),
  .min_version = getOption("bbr.bbi_min_version"),
  .function = NULL
)

test_bbi_version(
  .bbi_exe_path = getOption("bbr.bbi_exe_path"),
  .min_version = getOption("bbr.bbi_min_version")
)
}
\arguments{
\item{.bbi_exe_path}{Path to bbi executable}

\item{.min_version}{The minimum allowed version. Defaults to
\code{getOption("bbr.bbi_min_version")}.}

\item{.function}{The relevant function that is version-constrained. If not
\code{NULL}, notifies user in error message.}
}
\value{
\itemize{
\item \code{assert_bbi_version()} returns the value of \code{.bbi_exe_path}.
\item \code{test_bbi_version()} returns \code{TRUE} if the installed bbi satisfies the
constraint and \code{FALSE} otherwise.
}
}
\description{
Check that bbi satisfies a version constraint
}
\keyword{internal}
