% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-file.R
\name{assert_nm_table_format}
\alias{assert_nm_table_format}
\title{Check if a \code{NONMEM} table file contains one or more tables in the
specified format}
\usage{
assert_nm_table_format(.path, table_pattern = "^TABLE NO")
}
\arguments{
\item{.path}{a path to a table file.}

\item{table_pattern}{character string defining the start of a new
table (regex accepted).}
}
\value{
logical
}
\description{
Check if a \code{NONMEM} table file contains one or more tables in the
specified format
}
\seealso{
nm_file_multi_table
}
\keyword{internal}
