% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_model_class}
\alias{assign_model_class}
\alias{assign_process_class}
\title{Assign `bbi_{.model_type}_model` class}
\usage{
assign_model_class(.mod_list, .model_type)

assign_process_class(res)
}
\arguments{
\item{.mod_list}{List to attempt to assign the class to}

\item{.model_type}{Type of model.}

\item{res}{List to attempt to assign the class to}
}
\description{
First checks to be sure the required keys are all present

First checks to be sure the required keys are all present
}
