% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_result_class}
\alias{assign_result_class}
\title{Checks that object (list) has all the required elements to become an S3 object of class `bbi_{.model_type}_result` and then, if it does, assigns the class.}
\usage{
assign_result_class(.mod_list, .model_type)
}
\arguments{
\item{.mod_list}{The candidate list to assign the class to}

\item{.model_type}{The type of model}
}
\description{
Checks that object (list) has all the required elements to become an S3 object of class `bbi_{.model_type}_result` and then, if it does, assigns the class.
}
