% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{bbi_current_release}
\alias{bbi_current_release}
\title{Get version number of bbi current release}
\usage{
bbi_current_release(.bbi_url = NULL)
}
\arguments{
\item{.bbi_url}{(Optional) URL for a bbi release artifact to strip version
number out of. If \code{NULL}, the default, will fetch the URL with
\code{current_release_url()}.}
}
\description{
Helper function to get version number of most recent release of bbi from GitHub.
}
