% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{bbi_dry_run}
\alias{bbi_dry_run}
\title{Babylon dry_run}
\usage{
bbi_dry_run(.cmd_args, .dir)
}
\description{
Creates a `babylon_process` object with all the required keys, without actually running the command
}
