% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{bbi_help}
\alias{bbi_help}
\title{Executes (`bbi --help`) with processx::run and prints the output string}
\usage{
bbi_help()
}
\description{
Executes (`bbi --help`) with processx::run and prints the output string
}
