% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{bbi_init}
\alias{bbi_init}
\title{Initialize babylon}
\usage{
bbi_init(
  .dir,
  .nonmem_dir,
  .nonmem_version = NULL,
  .no_default_version = FALSE
)
}
\arguments{
\item{.dir}{Path to directory to run \code{init} in (and put the resulting \code{babylon.yml} file)}

\item{.nonmem_dir}{Path to directory with the NONMEM installation.}

\item{.nonmem_version}{Character scalar for default version of NONMEM to use. If left NULL, function will exit and tell you which versions were found in \code{.nonmem_dir}}

\item{.no_default_version}{If \code{TRUE}, force creation of babylon.yaml with \strong{no default NONMEM version}. \code{FALSE} by default, and using \code{TRUE} is \emph{not} encouraged.}
}
\description{
Executes \verb{bbi init ...} in specified directory. This creates a \code{babylon.yml} file, which contains defaults
for many configurable \code{babylon} settings, in that directory.
}
\details{
For \code{rbabylon} to make any calls out to \code{bbi} (for example in \code{submit_model()} or \code{model_summary()}) it must find a
\code{babylon.yml} file in one of the following places:
\itemize{
\item The directory specified in \code{options("rbabylon.model_directory")}
\item The working directory, if \code{options("rbabylon.model_directory")} is \code{NULL}
\item A path passed to the \code{.config_path} argument of the functions mentioned above
}

The recommended behavior is to set \code{options("rbabylon.model_directory")}, ideally in your \code{.Rprofile},
and then call \code{bbi_init(.dir = getOption("rbabylon.model_directory"), ...)}. This only has to be done once.
}
