% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{bbi_version}
\alias{bbi_version}
\title{Get version of installed babylon}
\usage{
bbi_version(.bbi_exe_path = getOption("rbabylon.bbi_exe_path"))
}
\arguments{
\item{.bbi_exe_path}{Path to babylon executable}
}
\value{
String giving the version of the bbi binary installed at \code{.bbi_exe_path}
}
\description{
Get version of installed babylon
}
\examples{
\dontrun{
bbi_version()
}
}
