% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\docType{data}
\name{BBR_ESTIMATES_INHERIT}
\alias{BBR_ESTIMATES_INHERIT}
\title{Supported estimates to inherit from parent model}
\format{
An object of class \code{character} of length 3.
}
\usage{
BBR_ESTIMATES_INHERIT
}
\description{
Supported estimates to inherit from parent model
}
\keyword{datasets}
