% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap a model}
\usage{
bootstrap(
  .parent_mod,
  .bootstrap_start,
  .bootstrap_count,
  .key_col,
  .sample_rate
)
}
\arguments{
\item{.parent_mod}{A `bbi_{.model_type}_model` object to bootstrap}

\item{.bootstrap_start}{Integer to use for the name of the first bootstrapped model}

\item{.bootstrap_count}{Number of bootstraps to perform. For example if `.bootstrap_start = 100` and `.bootstrap_count = 50` then 50 models, named 100-149 will be created.}

\item{.key_col}{Character scaler specifying the key column to sample on.}

\item{.sample_rate}{Number between 0 and 1 specifying the size of each resample. For example `.sample_rate = 0.75` will sample 75% of the data for each bootstrap.}
}
\description{
Resamples data set a specified number of times and creates new models that are copies of `.parent_mod` but point to a resampled data set
}
