% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{bootstrap_model}
\alias{bootstrap_model}
\title{Create a boostrap model object from an existing model}
\usage{
bootstrap_model(
  .mod,
  n = 100,
  strat_cols = NULL,
  replace = TRUE,
  seed = 1234,
  overwrite = FALSE
)
}
\arguments{
\item{.mod}{a \code{bbr} model object}

\item{n}{number of unique sampled keys, defaults to match dataset}

\item{strat_cols}{columns to maintain proportion for stratification}

\item{replace}{whether to stratify with replacement}

\item{seed}{a seed for sampling the data}

\item{overwrite}{Logical to specify whether or not to overwrite existing
model output from a previous run.}
}
\description{
Create a boostrap model object from an existing model
}
