% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{build_bbi_install_path}
\alias{build_bbi_install_path}
\title{Build default path to install bbi to}
\usage{
build_bbi_install_path()
}
\description{
Build default path to install bbi to
}
\keyword{internal}
