% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{build_bbi_param_list}
\alias{build_bbi_param_list}
\title{Group models for bbi submission}
\usage{
build_bbi_param_list(.mods, .bbi_args = NULL)
}
\arguments{
\item{.mods}{A list of S3 objects of class \code{bbi_nonmem_model}}

\item{.bbi_args}{A named list specifying arguments to pass to babylon formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}. Run \code{\link[=print_bbi_args]{print_bbi_args()}} to see valid arguments.}
}
\value{
A list whose elements correspond to groups of models that can be
submitted in a single bbi call. Each element is itself a list with elements
\itemize{
\item \code{bbi_args}: a set of CLI arguments
\item \code{models}: a list of the elements of \code{.mods} in the group
}
}
\description{
Multiple models can be submitted in a single bbi call if those models share a
working directory and a set of CLI arguments.
}
\keyword{internal}
