% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-data.R
\name{build_data}
\alias{build_data}
\alias{build_data.bbi_stan_model}
\title{Build model input data}
\usage{
build_data(.mod, .out_path = NULL, ...)

\method{build_data}{bbi_stan_model}(.mod, .out_path = NULL, ...)
}
\arguments{
\item{.mod}{a \verb{bbi_\{.model_type\}_model} object.}

\item{.out_path}{If \code{NULL}, the default, does not write any data to disk.
Otherwise, pass a file path where the resulting data object should be
written.}

\item{...}{Arguments passed through to methods (currently none).}
}
\value{
Invisibly returns the data object returned from the relevant build function.
Also optionally saves the object to disk if a valid file path is passed to
\code{.out_path}.

The data object is returned invisibly because sometimes you will want to call
this function solely for the side effect of writing the data to disk.
}
\description{
Some types of models carry around code that builds the input data for the
modeling, instead of relying on a single static file on disk. When necessary,
\strong{this function is called internally by \code{\link[=submit_model]{submit_model()}}} to build the data
for running the model, but they can also be called manually by users for
other purposes.
}
\details{
\strong{Currently, only \code{bbi_stan_model} objects have this implemented.}
The \verb{<run>-standata.R} file contains the code for building the Stan input
data.
}
\section{Methods (by class)}{
\itemize{
\item \code{bbi_stan_model}: executes the function in \verb{<run>-standata.R} and
returns the list that will be passed to the \code{data} argument of
\code{cmdstanr::sample()}. Also optionally writes the returned data list to json
with \code{\link[cmdstanr:write_stan_json]{cmdstanr::write_stan_json()}} if an \code{.out_path} is passed.
}}

