% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-labels.R
\name{build_matrix_indices}
\alias{build_matrix_indices}
\title{Builds matrix indices labels}
\usage{
build_matrix_indices(.is_diag)
}
\arguments{
\item{.is_diag}{Boolean vector denoting whether each element is a diagonal}
}
\description{
Builds matrix indices labels
}
