% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{build_path_from_model}
\alias{build_path_from_model}
\title{Builds the absolute path to a file in the output directory from components of the \verb{bbi_\{.model_type\}_model} object}
\usage{
build_path_from_model(.mod, .extension)
}
\arguments{
\item{.mod}{Model to use, either a \verb{bbi_\{.model_type\}_model} or \verb{bbi_\{.model_type\}_summary} object.}

\item{.extension}{file extension to append (for example \code{lst}, \code{ext}, \code{grd}, etc.)}
}
\description{
Builds the absolute path to a file in the output directory from components of the \verb{bbi_\{.model_type\}_model} object
}
