% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan.R
\name{build_stan_bbi_config}
\alias{build_stan_bbi_config}
\title{Build bbi_config.json for Stan models}
\usage{
build_stan_bbi_config(.mod, .write)
}
\arguments{
\item{.mod}{a \code{bbi_stan_model} object}

\item{.write}{Logical scalar for whether to write the resulting list to
a file at \code{file.path(get_output_dir(.mod), "bbi_config.json")}}
}
\description{
Contains information, including hashes and configuration,
for successfully run models.
}
\keyword{internal}
