% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-log.R
\name{calc_dofv_impl}
\alias{calc_dofv_impl}
\title{Implementation function for calculating the difference in objective function
value (OFV).}
\usage{
calc_dofv_impl(.mod, .mod2 = NULL)
}
\arguments{
\item{.mod}{The \code{bbi_nonmem_model} to calculate \code{dofv} for.}

\item{.mod2}{If a \code{bbi_nonmem_model} object is passed, calculate the
difference in objective function value (\code{dofv}) between \code{.mod} and \code{.mod2}.
If \code{.mod2 = NULL}, the default, calculate the difference between \code{.mod} and
the model at \code{get_based_on(.mod)}.}
}
\description{
Implementation function for calculating the difference in objective function
value (OFV).
}
\keyword{internal}
