% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-handling.R
\name{check_and_modify_pd}
\alias{check_and_modify_pd}
\title{Check and modify a matrix for positive-definiteness}
\usage{
check_and_modify_pd(mat, corr = FALSE, digits = 3)
}
\arguments{
\item{mat}{a symmetric matrix, or a matrix with only lower triangular values
specified (assumed to be symmetric).}

\item{corr}{logical indicating if the matrix should be a correlation matrix.}

\item{digits}{Number of significant digits to round estimates to.}
}
\description{
Check and modify a matrix for positive-definiteness
}
\keyword{internal}
