% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_bbi_run_log_df_object}
\alias{check_bbi_run_log_df_object}
\title{Private helper to check if an object inherits a run log class and error if not}
\usage{
check_bbi_run_log_df_object(.df)
}
\arguments{
\item{.df}{The object to check}
}
\description{
Private helper to check if an object inherits a run log class and error if not
}
\keyword{internal}
