% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbr.R
\name{check_bbi_version_constraint}
\alias{check_bbi_version_constraint}
\title{Check that bbi satisfies a version constraint}
\usage{
check_bbi_version_constraint(
  .bbi_exe_path = getOption("rbabylon.bbi_exe_path")
)
}
\arguments{
\item{.bbi_exe_path}{Path to bbi executable}
}
\value{
\code{NULL} if \code{.bbi_exe_path} satisfies the constraint
}
\description{
Check that bbi satisfies a version constraint
}
\keyword{internal}
