% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-cor.R
\name{check_cor_threshold}
\alias{check_cor_threshold}
\title{Check if any correlations are over threshold and warn if so}
\usage{
check_cor_threshold(cor_mat, threshold)
}
\description{
Check if any correlations are over threshold and warn if so
}
\keyword{internal}
