% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_file}
\alias{check_file}
\title{Reads the head and tail of specified file and prints it the console and/or returns it as a character vector.}
\usage{
check_file(.file, .head = 3, .tail = 5, .print = TRUE, .return = FALSE)
}
\arguments{
\item{.file}{Character scaler of path to file to read}

\item{.head}{Integer for number of lines to read from the top of the file}

\item{.tail}{Integer for number of lines to read from the bottom of the file}

\item{.print}{Boolean for whether to print resulting head and tail to console. Defaults to TRUE.}

\item{.return}{Boolean for whether to return resulting head and tail as a character vector. Defaults to FALSE.}
}
\description{
Reads the head and tail of specified file and prints it the console and/or returns it as a character vector.
}
