% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{check_for_existing_model}
\alias{check_for_existing_model}
\title{Private helper to look for existing model and overwrite if necessary}
\usage{
check_for_existing_model(.path, .overwrite)
}
\arguments{
\item{.path}{Path to save the new model. Will be the path to the model file
and YAML file (both without extension), and the path to the output
directory.}

\item{.overwrite}{If \code{FALSE}, the default, error if a file already exists at
\code{.yaml_path}. If \code{TRUE} overwrite existing file and output directory, if they exist.}
}
\description{
Private helper to look for existing model and overwrite if necessary
}
\keyword{internal}
