% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{check_for_model_tree_pkgs}
\alias{check_for_model_tree_pkgs}
\title{Checks if all packages needed for \code{\link[=model_tree]{model_tree()}} are present}
\usage{
check_for_model_tree_pkgs()
}
\description{
Returns a vector with the missing packages, or returns NULL if all are
present.
}
\keyword{internal}
