% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{check_lst_file}
\alias{check_lst_file}
\title{Helper function to look for .lst function in a directory}
\usage{
check_lst_file(.x)
}
\description{
Helper function to look for .lst function in a directory
}
