% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{check_mode_argument}
\alias{check_mode_argument}
\title{Private helper to check if \code{.mode} arg to \code{submit_model()} is valid}
\usage{
check_mode_argument(.mode)
}
\description{
Private helper to check if \code{.mode} arg to \code{submit_model()} is valid
}
