% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{check_model_for_sim}
\alias{check_model_for_sim}
\title{Check that a \code{bbr} model can be used for simulation}
\usage{
check_model_for_sim(.mod, replicate_var = "REPI")
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} object.}

\item{replicate_var}{a variable name for storing each iteration. Cannot match
a variable previously defined in \verb{$PRED} or \verb{$ERROR}.}
}
\value{
\code{TRUE} invisibly
}
\description{
Checks that a \code{bbr} model object is in the correct format to generate a
\code{bbi_nmsim_model} object.
}
\keyword{internal}
