% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{check_model_tree}
\alias{check_model_tree}
\title{Perform checks regarding whether a tree can be made}
\usage{
check_model_tree(.network_df)
}
\description{
Perform checks regarding whether a tree can be made
}
