% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-model-up-to-date.R
\name{check_model_up_to_date}
\alias{check_model_up_to_date}
\alias{check_model_up_to_date.bbi_nonmem_model}
\alias{check_model_up_to_date.bbi_nonmem_summary}
\alias{check_model_up_to_date.bbi_stan_model}
\alias{check_model_up_to_date.bbi_stan_summary}
\title{Check model and data up to date with outputs}
\usage{
check_model_up_to_date(.mod, .build_data, ...)

\method{check_model_up_to_date}{bbi_nonmem_model}(.mod, .build_data = FALSE, ...)

\method{check_model_up_to_date}{bbi_nonmem_summary}(.mod, .build_data = FALSE, ...)

\method{check_model_up_to_date}{bbi_stan_model}(.mod, .build_data = TRUE, ...)

\method{check_model_up_to_date}{bbi_stan_summary}(.mod, .build_data = TRUE, ...)
}
\arguments{
\item{.mod}{a \verb{bbi_\{.model_type\}_model} or \verb{bbi_\{.model_type\}_summary} object}

\item{.build_data}{\strong{Only relevant to Stan models.} If \code{TRUE}, the default,
run \code{-standata.R} and save the output to a temp file and check the hash of
\emph{the temp file} against the \code{bbi_config.json} hash. This option actually
runs the code and, importantly, verifies that the input data to
\code{-standata.R} has not changed either. If \code{FALSE}, check the hashes of the
\code{-standata.R} and \code{-standata.json} against the hashes in \code{bbi_config.json}
but do \emph{not} run the \code{-standata.R} script.}

\item{...}{Arguments passed through (currently none).}
}
\value{
A (named) logical vector of length 2. The first element (named
\code{"model"}) refers to the model files mentioned above. The second element
(named \code{"data"}) refers to the data files mentioned above. For both
elements, they will be \code{TRUE} if nothing has changed, \code{FALSE} if anything
has changed.
}
\description{
Functions for checking that the model outputs on disk match the md5 hashes
stored in \code{bbi_config.json} at run time. In other words, checking that the
model and data files have \emph{not} changed since the model was last run. See
Details section for what specific files are checked.
}
\details{
Different files are checked depending on what type of model is being checked.

\strong{For NONMEM models}
\itemize{
\item The model file (control stream)
\item The data file (referenced in \verb{$DATA} within the control stream)
}

\strong{For Stan models}
\itemize{
\item The model files:
\itemize{
\item <run>.stan
\item <run>-init.R
\item <run>-stanargs.R
}
\item The data files (see \code{.build_data} argument):
\itemize{
\item <run>-standata.R
\item <run>-standata.json
}
}
}
