% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_nonmem_args}
\alias{check_nonmem_args}
\title{Checks that all passed NONMEM command line args are valid and formats}
\usage{
check_nonmem_args(.args)
}
\arguments{
\item{.args}{A named list of .args to check}
}
\value{
character string, output from format_cmd_args()
}
\description{
Checks that all passed NONMEM command line args are valid and formats
}
