% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_nonmem_output_file}
\alias{check_nonmem_output_file}
\title{Checks nonmem output file that's a whitespace-delimited file (for instance .grd or .ext)}
\usage{
check_nonmem_output_file(.path, .plot = FALSE, .iter_floor = 0)
}
\arguments{
\item{.path}{Character scalar path to the gradient file}

\item{.plot}{Boolean for whether to return a plot instead of a tibble}

\item{.iter_floor}{Filters gradient file to only iterations GREATER THAN this value. Default is 0.}
}
\description{
Checks nonmem output file that's a whitespace-delimited file (for instance .grd or .ext)
}
