% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_nonmem_progress}
\alias{check_nonmem_progress}
\title{Check the progress of a NONMEM run from the bbi_nonmem_result object}
\usage{
check_nonmem_progress(.res, .ext_wait = 30)
}
\arguments{
\item{.res}{`bbi_nonmem_summary` object to check on}

\item{.ext_wait}{Integer number of seconds to wait for an .ext file to be there before exiting with FALSE}
}
\description{
Check the progress of a NONMEM run from the bbi_nonmem_result object
}
