% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_nonmem_table_bbi}
\alias{check_nonmem_table_bbi}
\title{Private helper to pull a NONMEM table from a model}
\usage{
check_nonmem_table_bbi(.mod, .iter_floor, .extension)
}
\description{
Private helper to pull a NONMEM table from a model
}
\keyword{internal}
