% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_nonmem_table_output}
\alias{check_nonmem_table_output}
\alias{check_grd}
\alias{check_grd.character}
\alias{check_grd.bbi_nonmem_model}
\alias{check_ext}
\alias{check_ext.character}
\alias{check_ext.bbi_nonmem_model}
\title{Checks nonmem output file that's a whitespace-delimited file (for instance .grd or .ext)}
\usage{
check_nonmem_table_output(.path, .x_var = NULL, .x_floor = NULL)

check_grd(.x, ...)

\method{check_grd}{character}(.path, .iter_floor = 0)

\method{check_grd}{bbi_nonmem_model}(.mod, .iter_floor = 0)

check_ext(.x, ...)

\method{check_ext}{character}(.path, .iter_floor = 0)

\method{check_ext}{bbi_nonmem_model}(.mod, .iter_floor = 0)
}
\arguments{
\item{.path}{path to ext file}

\item{.x_var}{name of variable to filter with `.x_floor`}

\item{.x_floor}{Filters file to only rows with `.x_var` GREATER THAN this value.}

\item{.x}{generic .x}

\item{...}{args passed through}

\item{.iter_floor}{Filters file to only rows with `ITERATION` GREATER THAN this value.}

\item{.mod}{`bbi_nonmem_model` object}
}
\description{
Checks nonmem output file that's a whitespace-delimited file (for instance .grd or .ext)

S3 generic for checking grd file

S3 dispatch for checking grd file

S3 dispatch for checking grd file

S3 generic for checking ext file

S3 dispatch for checking ext file

S3 dispatch for checking ext file
}
