% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_os}
\alias{check_os}
\title{Check what operating system R is running on}
\usage{
check_os()
}
\value{
String, either "linux", "darwin", "windows"
}
\description{
Check what operating system R is running on
}
\keyword{internal}
