% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_output_dir}
\alias{check_output_dir}
\alias{check_output_dir.character}
\alias{check_output_dir.bbi_nonmem_model}
\alias{check_output_dir.bbi_nonmem_summary}
\alias{check_output_dir.bbi_stan_model}
\alias{check_output_dir.bbi_stan_summary}
\title{Check output directory}
\usage{
check_output_dir(.mod, ...)

\method{check_output_dir}{character}(.mod, ...)

\method{check_output_dir}{bbi_nonmem_model}(.mod, ...)

\method{check_output_dir}{bbi_nonmem_summary}(.mod, ...)

\method{check_output_dir}{bbi_stan_model}(.mod, ...)

\method{check_output_dir}{bbi_stan_summary}(.mod, ...)
}
\arguments{
\item{.mod}{Either a \verb{bbi_\{.model_type\}_model} object or a file path to an output directory}

\item{...}{values to pass to \code{fs::dir_ls}}
}
\description{
List files in the output directory to glance at where the process is
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Takes a file path to the output directory.

\item \code{bbi_nonmem_model}: Takes a \code{bbi_nonmem_model} object.

\item \code{bbi_nonmem_summary}: Takes a \code{bbi_nonmem_summary} object.

\item \code{bbi_stan_model}: Takes a \code{bbi_stan_model} object.

\item \code{bbi_stan_summary}: Takes a \code{bbi_stan_summary} object.
}}

