% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_output_dir_bbi}
\alias{check_output_dir_bbi}
\title{Implementation function for check_output_dir}
\usage{
check_output_dir_bbi(.mod, ...)
}
\description{
Implementation function for check_output_dir
}
\keyword{internal}
