% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{check_record_replacements}
\alias{check_record_replacements}
\title{Check that the replacement values are the same length as the record}
\usage{
check_record_replacements(val_recs, new_values)
}
\arguments{
\item{val_recs}{list of \code{nmrec} record options.}

\item{new_values}{vector of replacement values. Should be the same length as
\code{val_recs}}
}
\description{
Check that the replacement values are the same length as the record
}
\keyword{internal}
