% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-threads.R
\name{check_record_val}
\alias{check_record_val}
\title{Check if a record option matches a specific value}
\usage{
check_record_val(.record, .name, .value)
}
\arguments{
\item{.record}{an \code{nmrec} record object}

\item{.name}{Name of option to select. Any valid spelling of the option name
is allowed.}

\item{.value}{Value of option to check for. For value options, the specified value is anything
that can be formatted to a string (i.e. \code{10} in \code{NBURN=10}). For included flag options, the
default value is \code{TRUE} (\code{NOABORT}).}
}
\description{
Check if a record option matches a specific value
}
\keyword{internal}
