% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan.R
\name{check_stan_model}
\alias{check_stan_model}
\title{Checks a \code{bbi_stan_model} for necessary files}
\usage{
check_stan_model(.mod, .error = FALSE)
}
\arguments{
\item{.mod}{A \code{bbi_stan_model} object}

\item{.error}{If \code{FALSE}, the default, will warn if any necessary files are missing.
If \code{TRUE} will error instead.}
}
\description{
Checks a \code{bbi_stan_model} object to make sure it can find
all of the files necesssary to submit the model.
}
\details{
Will look for the following:
\itemize{
\item \code{build_path_from_model(.mod, "-standata.R")}
\item \code{build_path_from_model(.mod, "-init.R")}
\item \code{build_path_from_model(.mod, ".stan")}
}
}
