% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_status_code}
\alias{check_status_code}
\title{Checks status code from processx::run output}
\usage{
check_status_code(.output)
}
\arguments{
\item{.path}{Full path to a model that has been run that user wants to summarize}

\item{.model_type}{Type of model to summarize. Currently only supports "nonmem"}
}
\description{
Checks status code from processx::run output
}
