% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_summary_object}
\alias{check_summary_object}
\title{Private helper to check if an object inherits a model summary class and error if not}
\usage{
check_summary_object(.s)
}
\arguments{
\item{.s}{The object to check}
}
\description{
Private helper to check if an object inherits a model summary class and error if not
}
\keyword{internal}
