% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-threads.R
\name{check_threads}
\alias{check_threads}
\title{Check estimation time for models run with various threads values}
\usage{
check_threads(mods)
}
\arguments{
\item{mods}{list of bbi model objects created by \code{test_threads()}}
}
\value{
A tibble with columns \code{threads} (number of threads) and \code{time}
(elapsed estimation time in seconds for test models).
}
\description{
Check estimation time for models run with various threads values
}
