% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-up-to-date.R
\name{check_up_to_date}
\alias{check_up_to_date}
\alias{check_up_to_date.bbi_nonmem_model}
\alias{check_up_to_date.bbi_nonmem_summary}
\title{Check model and data up to date with outputs}
\usage{
check_up_to_date(.mod, ...)

\method{check_up_to_date}{bbi_nonmem_model}(.mod, ...)

\method{check_up_to_date}{bbi_nonmem_summary}(.mod, ...)
}
\arguments{
\item{.mod}{a \verb{bbi_\{.model_type\}_model} or \verb{bbi_\{.model_type\}_summary} object}

\item{...}{Arguments passed through (currently none).}
}
\value{
A (named) logical vector of length 2. The first element (named
\code{"model"}) refers to the model files mentioned above. The second element
(named \code{"data"}) refers to the data files mentioned above. For both
elements, they will be \code{TRUE} if nothing has changed, \code{FALSE} if anything
has changed.
}
\description{
Functions for checking that the model outputs on disk match the md5 hashes
stored in \code{bbi_config.json} at run time. In other words, checking that the
model and data files have \emph{not} changed since the model was last run. See
Details section for what specific files are checked.
}
\details{
Different files are checked depending on what type of model is being checked.

\strong{For NONMEM models}
\itemize{
\item The model file (control stream)
\item The data file (referenced in \verb{$DATA} within the control stream)
}

\strong{Currently only NONMEM implemented.}
}
