% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-up-to-date.R
\name{check_up_to_date_analysis}
\alias{check_up_to_date_analysis}
\title{Private implementation to check that NONMEM analysis (bootstrap or SSE) is
up-to-date}
\usage{
check_up_to_date_analysis(.bbi_object)
}
\arguments{
\item{.bbi_object}{the object to check. Could be
a \verb{bbi_\{.model_type\}_model} object,
a \verb{bbi_\{.model_type\}_summary} object,
or a \code{bbi_log_df} tibble.}
}
\description{
Private implementation to check that NONMEM analysis (bootstrap or SSE) is
up-to-date
}
\keyword{internal}
