% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-up-to-date.R
\name{check_up_to_date_nonmem}
\alias{check_up_to_date_nonmem}
\title{Private implementation to check that NONMEM model is up-to-date}
\usage{
check_up_to_date_nonmem(.mod)
}
\description{
Specifically, check that control stream and data file on disk have not
changed since the model was run. This is accomplished by taking their md5
hashes and comparing it to the hashes stored in \code{bbi_config.json}.
}
\keyword{internal}
