% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-labels.R
\name{clean_ctl}
\alias{clean_ctl}
\title{Parse control stream to a named list of the blocks}
\usage{
clean_ctl(x)
}
\arguments{
\item{x}{control stream as a character scaler with lines separated by newline characters}
}
\description{
Taken directly from tidynm package.
}
