% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-run-utils.R
\name{cleanup_analysis_run}
\alias{cleanup_analysis_run}
\title{Cleanup analysis run directory}
\usage{
cleanup_analysis_run(.run, .force = FALSE)
}
\arguments{
\item{.run}{A \code{bbi_nmboot_model} or \code{bbi_nmsse_model} object.}

\item{.force}{logical (T/F). If \code{TRUE}, do not prompt the user if they want to delete the models.}
}
\description{
This will delete all child models, and only keep the information
you need to read in estimates or summary information
}
\details{
The intent of this function is to help reduce the number of files you need to
commit via version control. Collaborators will be able to read in the
analysis model and summary objects without needing individual run files.
\itemize{
\item Note that this will prevent \code{force_resummarize = TRUE} from working
}

\strong{This should only be done} if you no longer need to re-summarize, as this
will clean up (delete) the \emph{individual} analysis model files
}
\keyword{internal}
