% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-threads.R
\name{cleanup_mods}
\alias{cleanup_mods}
\title{Remove model files created via \code{\link[=test_threads]{test_threads()}}}
\usage{
cleanup_mods(mods)
}
\arguments{
\item{mods}{list of bbi model objects created by \code{test_threads()}}
}
\description{
Remove model files created via \code{\link[=test_threads]{test_threads()}}
}
