% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{cli_install_commands}
\alias{cli_install_commands}
\title{Private helper function for building commands to install bbi on Linux or Unix}
\usage{
cli_install_commands(.path, .bbi_url)
}
\arguments{
\item{.path}{Absolute path to install bbi to}

\item{.bbi_url}{Full url to download tarball from}
}
\description{
Private helper function for building commands to install bbi on Linux or Unix
}
\keyword{internal}
