% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{color_model_type}
\alias{color_model_type}
\title{Function to color the model type.}
\usage{
color_model_type(.mod)
}
\arguments{
\item{.mod}{a \code{bbi_base_model} object}
}
\description{
Function to color the model type.
}
\keyword{internal}
