% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{color_status}
\alias{color_status}
\title{Function to color the status. Green for finished; red otherwise.}
\usage{
color_status(status)
}
\arguments{
\item{status}{Character string. Either \code{'Finished Running'}, \code{'Incomplete Run'},
or \code{'Not Run'}.}
}
\description{
Function to color the status. Green for finished; red otherwise.
}
\keyword{internal}
