% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_directory_path}
\alias{combine_directory_path}
\title{Combine a directory and path}
\usage{
combine_directory_path(.directory, .path)
}
\arguments{
\item{.directory}{Character scaler for the directory}

\item{.path}{Character scaler for the path to the file (could be just the file name, or could be in a subdirectory)}
}
\description{
Concatenates a directory and path, with some specific rules:
If .directory is NULL, set to working directory.
If .path is an absolute path, ignore .directory and just return .path,
otherwise return `file.path(.directory, .path)`
}
