% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_list_objects}
\alias{combine_list_objects}
\title{Combines two named lists. Shared keys will have their contents concatenated.
Any S3 classes will be inherited with those from .new_list given precedence.}
\usage{
combine_list_objects(.new_list, .old_list)
}
\arguments{
\item{.new_list}{A named list of arguments for bbi, passed into submit_model function call}

\item{.old_list}{A named list of arguments for bbi, parsed from user input yaml}
}
\value{
The combined list
}
\description{
Combines two named lists. Shared keys will have their contents concatenated.
Any S3 classes will be inherited with those from .new_list given precedence.
}
