% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-nmtran.R
\name{compare_nmtran}
\alias{compare_nmtran}
\title{Compare different NONMEM control stream configurations.}
\usage{
compare_nmtran(.mod, .mod_compare, .config_path = NULL, nmtran_exe = NULL)
}
\description{
Runs \code{run_nmtran()} on two models and compares the output, denoting whether
they evaluate to the same model via \code{NMTRAN}.
}
\details{
Say you wanted to test whether diagonal matrices could specify standard
deviation for one value, and variance for another

The \strong{reference model} would have this block:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{$OMEGA
0.05 STANDARD   ; iiv CL
0.2     ; iiv V2
}\if{html}{\out{</div>}}

The \strong{new model} would have this block:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{$OMEGA
0.05 STANDARD   ; iiv CL
0.2 VAR    ; iiv V2
}\if{html}{\out{</div>}}

Comparing the two (see below), we find no differences. This means that adding
\code{VAR} to the second ETA value had no impact, and the two models would evaluate
the same.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{> compare_nmtran(MOD1, MOD_COMPARE)
Running NMTRAN with NONMEM version `nm75`

No differences found
character(0)
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
# Starting model - set a reference
open_model_file(MOD1)

# Make new model
MOD_COMPARE <- copy_model_from(MOD1)

# Make a change
open_model_file(MOD_COMPARE)

# Compare NMTRAN evaluation
compare_nmtran(MOD1, MOD_COMPARE)

# delete new model at the end
delete_models(MOD_COMPARE, .tags = NULL, .force = TRUE)
}

}
\keyword{internal}
