% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan.R
\name{compile_stanmod}
\alias{compile_stanmod}
\title{Private helper to compile a stan model and save a gitignore that ignores the binary}
\usage{
compile_stanmod(.mod)
}
\description{
Private helper to compile a stan model and save a gitignore that ignores the binary
}
