% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-log.R
\name{config_log}
\alias{config_log}
\alias{add_config}
\title{Parse babylon configs to log}
\usage{
config_log(.base_dir = get_model_directory(), .recurse = TRUE)

add_config(.log_df)
}
\arguments{
\item{.base_dir}{Base directory to look in for models. Defaults to \code{\link[=get_model_directory]{get_model_directory()}}.}

\item{.recurse}{If \code{TRUE}, the default, search recursively in all subdirectories. Passed through to \code{fs::dir_ls()} -- If a positive number, the number of levels to recurse.}

\item{.log_df}{a \code{bbi_run_log_df} tibble (the output of \code{\link[=run_log]{run_log()}})}
}
\value{
\code{config_log()} will return a tibble with one row per \code{bbi_config.json} found in \code{.base_dir} (and optionally subdirectories).

\code{add_config()} takes a \code{bbi_run_log_df} tibble (the output from \code{\link[=run_log]{run_log()}}) and returns the input tibble,
with all the columns from \code{config_log()} joined onto it.
}
\description{
Parses \code{bbi_config.json} files into a log tibble.
This file is created by babylon, in the model output folder.
It stores metadata about the execution of a model run,
including md5 hashes of the model file (control stream) and data file at the time the model was run.
}
\seealso{
\code{\link[=run_log]{run_log()}}
}
