% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-log.R
\name{config_log_impl}
\alias{config_log_impl}
\title{Build config log}
\usage{
config_log_impl(.mods)
}
\arguments{
\item{.mods}{List of model objects. The function will attempt to find a \code{bbi_config.json} file for each model.}
}
\description{
Private implementation function for building the config log from a list of model objects.
This is called by both \code{\link[=config_log]{config_log()}} and \code{\link[=add_config]{add_config()}}.
}
\keyword{internal}
