% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-log.R
\name{config_log_make_entry}
\alias{config_log_make_entry}
\alias{config_log_make_entry.default}
\alias{config_log_make_entry.bbi_nonmem_model}
\title{Prepare a model-specific config log entry}
\usage{
config_log_make_entry(.mod, config, fields = NULL)

\method{config_log_make_entry}{default}(.mod, config, fields = NULL)

\method{config_log_make_entry}{bbi_nonmem_model}(.mod, config, fields = NULL)
}
\arguments{
\item{.mod}{A model object.}

\item{config}{The raw configuration read from \code{bbi_config.json}.}

\item{fields}{Requested fields to include in the config. If \code{NULL}, a default
set of fields for the particular model type should be used.}
}
\value{
A two element list. The first element is named "config" and contains
the prepared config object. The second is named "fields" and is a character
vector of fields that includes those specified via the \code{fields} parameter
as well as any additional fields that were automatically tacked on.
}
\description{
\code{\link[=config_log]{config_log()}} relies on \code{\link[=config_log_entry]{config_log_entry()}} to create a entry.
\code{\link[=config_log_entry]{config_log_entry()}} reads in configuration and relies on this method to
prepare and tailor the config object for a given model type.
}
