% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{convert_psn}
\alias{convert_psn}
\title{Convert the output of a PsN run to a bbr model}
\usage{
convert_psn(
  .psn_mod_dir,
  .psn_run_dir,
  .bbr_mod_dir = NULL,
  .method = c("execute", "bootstrap", "scm"),
  .description = NULL,
  .based_on = NULL,
  .tags = NULL,
  .star = NULL,
  .cleanup_psn = FALSE,
  .overwrite = FALSE
)
}
\arguments{
\item{.psn_mod_dir}{Directory containing where the model and table files exist.
Note that the model file must have a \code{.mod} or \code{.ctl} file extension.}

\item{.psn_run_dir}{Directory containing where the model was run (i.e. the generated \code{modelfit_dir}).}

\item{.bbr_mod_dir}{Directory to copy the generated bbr model to. If \code{NULL}, defaults to \code{.psn_mod_dir}.}

\item{.method}{How the model was run. Only \code{'execute'} is currently supported}

\item{.description}{Character scalar description of new model run. This will
be stored in the yaml (and can be viewed later in \code{run_log()}).}

\item{.based_on}{Character scalar or vector of paths to other models that
this model was "based on." These are used to reconstuct model developement
and ancestry. \strong{Paths must be relative to \code{.yaml_path}.}}

\item{.tags}{A character scalar or vector with any user tags to be added to
the YAML file}

\item{.star}{Boolean, marks model to indicate special interest level.}

\item{.cleanup_psn}{logical (T/F). Whether to delete relevant PsN files after the bbr model is created}

\item{.overwrite}{logical (T/F). Whether to overwrite files in \code{.bbr_mod_dir}}
}
\description{
Convert the output of a PsN run to a bbr model
}
