% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{convert_psn_data_path}
\alias{convert_psn_data_path}
\title{Convert data path during conversion of PsN to bbr model submission}
\usage{
convert_psn_data_path(.mod_path, .psn_mod_dir, .bbr_run_dir)
}
\arguments{
\item{.mod_path}{Path to a NONMEM control stream file.}

\item{.psn_mod_dir}{Directory containing where the model and table files exist.
Note that the model file must have a \code{.mod} or \code{.ctl} file extension.}

\item{.bbr_run_dir}{Directory to copy the generated bbr run files to.}
}
\description{
Convert data path during conversion of PsN to bbr model submission
}
\keyword{internal}
