% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{convert_psn_execute}
\alias{convert_psn_execute}
\title{Convert the output of a PsN execute run to a bbr model}
\usage{
convert_psn_execute(.psn_mod_dir, .psn_run_dir, .bbr_mod_dir)
}
\arguments{
\item{.psn_mod_dir}{Directory containing where the model and table files exist.
Note that the model file must have a \code{.mod} or \code{.ctl} file extension.}

\item{.psn_run_dir}{Directory containing where the model was run (i.e. the generated \code{modelfit_dir}).}

\item{.bbr_mod_dir}{Directory to copy the generated bbr model to. If \code{NULL}, defaults to \code{.psn_mod_dir}.}
}
\description{
Convert the output of a PsN execute run to a bbr model
}
\keyword{internal}
