% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-model-from.R
\name{copy_control_stream}
\alias{copy_control_stream}
\title{Copy a NONMEM control stream file}
\usage{
copy_control_stream(
  .parent_model_path,
  .new_model_path,
  .overwrite,
  .update_model_file = FALSE
)
}
\arguments{
\item{.parent_model_path}{Path to the control stream to copy}

\item{.new_model_path}{Path to copy the new control stream to}

\item{.overwrite}{If \code{TRUE}, overwrite existing file at \code{.new_model_path}. If \code{FALSE} and file exists at \code{.new_model_path} error.}

\item{.update_model_file}{\strong{Only relevant to NONMEM models.} If \code{TRUE}, the
default, update the newly created model file. If \code{FALSE}, new model file
will be an exact copy of its parent. For a NONMEM model, this currently
means only the \verb{$PROBLEM} line in the new control stream will be updated to
read \verb{See \{.new_model\}.yaml. Created by bbr.}.}
}
\description{
Helper function to copy a NONMEM control stream file and optionally update the new file.
Note that any file existing at \code{.new_model_path} will be overwritten.
}
\keyword{internal}
