% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-model-from.R
\name{copy_nonmem_model_from}
\alias{copy_nonmem_model_from}
\title{Copy model from an existing NONMEM model}
\usage{
copy_nonmem_model_from(
  .parent_mod,
  .new_model,
  .description,
  .based_on_additional = NULL,
  .add_tags = NULL,
  .inherit_tags = FALSE,
  .update_model_file = TRUE,
  .overwrite = FALSE
)
}
\arguments{
\item{.parent_mod}{S3 object of class `bbi_nonmem_model` to be used as the basis for copy.}

\item{.description}{Description of new model run. This will be stored in the yaml (to be used later in `create_run_log()`) and optionally passed into the `$PROBLEM` of the new control stream.}

\item{.update_model_file}{Boolean for whether to update the `$PROBLEM` line in the new control stream. By default it is TRUE, but if FALSE is passed `{.new_model}.[mod|ctl]` will be an exact copy of its parent control stream.}
}
\value{
S3 object of class `bbi_nonmem_model` that can be passed to `submit_nonmem_model()`
}
\description{
Private implementation function called by `copy_model_from()` dispatches.
Create new .mod/ctl and new .yaml files based on a previous model. Used for iterating on model development.
Also fills in necessary YAML fields for using `create_run_log()` later.
}
\keyword{internal}
