% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{copy_omegas}
\alias{copy_omegas}
\title{Copy omega records from one model to another}
\usage{
copy_omegas(.mod_inherit_path, .mod_path, .new_omegas)
}
\arguments{
\item{.mod_inherit_path}{model path to inherit properties from.}

\item{.mod_path}{model path to overwrite}

\item{.new_omegas}{list of new omega matrices, or a single omega matrix}
}
\description{
Copy omega records from one model to another
}
\keyword{internal}
