% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{copy_psn_tables}
\alias{copy_psn_tables}
\title{Get NONMEM table files from control file}
\usage{
copy_psn_tables(.mod_path, .psn_run_dir, .bbr_run_dir, .overwrite = FALSE)
}
\arguments{
\item{.mod_path}{Path to a NONMEM control stream file.}

\item{.psn_run_dir}{PsN execution directory where the model was run (e.g., \code{modelfit_dir/NM_run1}).}

\item{.bbr_run_dir}{Directory to copy the \strong{generated bbr run files} to. Subdirectory of \code{.bbr_dir}}

\item{.overwrite}{Logical (T/F). Whether to overwrite files if they already exist.}
}
\value{
a vector of file paths
}
\description{
Get NONMEM table files from control file
}
\keyword{internal}
