% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{copy_record_opt}
\alias{copy_record_opt}
\title{Overwrite an \code{nmrec} record option}
\usage{
copy_record_opt(val_recs, new_values, .bounds_opts)
}
\arguments{
\item{val_recs}{list of \code{nmrec} record options.}

\item{new_values}{vector of replacement values. Should be the same length as
\code{val_recs}}
}
\description{
Overwrite an \code{nmrec} record option
}
\keyword{internal}
