% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{copy_sigmas}
\alias{copy_sigmas}
\title{Copy sigma records from one model to another}
\usage{
copy_sigmas(.mod_inherit_path, .mod_path, .new_sigmas)
}
\arguments{
\item{.mod_inherit_path}{model path to inherit properties from.}

\item{.mod_path}{model path to overwrite}

\item{.new_sigmas}{list of new sigma matrices, or a single sigma matrix}
}
\description{
Copy sigma records from one model to another
}
\keyword{internal}
