% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{copy_thetas}
\alias{copy_thetas}
\title{Copy theta records from one model to another}
\usage{
copy_thetas(.mod_inherit_path, .mod_path, .new_thetas)
}
\arguments{
\item{.mod_inherit_path}{model path to inherit properties from.}

\item{.mod_path}{model path to overwrite}

\item{.new_thetas}{list of new theta vectors, or a single theta vector}
}
\description{
Copy theta records from one model to another
}
\keyword{internal}
