% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-cor.R
\name{cov_cor}
\alias{cov_cor}
\alias{cov_cor.bbi_nonmem_model}
\alias{cov_cor.bbi_nonmem_summary}
\title{NONMEM Covariance and Correlation Matrices}
\usage{
cov_cor(.mod, .threshold = NULL, ..., .dry_run = FALSE)

\method{cov_cor}{bbi_nonmem_model}(.mod, .threshold = NULL, ..., .dry_run = FALSE)

\method{cov_cor}{bbi_nonmem_summary}(.mod, .threshold = NULL, ..., .dry_run = FALSE)
}
\arguments{
\item{.mod}{Model to check.}

\item{.threshold}{Numeric scalar between 0 and 1 (defaults to \code{0.95}). Will
print a warning if the absolute values of any of the \emph{off-diagonals} in the
correlation matrix are above this threshold.}

\item{...}{args passed through to \code{\link[=bbi_exec]{bbi_exec()}}}

\item{.dry_run}{show what the command would be without actually running it}
}
\value{
A named list of matrices.
}
\description{
Pulls in the covariance and correlation matrices (for the final estimation method) for NONMEM models.
If no \code{.cov} or \code{.cor} files are found for the relevant model, will error informing the user.
}
\section{Methods (by class)}{
\itemize{
\item \code{bbi_nonmem_model}: Get cov and cor from \code{bbi_nonmem_model} object

\item \code{bbi_nonmem_summary}: Get cov and cor from \code{bbi_nonmem_summary} object (output of \code{model_summary()})
}}

\seealso{
\code{\link[=model_summary]{model_summary()}}, \code{\link[=check_cor_threshold]{check_cor_threshold()}}
}
