% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crash_model_run}
\alias{crash_model_run}
\title{Crash a specific job on the grid}
\usage{
crash_model_run(run_name)
}
\arguments{
\item{run_name}{character vector. The \code{name(s)} of the job(s) you want to end.}
}
\description{
Wrapper for \code{qdel}
}
\details{
It is advised to run \code{\link[=fetch_model_runs]{fetch_model_runs()}} (or \code{qstat} in your terminal) before calling this function to avoid accidentally deleting the wrong job.
}
\examples{
\dontrun{
crash_model_run("Run_1")
crash_model_run(c("Run_1", "Run_4"))
}

}
