% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_bbi_object}
\alias{create_bbi_object}
\alias{create_model_object}
\alias{create_summary_object}
\alias{create_process_object}
\alias{create_run_log_object}
\title{Create bbi object}
\usage{
create_model_object(.mod_list)

create_summary_object(res, .model_type = SUPPORTED_MOD_TYPES)

create_process_object(res)

create_run_log_object(log_df)
}
\arguments{
\item{.mod_list}{List with the required information to create a model object}

\item{res}{List to attempt to assign the class to}

\item{.model_type}{Character scaler of a valid model type (currently either \code{nonmem} or \code{stan})}

\item{log_df}{data.frame or tibble to attempt to assign the class to}
}
\description{
Constructor functions for objects of fundamental rbabylon classes.
}
\section{Functions}{
\itemize{
\item \code{create_model_object}: Creates list object of class \verb{bbi_\{.model_type\}_model} from named list with \code{MODEL_REQ_INPUT_KEYS}

\item \code{create_summary_object}: Create list object of \verb{bbi_\{.model_type\}_summary} class, first checking that all the required keys are present.

\item \code{create_process_object}: Create list object of \code{babylon_process} class, first checking that all the required keys are present.

\item \code{create_run_log_object}: Create list object of \code{bbi_run_log_df} class, first checking that all the required columns are present.
}}

\keyword{internal}
