% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create_model}
\alias{create_model}
\title{Create new model spec by specifying relevant information as arguments
Also creates necessary YAML file for using `create_run_log()` later.}
\usage{
create_model(
  .model_path,
  .yaml_path,
  .description,
  .based_on = "",
  .tags = "",
  .bbi_args = list(),
  .model_type = c("nonmem", "stan")
)
}
\arguments{
\item{.model_path}{Path to control stream file for the model}

\item{.yaml_path}{Path to save resulting model YAML file to}

\item{.description}{Description of new model run. This will be stored in the yaml (to be used later in `create_run_log()`) and optionally passed into the `$PROBLEM` of the new control stream.}

\item{.based_on}{A character scaler or vector of run id's (model names) that this model was "based on." These are used to reconstuct model developement and ancestry.}

\item{.tags}{A character scaler or vector with any user tags to be added to the YAML file}

\item{.bbi_args}{A named list specifying arguments to pass to babylon formatted like `list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)`. Run `print_nonmem_args()` to see valid arguments. These will be written into YAML file.}

\item{.model_type}{string to specify type of model being created (used for S3 class). Currently only `'nonmem'` is supported.}
}
\value{
S3 object of class `bbi_{.model_type}_spec` that can be passed to `submit_model()`
}
\description{
Create new model spec by specifying relevant information as arguments
Also creates necessary YAML file for using `create_run_log()` later.
}
