% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create_model_from_yaml}
\alias{create_model_from_yaml}
\title{Create new model spec from a model yaml}
\usage{
create_model_from_yaml(.yaml_path)
}
\arguments{
\item{.yaml_path}{Character scaler for the path to the YAML model file that will be used}

\item{.model_type}{Character scaler to specify type of model being created (used for S3 class). Currently only `'nonmem'` is supported.}
}
\value{
S3 object of class `bbi_{.model_type}_spec` that can be passed to `submit_model()`
}
\description{
Create new model spec from a model yaml
}
