% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_model_hook}
\alias{create_model_hook}
\alias{create_model_hook.default}
\alias{create_model_hook.bbi_nonmem_model}
\alias{create_model_hook.bbi_nmboot_model}
\alias{create_model_hook.bbi_nmsse_model}
\alias{create_model_hook.bbi_nmsim_model}
\title{Perform model-type specific setup}
\usage{
create_model_hook(.mod, ...)

\method{create_model_hook}{default}(.mod, ...)

\method{create_model_hook}{bbi_nonmem_model}(.mod, ...)

\method{create_model_hook}{bbi_nmboot_model}(.mod, ...)

\method{create_model_hook}{bbi_nmsse_model}(.mod, ...)

\method{create_model_hook}{bbi_nmsim_model}(.mod, ...)
}
\arguments{
\item{.mod}{Model object.}

\item{...}{All but the first argument that was passed to
\code{create_model_object()}.}
}
\description{
This method is called by \code{create_model_object()} after it has created an
object of type \verb{bbi_\{type\}_model}, and, if the caller requested writing the
model yaml to disk, this method will be called before that.
}
\details{
Every \verb{bbi_\{type\}_model} should define a \verb{create_model_hook.bbi_\{type\}_model}
method, even if no model-specific setup is required.
}
\keyword{internal}
