% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create_model_object}
\alias{create_model_object}
\title{Creates S3 object of class `bbi_{.model_type}_model` from list with `MODEL_REQ_INPUT_KEYS`}
\usage{
create_model_object(.mod_list)
}
\arguments{
\item{.mod_list}{List with the required information to create a model object}
}
\value{
S3 object of class `bbi_{.model_type}_model` that can be passed to `submit_model()`, `model_summary()` etc.
}
\description{
Creates S3 object of class `bbi_{.model_type}_model` from list with `MODEL_REQ_INPUT_KEYS`
}
