% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create_run_log}
\alias{create_run_log}
\title{Parses model outputs into a}
\usage{
create_run_log(.base_dir, .recurse = TRUE)
}
\value{
tibble with information on each run
}
\description{
Parses model outputs into a
}
