% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create_spec}
\alias{create_spec}
\title{Creates S3 object of class `bbi_{.model_type}_spec` from list with `SPEC_REQ_INPUT_KEYS`}
\usage{
create_spec(.mod_list)
}
\arguments{
\item{.mod_list}{List with the required information to create a spec object}
}
\value{
S3 object of class `bbi_{.model_type}_spec` that can be passed to `submit_model`
}
\description{
Creates S3 object of class `bbi_{.model_type}_spec` from list with `SPEC_REQ_INPUT_KEYS`
}
