% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dev_error}
\alias{dev_error}
\title{Construct error message that users shouldn't see}
\usage{
dev_error(.msg)
}
\description{
Contains a note to file an issue if ever encountered by a user.
}
\keyword{internal}
