% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{download_bbi}
\alias{download_bbi}
\title{Download bbi executable to the specified path.}
\usage{
download_bbi(.bbi_url, .path)
}
\arguments{
\item{.bbi_url}{URL pointing to the tar.gz for a particular release}

\item{.path}{the final destination of bbi executable}
}
\description{
Download bbi executable to the specified path.
}
\keyword{internal}
