% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_with_retry}
\alias{download_with_retry}
\title{Call \code{utils::download.file()}, retrying once on failure.}
\usage{
download_with_retry(...)
}
\arguments{
\item{...}{arguments to pass to \code{utils::download.file()}}
}
\description{
Call \code{utils::download.file()}, retrying once on failure.
}
\keyword{internal}
