% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-join.R
\name{drop_dups}
\alias{drop_dups}
\title{Drop duplicate columns to prepare for join}
\usage{
drop_dups(.new_table, .dest_table, .join_col, .table_name)
}
\description{
Drop duplicate columns to prepare for join
}
\keyword{internal}
