% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-nmtran.R
\name{execute_nmtran}
\alias{execute_nmtran}
\title{Execute NMTRAN in a given directory}
\usage{
execute_nmtran(nmtran_exe, mod_path, dir = NULL)
}
\arguments{
\item{nmtran_exe}{Path to \code{NMTRAN} executable.}

\item{mod_path}{Path of a model to evaluate. Should be relative to \code{dir}.}

\item{dir}{Directory in which to execute the command.}
}
\description{
Execute NMTRAN in a given directory
}
\keyword{internal}
