% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{extract_data_path}
\alias{extract_data_path}
\title{Helper function to extract the data path from the $DATA line of a raw control stream}
\usage{
extract_data_path(.mod)
}
\description{
Helper function to extract the data path from the $DATA line of a raw control stream
}
