% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-log.R
\name{extract_from_summary}
\alias{extract_from_summary}
\alias{extract_param_count}
\alias{extract_details}
\alias{extract_heuristics}
\alias{extract_ofv}
\alias{extract_condition_number}
\title{Extract from summary object}
\usage{
extract_param_count(.s)

extract_details(.s)

extract_heuristics(.s)

extract_ofv(.s)

extract_condition_number(.s)
}
\arguments{
\item{.s}{The summary object to extract from}

\item{.subfield}{The field to extract from within "ofv". Defaults to the objective function value with no constant added, but the other fields are available too.}
}
\description{
These are all helper functions to extract a specific field or sub-field from a \verb{bbi_\{.model_type\}_summary} object.
}
\section{Functions}{
\itemize{
\item \code{extract_param_count}: Extract count of non-fixed parameters

\item \code{extract_details}: Extract \code{run_details} field

\item \code{extract_heuristics}: Extract \code{run_heuristics} field

\item \code{extract_ofv}: Extract objective function value (without constant) for the final estimation method

\item \code{extract_condition_number}: Extract condition number for the final estimation method
}}

