% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-log.R
\name{file_matches}
\alias{file_matches}
\title{Compare a file to an MD5 sum}
\usage{
file_matches(path, md5)
}
\arguments{
\item{path}{String giving the path to the file.}

\item{md5}{String giving expected MD5 sum.}
}
\value{
\code{TRUE} if \code{path} matches \code{md5}, otherwise \code{FALSE} (including if
\code{path} doesn't exist).
}
\description{
Compare a file to an MD5 sum
}
\keyword{internal}
