% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_matches_string}
\alias{file_matches_string}
\title{Compare a file to a string}
\usage{
file_matches_string(path, string, append = "\\n")
}
\arguments{
\item{path}{String giving the path to the file.}

\item{string}{Character scalar to hash and compare to hash of file}

\item{append}{Character scalar to append to the end of string. Defaults
to \code{"\\n"} because helpers like \code{writeLines()} and \code{readr::write_lines()}
automatically append a \code{"\\n"} to a string when writing to a file.}
}
\value{
\code{TRUE} if \code{path} matches \code{string}, otherwise \code{FALSE} (including if
\code{path} doesn't exist).
}
\description{
Similar to \link{file_matches} but takes a string that
will be hashed to MD5 before comparing. By default
it also appends a new line to the end of the string
before hashing (because most file writer functions
do this automatically).
}
\keyword{internal}
