% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-nm-data.R
\name{filter_nm_data}
\alias{filter_nm_data}
\title{Filter \code{NONMEM} input data based on \code{IGNORE} and \code{ACCEPT} record options}
\usage{
filter_nm_data(.mod, data = nm_data(.mod))
}
\arguments{
\item{.mod}{A \code{bbi_nonmem_model} object}

\item{data}{A starting dataset}
}
\description{
Filter \code{NONMEM} input data based on \code{IGNORE} and \code{ACCEPT} record options
}
\keyword{internal}
