% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{filter_prior_records}
\alias{filter_prior_records}
\title{Filter out prior records}
\usage{
filter_prior_records(.records)
}
\arguments{
\item{.records}{list of \code{nmrec} record objects}
}
\description{
Filter out prior records
}
\details{
\code{bbr} does not support the inheritance of priors at this time. To simplify
the inheritance of previous parameter estimates, filter prior blocks out,
and ensure the returned object is still a list of records.
}
\keyword{internal}
