% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_config_file_path}
\alias{find_config_file_path}
\title{Private helper to search for babylon.yaml config files}
\usage{
find_config_file_path(.config_path, .model_dir)
}
\arguments{
\item{.config_path}{Path to config, possibly relative}

\item{.model_dir}{absolute path to directory where model is be run from}
}
\value{
path to babylon.yaml in `.config_path`, relative to `.model_dir`
}
\description{
Private helper to search for babylon.yaml config files
}
