% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-log.R
\name{find_models}
\alias{find_models}
\title{Search for model YAML files and read them}
\usage{
find_models(.base_dir, .recurse)
}
\arguments{
\item{.base_dir}{Directory to search for model YAML files.}

\item{.recurse}{If \code{TRUE} search recursively in subdirectories as well.}
}
\description{
Private helper function that searches from a base directory for any YAML files (excluding \code{babylon.yaml})
and attempts to read them to a model object with \code{safe_read_model()}.
}
\keyword{internal}
