% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{find_nonmem_model_file_path}
\alias{find_nonmem_model_file_path}
\title{Find the path to a NONMEM model file}
\usage{
find_nonmem_model_file_path(.path, .check_exists = TRUE)
}
\arguments{
\item{.path}{File path to a NONMEM model file (control stream) with either
\code{.ctl} or \code{.mod} extension.}

\item{.check_exists}{If \code{TRUE}, the default, will throw an error if the file does not exist}
}
\value{
The path to the model file, if it exists. If the model file does not
exist and \code{.check_exists = FALSE}, a string that is \code{.path} with the \code{.ctl}
extension.
}
\description{
NONMEM recognizes two model file extensions, \code{.ctl} and (less common) \code{.mod}.
Look for these at \code{.path} and if exactly one exists, return it. If both
exist, throw an error, because there is no way to know which should be used.
If neither exists and \code{.check_exists = TRUE}, throw an error.
}
\keyword{internal}
