% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-handling.R
\name{fmt_mat_zeros}
\alias{fmt_mat_zeros}
\title{Make matrix symmetric and format other NA values as zeros}
\usage{
fmt_mat_zeros(mat)
}
\arguments{
\item{mat}{a symmetric matrix, or a matrix with only lower triangular values
specified (assumed to be symmetric).}
}
\value{
a list containing the formatted matrix, and original \code{NA} indices
}
\description{
Make matrix symmetric and format other NA values as zeros
}
\keyword{internal}
