% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-estimates.R
\name{fmt_record_num}
\alias{fmt_record_num}
\title{Get record number of initial estimates}
\usage{
fmt_record_num(inits)
}
\arguments{
\item{inits}{initial estimates object as returned by \verb{nmrec::extract_*}
functions. Should be a vector or matrix with \code{nmrec_record_size} attributes.}
}
\value{
a vector indicating the which record the initial estimate belongs to
}
\description{
Get record number of initial estimates
}
\details{
\code{attr(inits, "nmrec_record_size")} stores a vector of sizes, one for each
occurrence of a given record type.
This function creates a vector indicating the which record the initial
estimate belongs to, and correlates to the order of the initial estimates.
\itemize{
\item For \code{THETA} records, this is simply \code{rep(seq_along(sizes), sizes)}
\item For \code{OMEGA} or \code{SIGMA} records, it's a bit more complicated. Here we need
to subset the columns (not the rows) of the full lower triangular matrix to
update the \code{size} to include values created when diagonally concatenating
multiple matrix-type records.
}

This function is intended to be used with \code{matrix_to_df()} for matrix-type
records.
}
\keyword{internal}
