% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_cmd_args}
\alias{format_cmd_args}
\title{Formats command line args from a named list to a string as it would be passed on the command line}
\usage{
format_cmd_args(.args, .collapse = FALSE)
}
\arguments{
\item{.args}{A named list of .args to check}

\item{.collapse}{Boolean for whether to collapse return vector to a single string (FALSE by default)}
}
\value{
character vector of command args
}
\description{
Formats command line args from a named list to a string as it would be passed on the command line
}
