% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-param.R
\name{format_matrix}
\alias{format_matrix}
\title{Format vector into matrix using positional vector}
\usage{
format_matrix(.values, .labels, .type = c("OMEGA", "SIGMA"))
}
\arguments{
\item{.values}{vector of values. Values should be in the same order presented in model_summary}

\item{.labels}{parameter names corresponding to the values. Order matters here as well.}

\item{.type}{matrix type. Either "OMEGA" or "SIGMA"}
}
\description{
Format vector into matrix using positional vector
}
\details{
The only information gathered from \code{.labels} is the positioning (i.e. (2,3), (1,1), etc.)
}
\keyword{internal}
