% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{format_model_type}
\alias{format_model_type}
\title{Format the model type for use in \code{\link[=model_tree]{model_tree()}}}
\usage{
format_model_type(model_type, fmt_html = FALSE, ...)
}
\arguments{
\item{model_type}{one of the \code{bbr} supported model types (\code{nonmem}, \code{nboot},
\code{nsim}). Other model types are also acceptable, but won't have custom formatting.}

\item{fmt_html}{Logical (\code{T}/\code{F}). If \code{TRUE}, apply coloring and format as \code{HTML}.}

\item{...}{Additional arguments passed to \code{\link[=style_html]{style_html()}}.}
}
\description{
Format the model type for use in \code{\link[=model_tree]{model_tree()}}
}
\keyword{internal}
