% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_nonmem_args}
\alias{format_nonmem_args}
\title{Checks that all passed NONMEM command line args are valid and formats}
\usage{
format_nonmem_args(.args)
}
\arguments{
\item{.args}{A named list of .args to check}
}
\description{
Checks that all passed NONMEM command line args are valid and formats
}
