% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_absolute_model_path}
\alias{get_absolute_model_path}
\title{Get the absolute model path from a model object}
\usage{
get_absolute_model_path(.mod)
}
\arguments{
\item{.mod}{A bbi model or summary object.}
}
\value{
The path to the working directory.
}
\description{
This is a simple wrapper around pulling the
\code{absolute_model_path} from the model object. It
only exists to protect against future refactors
that might change the structure of the model object.
}
\keyword{internal}
