% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-run-utils.R
\name{get_analysis_models}
\alias{get_analysis_models}
\title{Read in all analysis run model objects}
\usage{
get_analysis_models(.run)
}
\arguments{
\item{.run}{An analysis run (\code{bbi_nmboot_model}, \code{bbi_nmsse_model}) or
analysis run summary (\code{bbi_nmboot_summary}) object.}
}
\description{
Read in all analysis run model objects
}
\keyword{internal}
