% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_analysis_spec}
\alias{get_analysis_spec}
\title{Read in and format an analysis specification file, such as for bootstraps or
SSE.}
\usage{
get_analysis_spec(.run)
}
\arguments{
\item{.run}{An analysis run (\code{bbi_nmboot_model}, \code{bbi_nmsse_model}) or
analysis run summary (\code{bbi_nmboot_summary}) object.}
}
\value{
a list
}
\description{
Tabulates analysis metadata and all relevant \code{bbi_nonmem_model} model files
from an analysis specification file.
}
\keyword{internal}
