% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_analysis_sum_path}
\alias{get_analysis_sum_path}
\alias{get_analysis_sum_path.bbi_nmboot_model}
\alias{get_analysis_sum_path.bbi_nmsse_model}
\title{Get the relevant analysis run summary path}
\usage{
get_analysis_sum_path(.mod, .check_exists = TRUE)

\method{get_analysis_sum_path}{bbi_nmboot_model}(.mod, .check_exists = TRUE)

\method{get_analysis_sum_path}{bbi_nmsse_model}(.mod, .check_exists = TRUE)
}
\arguments{
\item{.mod}{a \verb{bbi_\{.model_type\}_model} object.}

\item{.check_exists}{If \code{TRUE}, the default, will throw an error if the file
does not exist.}
}
\description{
Get the relevant analysis run summary path
}
\section{Methods (by class)}{
\itemize{
\item \code{get_analysis_sum_path(bbi_nmboot_model)}: Get the summary RDS file path from a
\code{bbi_nmboot_model} object

\item \code{get_analysis_sum_path(bbi_nmsse_model)}: Get the summary RDS file path from a
\code{bbi_nmsse_model} object

}}
\keyword{internal}
