% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-threads.R
\name{get_block_idx}
\alias{get_block_idx}
\title{Get location of specific code blocks in a ctl file}
\usage{
get_block_idx(.mod_lines, .block = "EST")
}
\arguments{
\item{.mod_lines}{ctl lines returned from readLines}
}
\description{
Get location of specific code blocks in a ctl file
}
\keyword{internal}
