% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{get_block_labels}
\alias{get_block_labels}
\title{Get block label from parsed control stream file}
\usage{
get_block_labels(.blocks)
}
\arguments{
\item{.block}{character vector or list of vectors containing the lines of the control stream file.
Must contain a block header (e.g. $OMEGA)}
}
\description{
Get block label from parsed control stream file
}
\keyword{internal}
