% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-threads.R
\name{get_est_idx}
\alias{get_est_idx}
\title{Get location of $EST blocks}
\usage{
get_est_idx(.mod_lines)
}
\arguments{
\item{.mod_lines}{ctl lines returned from readLines}
}
\description{
Get location of $EST blocks
}
\keyword{internal}
