% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-log.R
\name{get_from_summay_log}
\alias{get_from_summay_log}
\title{Extract a elements or format parameter estimates from the \code{bbi_summary} object
in a summary log}
\usage{
get_from_summay_log(
  .log_df,
  element = c(SUMMARY_DETAILS, SUMMARY_HEURISTICS, "parameter_estimates")
)
}
\arguments{
\item{.log_df}{A \code{bbi_summary_log_df} object.}

\item{element}{Element to extract. \code{"run_details"} and \code{"run_heuristics"} are
extracted directly. Passing \code{"parameter_estimates"} will force a call to
\code{\link[=param_estimates]{param_estimates()}} using the \code{bbi_summary} object in each row.}
}
\description{
Extract a elements or format parameter estimates from the \code{bbi_summary} object
in a summary log
}
\keyword{internal}
