% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-estimates.R
\name{get_initial_est}
\alias{get_initial_est}
\title{Retrieve initial parameter estimates}
\usage{
get_initial_est(.mod, flag_fixed = FALSE)
}
\arguments{
\item{.mod}{a \code{bbr} model object}

\item{flag_fixed}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, return which
parameters are fixed.}
}
\description{
Retrieve initial parameter estimates
}
\keyword{internal}
