% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{get_matrix_types}
\alias{get_matrix_types}
\title{Get matrix type for omega and sigma records}
\usage{
get_matrix_types(.blocks)
}
\description{
Get matrix type for omega and sigma records
}
\keyword{internal}
