% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_model_status}
\alias{get_model_status}
\alias{get_model_status.default}
\title{Get the model execution status}
\usage{
get_model_status(.mod, max_print = 10, ...)

\method{get_model_status}{default}(.mod, max_print = 10, ...)
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model}, \code{bbi_nmboot_model}, or list of
\code{bbi_nonmem_model} objects. Other packages (e.g., \code{bbr.bayes}) may add
additional methods.}

\item{max_print}{max number of models to explicitly print to the console. If
the number of finished or incomplete models are greater than this number,
just print the \emph{number} of models.}

\item{...}{Arguments passed to methods.}
}
\value{
invisibly returns a tibble of model ids and whether the model has
finished executing
}
\description{
Returns messages indicating which model(s) have finished executing and which
are still running.
}
\details{
If the result are saved to an R object, users can inspect the full list of
runs to determine exactly which models are still running vs. have finished
executing:

\if{html}{\out{<div class="sourceCode">}}\preformatted{res <- get_model_status(mod_list)
res
}\if{html}{\out{</div>}}
}
\seealso{
wait_for_nonmem check_nonmem_finished
}
