% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_model_working_directory}
\alias{get_model_working_directory}
\title{Get the working directory of a model object}
\usage{
get_model_working_directory(.mod)
}
\arguments{
\item{.mod}{A bbi model or summary object.}
}
\value{
The path to the working directory.
}
\description{
The working directory of a model object is the parent directory of the
absolute model path.
}
\keyword{internal}
