% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-model-from.R
\name{get_next_integer}
\alias{get_next_integer}
\title{Check directory for model files with integer names
and return the next integer. Used by copy_model_from().}
\usage{
get_next_integer(.parent_mod)
}
\description{
Check directory for model files with integer names
and return the next integer. Used by copy_model_from().
}
\keyword{internal}
