% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{get_nonmem_tables}
\alias{get_nonmem_tables}
\title{Get NONMEM table files from control file}
\usage{
get_nonmem_tables(.mod_path, .mod_dir)
}
\arguments{
\item{.mod_path}{path to a bbr model object}

\item{.mod_dir}{directory containing table files}
}
\description{
Get NONMEM table files from control file
}
\keyword{internal}
