% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-param.R
\name{get_omega}
\alias{get_omega}
\alias{get_sigma}
\alias{get_theta}
\title{Return a formatted object for a given parameter}
\usage{
get_omega(.summary)

get_sigma(.summary)

get_theta(.summary)
}
\arguments{
\item{.summary}{a \code{bbi_nonmem_summary} or \code{bbi_summary_list} object.}
}
\description{
\itemize{
\item \code{\link[=get_omega]{get_omega()}}: Return omega values as a labeled symmetric matrix
\item \code{\link[=get_sigma]{get_sigma()}}: Return sigma values as a labeled symmetric matrix
\item \code{\link[=get_theta]{get_theta()}}: Return theta values as a named vector
}
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode r">}}\preformatted{sum <- mod \%>\% model_summary()

sum \%>\% get_omega()
#>              OMEGA_1     OMEGA_2     OMEGA_3      OMEGA_4
#> OMEGA_1  0.010410500 0.000319542 0.001977200 -0.000189864
#> OMEGA_2  0.000319542 0.008278020 0.000872168  0.001005720
#> OMEGA_3  0.001977200 0.000872168 0.013139700  0.004040370
#> OMEGA_4 -0.000189864 0.001005720 0.004040370  0.010987300


sum \%>\% get_sigma()
#>         SIGMA_1
#> SIGMA_1       1


sum \%>\% get_theta()
#>     THETA1     THETA2     THETA3     THETA4     THETA5     THETA6     THETA7 
#>  3.3005000  3.2519700 -0.6122190 -0.2080120  0.7372940  1.1359900  0.3346290 
#>     THETA8     THETA9    THETA10    THETA11 
#>  0.1923520  0.6916970  2.2993700  0.0991684
}\if{html}{\out{</div>}}
}

