% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{get_param_block}
\alias{get_param_block}
\title{Get location and lines of specific code blocks in a ctl file}
\usage{
get_param_block(.mod_lines, .block = c("THETA", "OMEGA", "SIGMA"))
}
\arguments{
\item{.mod_lines}{ctl lines returned from readLines}

\item{.block}{character string defining which block in the control stream you are trying to parse}
}
\description{
Get location and lines of specific code blocks in a ctl file
}
\details{
This function is used in \code{inherit_param_estimates}, and therefore contains some
additional filtering rules:
\itemize{
\item Handles old and new method for specifying priors. These are filtered out
\item Will filter out any comments (as in they wont be overwritten)
\item Will remove any empty lines
}
}
\keyword{internal}
