% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{get_param_bound_type}
\alias{get_param_bound_type}
\title{Determine the type of bounds (or fixed parameter) for model parameters}
\usage{
get_param_bound_type(.record_opt)
}
\arguments{
\item{.record_opt}{an \code{nmrec_option_nested} object to check}
}
\description{
This function determines the type of bounds specified for model parameters,
limited to THETA, OMEGA, and SIGMA records in a NONMEM control stream file.
}
\keyword{internal}
