% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-param-inits.R
\name{get_param_inits}
\alias{get_param_inits}
\title{Retrieve and format the initial estimates}
\usage{
get_param_inits(.mod, flag_fixed = FALSE)
}
\arguments{
\item{.mod}{a \code{bbr} model object}

\item{flag_fixed}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, return which
parameters are fixed.}
}
\description{
Retrieve and format the initial estimates
}
