% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_data.R
\name{get_player_data}
\alias{get_player_data}
\title{Get data on one NBA player}
\usage{
get_player_data(slug)
}
\arguments{
\item{slug}{The slug for the player you're interested in. This
can be found using the \code{get_players()} function, and is part of 
the URL to the data of a player, e.g., 
if the URL is https://www.basketball-reference.com/players/a/abdelal01.html
then the slug is abdelal01.}
}
\value{
A data.frame with data from one player.

The columns include:

\describe{
\item{player}{Player name}
\item{season}{What season is this data from?}
\item{age}{Age of player in season}
\item{lg}{League, e.g. NBA, ABA}
\item{pos}{Position played}
\item{g}{Number of games played}
\item{gs}{Number of games as starter}
\item{mp}{Minutes played per game}
\item{fg}{Field goals per game}
\item{fga}{Field goal attempts per game}
\item{fg\%}{Field goal percentage}
\item{3p}{3-point field goals per game}
\item{3pa}{3-point attempts per game}
\item{3p\%}{3-point field goal percentage}
\item{2p}{2-point field goals per game}
\item{2pa}{2-point attempts per game}
\item{2p\%}{2-point field goal percentage}
\item{efg\%}{Effective field goal percentage}
\item{ft}{Free throws per game}
\item{fta}{Free throw attempts per game}
\item{ft\%}{Free throw percentage}
\item{orb}{Offensive rebounds per game}
\item{drb}{Defensive rebounds per game}
\item{trb}{Total rebounds per game}
\item{ast}{Assists per game}
\item{stl}{Steals per game}
\item{blk}{Blocks per game}
\item{tov}{Turnovers per game}
\item{pf}{Personal fouls per game}
\item{pts}{Points per game}
}
}
\description{
This function retrieves data for one specific NBA player
}
\examples{
abdelnaby_d <- get_player_data('abdelal01')
jordan_d <- get_player_data('jordami01')
}
