% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_players.R
\name{get_players}
\alias{get_players}
\title{Find information for all NBA players}
\usage{
get_players(initial)
}
\arguments{
\item{initial}{What's the last initial you're interested in? E.g., to find
information on John Stockton, the initial is "S".}
}
\value{
A data.frame with a row for each player with that last initial.

The columns include:

\describe{
\item{player}{Player name}
\item{from}{First year in league}
\item{to}{Last year in league}
\item{pos}{Position(s) played}
\item{ht}{Height in 'feet-inches' format}
\item{wt}{Weight in pounds}
\item{birth_date}{Birth date}
\item{college}{What college did the player attend?}
}
}
\description{
This function retrieves NBA player data for players by last initial
}
\examples{
d <- get_players("A")

}

