% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{get_psn_submission_info}
\alias{get_psn_submission_info}
\title{Get PsN execution information}
\usage{
get_psn_submission_info(.modelfit_dir, .psn_model_file = NULL)
}
\arguments{
\item{.modelfit_dir}{Directory containing where the model was run (i.e. the generated \code{modelfit_dir}).}

\item{.psn_model_file}{Model file used during PsN execution. Note that the model file must have a \code{.mod} or \code{.ctl} file extension.}
}
\description{
Retrieves PsN model execution information from either "version_and_option_info.txt" or "meta.yaml" depending on the version
of PsN the user has installed
}
\keyword{internal}
