% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{get_record_attr}
\alias{get_record_attr}
\title{Parse values per block}
\usage{
get_record_attr(.param_recs)
}
\arguments{
\item{.param_recs}{list of \code{nmrec} record objects specific to the parameter}
}
\description{
Parse values per block
}
\keyword{internal}
