% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season}
\alias{get_season}
\title{Scrape NBA player data by season}
\usage{
get_season(year)
}
\arguments{
\item{year}{The season to collect data for. If you want the 2016-2017 data,
use the latter year (2017).}
}
\value{
A data.frame with a row for each player X team combo in that season.
}
\description{
This function retrieves NBA player data for a specific season, specified by
year. There will be one row per player X team combination. That is, if a
player played for multiple teams in one season, they will show up on multiple
rows (one row for each team).
}
\details{
The columns include:

\describe{
\item{player}{Player name}
\item{pos}{Position}
\item{age}{Age in years}
\item{tm}{Team}
\item{g}{Games played}
\item{gs}{Games started}
\item{mp}{Minutes played}
\item{fg}{Field goals made}
\item{fga}{Field goals attempted}
\item{fg_pct}{Field goal shooting percentage}
\item{three_p}{Three point shots made}
\item{three_pa}{Three point shots attempted}
\item{three_p_pct}{Three point shooting percentage}
\item{two_p}{Two point shots made}
\item{two_pa}{Two point shots attempted}
\item{two_p_pct}{Two point shooting percentage}
\item{efg_pct}{Effective field goal percentage (adjusts for fact that 3
pointers are worth one more point than two pointers)}
\item{ft}{Free throws made}
\item{fta}{Free throw attempts}
\item{ft_pct}{Free throw percentage}
\item{orb}{Offensive rebounds}
\item{drb}{Defensive rebounds}
\item{trb}{Total rebounds}
\item{ast}{Assists}
\item{stl}{Steals}
\item{blk}{Blocks}
\item{tov}{Turnovers}
\item{pf}{Personal fouls}
\item{pts}{Points made}
}
}
\examples{
\dontrun{
get_season(2010)
}
}

