% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{get_sim_replicate_record}
\alias{get_sim_replicate_record}
\title{Determine whether \code{NONMEM} control stream is using a \verb{$PRED} or \verb{$ERROR} record}
\usage{
get_sim_replicate_record(.mod)
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} or \code{bbi_nmsim_model} object.}
}
\value{
an \code{nmrec} record type name
}
\description{
Determine whether \code{NONMEM} control stream is using a \verb{$PRED} or \verb{$ERROR} record
}
\keyword{internal}
