% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_sim_spec}
\alias{get_sim_spec}
\title{Read in and format the simulation specification file.}
\usage{
get_sim_spec(.mod)
}
\arguments{
\item{.mod}{A \code{bbi_nonmem_model}, \code{bbi_nmsim_model}, or \code{bbi_nonmem_summary}
object}
}
\value{
a list
}
\description{
Read in and format the simulation specification file.
}
\keyword{internal}
