% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_spec_path}
\alias{get_spec_path}
\alias{get_spec_path.bbi_nmboot_model}
\alias{get_spec_path.bbi_nmsim_model}
\title{Get the bootstrap specification file path}
\usage{
get_spec_path(.mod, .check_exists = TRUE)

\method{get_spec_path}{bbi_nmboot_model}(.mod, .check_exists = TRUE)

\method{get_spec_path}{bbi_nmsim_model}(.mod, .check_exists = TRUE)
}
\arguments{
\item{.mod}{a \verb{bbi_\{.model_type\}_model} object.}

\item{.check_exists}{If \code{TRUE}, the default, will throw an error if the file
does not exist.}
}
\description{
Get the bootstrap specification file path
}
\section{Methods (by class)}{
\itemize{
\item \code{get_spec_path(bbi_nmboot_model)}: Get the bootstrap specification file path

\item \code{get_spec_path(bbi_nmsim_model)}: Get the simulation specification file path

}}
\note{
this is not the configuration file generated by bbi
}
\keyword{internal}
