% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-handling.R
\name{get_sub_mat}
\alias{get_sub_mat}
\title{Separate a matrix into sub matrices}
\usage{
get_sub_mat(full_mat)
}
\arguments{
\item{full_mat}{initial estimates object as returned by \code{nmrec::extract_omega}
or \code{nmrec::extract_sigma}. Should be a matrix with \code{nmrec_record_size} attributes.}
}
\description{
Separate a matrix into sub matrices
}
