% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-param.R
\name{get_theta}
\alias{get_theta}
\title{Return theta values as a named vector}
\usage{
get_theta(.summary)
}
\arguments{
\item{.summary}{A \code{bbi_nonmem_summary} or \code{bbi_summary_list} object.}
}
\description{
Return theta values as a named vector
}
