% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{highlight_cell}
\alias{highlight_cell}
\title{Highlight cell in kable table}
\usage{
highlight_cell(.l, .i, .threshold)
}
\arguments{
\item{.l}{character scalar of the line to be formatted (a row of a kable table)}

\item{.i}{the index of the column to check}

\item{.threshold}{the threshold to check against. If value is greater than
.threshold then it is formatted as red.}
}
\value{
character vector of formatted values
}
\description{
Highlights in red numeric cells that are above the specified threshold.
}
\keyword{internal}
