% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{import_result}
\alias{import_result}
\title{Create a `bbi_nonmem_result` object from a file path (should work with either path to a model file, yaml file, and output folder)}
\usage{
import_result(.path)
}
\arguments{
\item{.path}{Character scaler with the file path to use}
}
\description{
Create a `bbi_nonmem_result` object from a file path (should work with either path to a model file, yaml file, and output folder)
}
