% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{inherit_param_estimates}
\alias{inherit_param_estimates}
\title{Inherit parameter estimates}
\usage{
inherit_param_estimates(.mod, .parent_mod = get_based_on(.mod))
}
\arguments{
\item{.mod}{New model object to overwrite}

\item{.parent_mod}{model path to inherit properties from}
}
\description{
Inherit parameter estimates
}
