% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-initial-est.R
\name{initial_estimates}
\alias{initial_estimates}
\title{Retrieve and format initial parameter estimates}
\usage{
initial_estimates(.mod, flag_fixed = TRUE)
}
\arguments{
\item{.mod}{a \code{bbr} model object}

\item{flag_fixed}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, return which
parameters are fixed.}
}
\description{
Retrieve and format initial parameter estimates
}
\note{
\code{NA} values indicate that they were not specified in the control stream file.
This is true for the initial estimate, \code{THETA} bounds, and whether a given
parameter is \code{FIXED} or not.
}
