% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{inspect_option_class}
\alias{inspect_option_class}
\title{Determine if record option contains a specific class (count if desired)}
\usage{
inspect_option_class(
  .record_opt,
  .class,
  .operation = c("any", "count", "index"),
  .inherits = c("all", "any")
)
}
\arguments{
\item{.record_opt}{an \code{nmrec_option_pos} object to check. \emph{Can be} nested
(inherit class \code{nmrec_option_nested}).}

\item{.class}{class or vector of classes to look for.}

\item{.operation}{which operation to perform (i.e. what to return)
\describe{
\item{\code{'any'}}{Returns Logical (\code{TRUE}/\code{FALSE}): whether the \code{.class} is found in the record at all}
\item{\code{'count'}}{Returns Numeric: frequency(ies) of \code{.class}}
\item{\code{'index'}}{Returns Numeric: index(es) of \code{.class}}
}}

\item{.inherits}{Either \code{'all'} or \code{'any'}. Only relevant if \code{class} is a vector.
If \code{.inherits = 'all'}, the classes must correspond to the same element.
If \code{.inherits = 'any'}, the classes can correspond to different elements.}
}
\description{
Determine if record option contains a specific class (count if desired)
}
\keyword{internal}
