% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-nm-data.R
\name{invert_operator}
\alias{invert_operator}
\title{Function to invert \code{R} operators in filter expressions}
\usage{
invert_operator(expr)
}
\arguments{
\item{expr}{A \code{\link[dplyr:filter]{dplyr::filter()}} expression}
}
\value{
the inverted expression
}
\description{
Function to invert \code{R} operators in filter expressions
}
\examples{
\dontrun{

invert_operator(c('A==2', 'B >= 4'))
#> [1] "A!=2"   "B <= 4"
}
}
\seealso{
\code{\link[=translate_nm_operator]{translate_nm_operator()}}, \code{\link[=translate_nm_expr]{translate_nm_expr()}}
}
\keyword{internal}
