% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimates.R
\name{is_diag}
\alias{is_diag}
\title{Check if diagonal index or not}
\usage{
is_diag(.name)
}
\arguments{
\item{.name}{A character scaler containing an index string}
}
\description{
Private helper to unpack an matrix index string like '(3,3)' is for a diagonal (i.e. if the numbers are the same)
}
