% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-handling.R
\name{is_positive_definite}
\alias{is_positive_definite}
\title{Check if matrix is positive-definite}
\usage{
is_positive_definite(mat)
}
\arguments{
\item{mat}{a square matrix}
}
\description{
Assumes matrix is a square matrix, is symmetric, and contains only numeric
values
}
\details{
References:
https://github.com/TomKellyGenetics/matrixcalc/blob/master/R/is.positive.definite.R
https://www.geeksforgeeks.org/fixing-non-positive-definite-correlation-matrices-using-r/
}
\keyword{internal}
