% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{is_valid_nonmem_extension}
\alias{is_valid_nonmem_extension}
\title{Private helper to check if file has a valid NONMEM control stream extension}
\usage{
is_valid_nonmem_extension(.path)
}
\arguments{
\item{.path}{File path to check}
}
\description{
Private helper to check if file has a valid NONMEM control stream extension
}
\keyword{internal}
