% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{linux_install_commands}
\alias{linux_install_commands}
\title{Private helper function for building commands to install bbi on Linux}
\usage{
linux_install_commands(.dir, .bbi_url)
}
\arguments{
\item{.dir}{Directory to install into}

\item{.bbi_url}{Full url to download tarball from}
}
\description{
Private helper function for building commands to install bbi on Linux
}
\keyword{internal}
