% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-run-utils.R
\name{make_analysis_model}
\alias{make_analysis_model}
\title{Set up a single bootstrap or SSE model run}
\usage{
make_analysis_model(mod_path, metadata)
}
\arguments{
\item{mod_path}{Absolute model path (no file extension) of a model run.}

\item{metadata}{List of parameters needed to create a model run.}
}
\description{
Set up a single bootstrap or SSE model run
}
\keyword{internal}
