% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-run-utils.R
\name{make_analysis_spec}
\alias{make_analysis_spec}
\title{Store bootstrap run details before submission}
\usage{
make_analysis_spec(run_models, metadata)
}
\arguments{
\item{run_models}{List of model objects created by \code{make_analysis_model()}.}

\item{metadata}{List of parameters needed to create a model run.}
}
\description{
Store bootstrap run details before submission
}
\details{
This is mainly meant to ensure traceability and enhance the portability of
bootstrap "model" objects.
}
\keyword{internal}
