% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{make_boot_mod}
\alias{make_boot_mod}
\title{Create a bootstrap model object}
\usage{
make_boot_mod(boot_runs, boot_args)
}
\arguments{
\item{boot_runs}{list of boostrap model objects created by \code{make_boot_run()}.}

\item{boot_args}{list of parameters needed to create a bootstrap model run.}
}
\description{
Create a bootstrap model object
}
\details{
The ctl is mainly meant to ensure traceability and allow for existing \code{bbr}
functionality. A control file (ctl extension) is not unique to \code{NONMEM},
and this one intentionally has no interaction with \code{NONMEM}/\code{bbi}.

\code{nmrec} is used elsewhere to extract this information to manage model file
and data path expectations.
}
\keyword{internal}
