% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{make_boot_run}
\alias{make_boot_run}
\title{Set up a single bootstrap model run}
\usage{
make_boot_run(mod_path, boot_args)
}
\arguments{
\item{mod_path}{Absolute model path (no file extension) of a bootstrap model run.}

\item{boot_args}{List of parameters needed to create a bootstrap model run.}
}
\description{
Set up a single bootstrap model run
}
\keyword{internal}
