% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{make_boot_spec}
\alias{make_boot_spec}
\title{Store bootstrap run details before submission}
\usage{
make_boot_spec(boot_models, boot_args)
}
\arguments{
\item{boot_models}{List of bootstrap model objects created by \code{make_boot_run()}.}

\item{boot_args}{List of parameters needed to create a bootstrap model run.}
}
\description{
Store bootstrap run details before submission
}
\details{
This is mainly meant to ensure traceability and enhance the portability of
bootstrap "model" objects.
}
\keyword{internal}
