% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{make_sim_spec}
\alias{make_sim_spec}
\title{Store simulation run details before submission}
\usage{
make_sim_spec(.mod, sim_args, .overwrite = FALSE)
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} or \code{bbi_nmsim_model} object.}

\item{sim_args}{named list defining the \code{seed}, and \code{n} simulations.}
}
\description{
Store simulation run details before submission
}
\keyword{internal}
