% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{make_tree_data}
\alias{make_tree_data}
\title{Construct dataset for use in \code{model_tree}}
\usage{
make_tree_data(
  .log_df,
  include_info = c("description", "star", "tags"),
  add_summary = TRUE
)
}
\arguments{
\item{.log_df}{a \code{bbr} run log}

\item{include_info}{vector of columns present in \code{.log_df} to include in the tooltip or
appended table.}

\item{add_summary}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, include key columns
from \code{\link[=model_summary]{model_summary()}} output.}
}
\value{
a dataframe
}
\description{
Construct dataset for use in \code{model_tree}
}
\details{
This function does the following things:
\itemize{
\item Performs checks on required columns
\item Unnests nested columns (such as \code{based_on} and \code{tags})
\itemize{
\item Both character (\code{'Two CMT'}) and list (\code{AMT: nmol}) tags are supported
}
\item Formats columns and performs prep work for creating a tooltip
\item Creates a model network using the \code{based_on} column. If additional
based_on models were used, the first one will be used to create the network;
the rest will be stored in a separate \code{addl_based_on} column.
\itemize{
\item This is represented via the \code{from} and \code{to} columns.
}
}
}
\keyword{internal}
