% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{make_tree_tooltip}
\alias{make_tree_tooltip}
\title{Create tooltip for interactive \code{\link[=model_tree]{model_tree()}}}
\usage{
make_tree_tooltip(.network_df)
}
\description{
Create tooltip for interactive \code{\link[=model_tree]{model_tree()}}
}
\keyword{internal}
