% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_list_recursive}
\alias{map_list_recursive}
\title{Recursively apply a function to each list element}
\usage{
map_list_recursive(.list, .func, .overwrite = TRUE)
}
\arguments{
\item{.list}{a list. Can be recursive}

\item{.func}{a function to apply to a list element}

\item{.overwrite}{whether or not to overwrite the list element with the specified function (otherwise just print)}
}
\description{
Recursively apply a function to each list element
}
\keyword{internal}
