% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-estimates.R
\name{matrix_to_df}
\alias{matrix_to_df}
\title{Format OMEGA or SIGMA matrix as data.frame}
\usage{
matrix_to_df(mat, type = c("omega", "sigma"))
}
\arguments{
\item{mat}{a matrix}

\item{type}{one of \code{c("omega","sigma")}}
}
\description{
Format OMEGA or SIGMA matrix as data.frame
}
\keyword{internal}
