% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-status.R
\name{mod_list_setup}
\alias{mod_list_setup}
\title{Coerce model object to a list of models for use in \code{get_model_status} and
\code{wait_for_nonmem}}
\usage{
mod_list_setup(.mod)
}
\arguments{
\item{.mod}{a \code{bbi_base_model} or list of \code{bbi_base_model} objects. Other
packages (e.g., \code{bbr.bayes}) may add additional methods.}
}
\description{
Coerce model object to a list of models (if not one already) for use in
\code{get_model_status} and \code{wait_for_nonmem}. For \code{bbi_nmboot_model} objects, the
list of individual models corresponding to that run will be returned.
}
\details{
\code{.mod} is a list of models regardless of input after this section
\itemize{
\item Support a single bbi_nonmem_model, a single bbi_nmboot_model, or a list
of bbi_base_models (bbi_nonmem_model or bbi_nmboot_model)
}
}
\keyword{internal}
