% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-handling.R
\name{mod_matrix}
\alias{mod_matrix}
\title{Modify a matrix to be a variance-covariance matrix, or the inverse operation}
\usage{
mod_matrix(mat, mat_opts, inverse = FALSE)
}
\arguments{
\item{mat}{a matrix}

\item{mat_opts}{tibble of matrix options}

\item{inverse}{Logical. If \code{TRUE}, perform the inverse operation. This assumes
the matrix is currently a variance-covariance matrix.}
}
\description{
Modify a matrix to be a variance-covariance matrix, or the inverse operation
}
\keyword{internal}
