% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diff.R
\name{model_diff}
\alias{model_diff}
\alias{model_diff.bbi_nonmem_model}
\title{Compare model files between models}
\usage{
model_diff(.mod, .mod2 = NULL, .file = "model", ..., .viewer = FALSE)

\method{model_diff}{bbi_nonmem_model}(.mod, .mod2 = NULL, .file = c("model"), ..., .viewer = FALSE)
}
\arguments{
\item{.mod}{The \verb{bbi_\{.model_type\}_model} to compare.}

\item{.mod2}{If a \verb{bbi_\{.model_type\}_model} object is passed, compare \code{.mod} to
\code{.mod2}. If \code{.mod2 = NULL}, the default, compare \code{.mod} to the model at
\code{get_based_on(.mod)}. See "\code{based_on} details" in Details section.}

\item{.file}{Defaults to \code{"model"} which compares the default model file for
that model type. For NONMEM models, the control stream is compared. Other
model types may accept additional values and support comparing multiple
files that can be compared. See "\code{.file} argument" of the specific S3
method for details.}

\item{...}{arguments passed through to methods. (Currently none.)}

\item{.viewer}{If \code{FALSE}, the default, prints diff to console or renders in
Rmd. If \code{TRUE}, render the diff in the Viewer window. Note: this option
will hold the console with \verb{Press ENTER to continue...} so it is only
recommended for interactive use.}
}
\value{
Returns a \code{"Diff"} object from the \code{diffobj} package that renders
when printed or called in the console.
}
\description{
By default, this compares a model's model file on disk to the model file of
its "parent" model (the model referenced in its \code{based_on} field).  It can
optionally take a second model (passed to \code{.mod2}), in which case it
will compare the first model to the second model, ignoring the \code{based_on}
field entirely.
}
\details{
\strong{Printing vs. knitting:} When using this function in an \code{.Rmd} file that
you intend to knit to HTML, be sure to \strong{set the chunk option \code{results = "asis"}} for the relevant chunk. If you do this, the \code{model_diff()} output
will render nicely in color in the HTML doc. Note \strong{this only works for HTML
output.} If you are are rendering to a different format (PDF, DOCX, etc.)
you can still use \code{model_diff()} \emph{without} \code{results = "asis"} and it will
render, although it won't look particularly nice.

\strong{\code{based_on} details:} As described above, if \code{.mod2} is \code{NULL} (the default),
then the \code{.mod} will be compared to the model returned by \code{get_based_on(.mod)}.
However, \emph{if there is more than one model} (or no models) in \code{.mod$based_on}, then user
will be prompted to explicitly pass the model they want to compare against to
the \code{.mod2} argument.
}
