% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diff.R
\name{model_diff_impl}
\alias{model_diff_impl}
\title{Private helper to diff two model files}
\usage{
model_diff_impl(model_A, model_B, .viewer)
}
\arguments{
\item{model_A}{Path to first model file}

\item{model_B}{Path to second model file}

\item{.viewer}{If \code{FALSE}, the default, prints diff to console or renders in
Rmd. If \code{TRUE}, render the diff in the Viewer window. Note: this option
will hold the console with \verb{Press ENTER to continue...} so it is only
recommended for interactive use.}
}
\description{
Private helper to diff two model files
}
\keyword{internal}
