% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_is_finished}
\alias{model_is_finished}
\title{Check if model run has finished}
\usage{
model_is_finished(.mod)
}
\value{
Logical (\code{TRUE}/\code{FALSE})
}
\description{
This function coerces the 3 statuses to a logical value.
}
\keyword{internal}
