% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-summaries.R
\name{model_summaries_concurrent}
\alias{model_summaries_concurrent}
\title{Summarize multiple models via single call to bbi}
\usage{
model_summaries_concurrent(.mods, .bbi_args, .fail_flags)
}
\description{
\verb{bbi nonmem summary} accepts more than one model as of v2.3.0, and v3.2.0
reworked the output so that the errors are included in the JSON structure.
This function consumes the v3.2.0 output, returning an object that can be
passed to \code{\link[=create_summary_list]{create_summary_list()}}.
}
\keyword{internal}
