% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{model_summary.bbi_nonmem_result}
\alias{model_summary.bbi_nonmem_result}
\title{Get model summary from `bbi_nonmem_result` object and optionally check if model is done before erroring}
\usage{
\method{model_summary}{bbi_nonmem_result}(.res, .wait = 30, .ext_wait = 30, ...)
}
\arguments{
\item{.res}{`bbi_nonmem_result` object for summary}

\item{.wait}{Integer number of seconds to wait trying `check_nonmem_progress()` for the model to be done.}

\item{.ext_wait}{Integer number of seconds to wait for an .ext file (i.e. for the model to start). Passed through to `check_nonmem_progress()`}

\item{...}{arguments passed through to `nonmem_summary()`}
}
\description{
Get model summary from `bbi_nonmem_result` object and optionally check if model is done before erroring
}
