% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{model_tree}
\alias{model_tree}
\alias{model_tree.character}
\alias{model_tree.bbi_log_df}
\title{Create a tree diagram of a modeling directory}
\usage{
model_tree(
  log_df,
  include_info = c("description", "star", "tags"),
  color_by = "run",
  add_summary = TRUE,
  static = FALSE
)

\method{model_tree}{character}(
  log_df,
  include_info = c("description", "star", "tags"),
  color_by = "run",
  add_summary = TRUE,
  static = FALSE
)

\method{model_tree}{bbi_log_df}(
  log_df,
  include_info = c("description", "star", "tags"),
  color_by = "run",
  add_summary = TRUE,
  static = FALSE
)
}
\arguments{
\item{log_df}{a \code{bbr} run log or a base directory to look in for models.}

\item{include_info}{vector of \code{\link[=run_log]{run_log()}} columns to include in the tooltip or
appended table.}

\item{color_by}{a run log column to color the nodes by. Can be helpful for
identifying which models are starred, have heuristics, etc. Defaults
to \code{"run"}.}

\item{add_summary}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, include key columns
from \code{\link[=model_summary]{model_summary()}} output.}

\item{static}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, render the plot as a
static image. This takes a little longer, as the interactive plot must
be saved as a PNG and loaded into the viewer.}
}
\description{
Create a tree diagram of a modeling directory
}
\details{
Uses the \code{based_on} attribute of each model to determine the tree network.
\itemize{
\item Additional \code{based_on} flags will be shown in the tooltip, using the first one
to create the tree network
}
}
