% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{model_tree_png}
\alias{model_tree_png}
\title{Save interactive \code{\link[=model_tree]{model_tree()}} as PNG and load in Rstudio viewer}
\usage{
model_tree_png(widget)
}
\description{
This function first saves and renders widget as HTML. It then screenshots
the webpage and saves the file to a temporary PNG.
}
\keyword{internal}
