% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{modify_based_on}
\alias{modify_based_on}
\alias{add_based_on}
\alias{replace_all_based_on}
\alias{remove_based_on}
\title{Modify based_on field in a model object}
\usage{
add_based_on(.mod, .based_on)

replace_all_based_on(.mod, .based_on)

remove_based_on(.mod, .based_on)
}
\arguments{
\item{.mod}{The \verb{bbi_\{.model_type\}_model} object to modify}

\item{.based_on}{Character vector of relative paths to add or remove}
}
\value{
The modified \verb{bbi_\{.model_type\}_model} object
}
\description{
Add, replace, or remove model identifiers from the \code{based_on}
field of a model object and corresponding YAML. The \code{based_on} field can be
modified at any time (i.e. before \emph{or} after a model is submitted).
}
\details{
The \code{based_on} field on a \verb{bbi_\{.model_type\}_model} object contains a
character vector identifying models which preceded \code{.mod} in the model
development process. This field is automatically populated with \code{.parent_mod}
when a model is created with \code{\link[=copy_model_from]{copy_model_from()}}.

The model identifiers in the \code{based_on} field are paths another model file
(without file extension) \emph{relative} to the location of \code{.mod}.

There is a \href{https://metrumresearchgroup.github.io/bbr/articles/using-based-on.html}{"Using the based_on field" vignette}
which demonstrates the motivation and usage of this field.
}
\section{Functions}{
\itemize{
\item \code{replace_all_based_on}: Replaces entire \code{based_on} field in a model object and corresponding YAML with new values.

\item \code{remove_based_on}: Remove specified \code{based_on} identifier(s) from a model object and corresponding YAML.
}}

\seealso{
\code{\link[=copy_model_from]{copy_model_from()}} \code{\link[=modify_tags]{modify_tags()}} \code{\link[=modify_notes]{modify_notes()}} \code{\link[=modify_description]{modify_description()}} \code{\link[=modify_bbi_args]{modify_bbi_args()}}
}
