% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{modify_bbi_args}
\alias{modify_bbi_args}
\alias{add_bbi_args}
\alias{replace_all_bbi_args}
\title{Modify bbi_args on a model object}
\usage{
add_bbi_args(.mod, .bbi_args)

replace_all_bbi_args(.mod, .bbi_args)
}
\arguments{
\item{.mod}{The \verb{bbi_\{.model_type\}_model} object to modify}

\item{.bbi_args}{named list of arguments to add to the model. See
\code{\link[=print_bbi_args]{print_bbi_args()}} for valid options.}
}
\value{
The modified \verb{bbi_\{.model_type\}_model} object
}
\description{
Add or replace the \code{bbi_args} field on a model object and corresponding YAML.
}
\details{
The \code{bbi_args} field on a \verb{bbi_\{.model_type\}_model} object contains a named
list of arguments that will be passed through to \code{bbi} to modify how a model
is processed (for instance, when it is submitted to be run with
\code{\link[=submit_model]{submit_model()}}). In many cases, it is preferable to store these arguments
with the model object, as opposed to passing them to
\code{submit_model(.bbi_args)} because then they will be applied every time that
model is run, for example if you need to re-run the model later.

The options for what can be included in this list (and a brief description of
what each does) can be seen by calling \code{\link[=print_bbi_args]{print_bbi_args()}}.

Note that there is no \code{replace_bbi_arg} (singular) to replace individual arguments.
This is because you can use \code{add_bbi_args()} and any arguments passed in will automatically
overwrite any previous value stored for that argument.
}
\section{Functions}{
\itemize{
\item \code{add_bbi_args}: Modifies model object and corresponding YAML
by adding named list passed to \code{.bbi_args}, overwriting any args that are
already present with the new values.

\item \code{replace_all_bbi_args}: Modifies model object and corresponding YAML
by replacing all \code{bbi_args} with named list passed to \code{.bbi_args}.
}}

\seealso{
\code{\link[=submit_model]{submit_model()}} \code{\link[=model_summary]{model_summary()}} \code{\link[=print_bbi_args]{print_bbi_args()}}
\code{\link[=modify_tags]{modify_tags()}} \code{\link[=modify_notes]{modify_notes()}} \code{\link[=modify_based_on]{modify_based_on()}} \code{\link[=modify_description]{modify_description()}} \code{\link[=modify_star]{modify_star()}}
}
