% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-records.R
\name{modify_data_path_ctl}
\alias{modify_data_path_ctl}
\title{Modify the specified data path in a control stream file}
\usage{
modify_data_path_ctl(mod, data_path)
}
\arguments{
\item{mod}{a bbr model object}

\item{data_path}{Data path to set in a \verb{$DATA} record.}
}
\description{
Modify the specified data path in a control stream file
}
\keyword{internal}
