% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{modify_data_str}
\alias{modify_data_str}
\title{Helper function to modify the raw $DATA and $INPUT lines in a control stream}
\usage{
modify_data_str(.mod, .new_data_path, .new_input = NULL)
}
\description{
Helper function to modify the raw $DATA and $INPUT lines in a control stream
}
