% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{modify_decisions}
\alias{modify_decisions}
\alias{add_decisions}
\alias{replace_decisions}
\title{Deprecated: Modify decisions on a model object}
\usage{
add_decisions(.mod, .decisions)

replace_decisions(.mod, .decisions)
}
\arguments{
\item{.mod}{The \verb{bbi_\{.model_type\}_model} object to modify}

\item{.decisions}{Character vector to add to \code{decisions} field}
}
\value{
The modified \verb{bbi_\{.model_type\}_model} object
}
\description{
The \code{decisions} field has been deprecated as of \verb{bbr 0.10.0} and
replaced by the \code{notes} field, to reflect the fact that users
will want to use this field throughout the modeling process, not only at the end
once some "decisions" have been reached. \code{add_decisions()} and
\code{replace_decisions()} now print a warning telling the user that they will be
deprecated in the future and encouraging use of their \verb{*_notes} counterparts.
They will begin erroring, instead of warning, in two releases, and then will be
removed two releases after that.
}
\section{Functions}{
\itemize{
\item \code{add_decisions}: \strong{Deprecated}  Append new decisions to the one(s) in a model object and corresponding YAML.

\item \code{replace_decisions}: \strong{Deprecated} Replaces \code{decisions} field in a model object and corresponding YAML with new values.
}}

\seealso{
\code{\link[=modify_notes]{modify_notes()}}
}
