% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{modify_description}
\alias{modify_description}
\alias{add_description}
\alias{replace_description}
\title{Modify description on a model object}
\usage{
add_description(.mod, .description)

replace_description(.mod, .description)
}
\arguments{
\item{.mod}{The \verb{bbi_\{.model_type\}_model} object to modify}

\item{.description}{Character scalar to put in \code{description} field}
}
\value{
The modified \verb{bbi_\{.model_type\}_model} object
}
\description{
Add or replace a description on a model object and corresponding YAML.
The description can be modified at any time (i.e. before \emph{or} after a model is submitted).
}
\details{
The \code{description} field on a \verb{bbi_\{.model_type\}_model} object contains a character
scalar with a brief description of the model. Note that it is \emph{not} a required field
and some users prefer to leave it blank and instead use the \code{tags} and/or \code{notes} fields
to describe their models instead.

Another pattern is to fill the \code{description} field \emph{only} on notable models,
for instance with \code{"Base Model"} or \code{"Final Model"}. This can be useful when
looking at a \code{bbi_run_log_df} (the tibble output from \code{\link[=run_log]{run_log()}}); for
example to look at only the most important models by calling
\code{run_log() \%>\% filter(!is.na(description))}.
}
\section{Functions}{
\itemize{
\item \code{add_description()}: Fills the description field in a model object and corresponding YAML, if it is currently empty.

\item \code{replace_description()}: Replaces description field in a model object and corresponding YAML with new description.

}}
\seealso{
\code{\link[=run_log]{run_log()}} \code{\link[=modify_tags]{modify_tags()}} \code{\link[=modify_notes]{modify_notes()}} \code{\link[=modify_based_on]{modify_based_on()}} \code{\link[=modify_bbi_args]{modify_bbi_args()}} \code{\link[=modify_star]{modify_star()}}
}
