% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{modify_model_field_lgl}
\alias{modify_model_field_lgl}
\title{Adds or removes boolean attribute from model object and YAML}
\usage{
modify_model_field_lgl(.mod, .field, .action = c("add", "remove"))
}
\arguments{
\item{.mod}{list object of model data}

\item{.field}{boolean attribute to be set}

\item{.action}{list of boolean variables to be set}
}
\description{
Adds or removes boolean attribute from model object and YAML
}
\keyword{internal}
