% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{modify_notes}
\alias{modify_notes}
\alias{add_notes}
\alias{replace_note}
\alias{replace_all_notes}
\alias{remove_notes}
\title{Modify notes on a model object}
\usage{
add_notes(.mod, .notes)

replace_note(.mod, .old_note, .new_note)

replace_all_notes(.mod, .notes)

remove_notes(.mod, .notes)
}
\arguments{
\item{.mod}{The \verb{bbi_\{.model_type\}_model} object to modify}

\item{.notes}{Character vector of notes to add or remove}

\item{.old_note}{Character scalar of note to be replaced}

\item{.new_note}{Character scalar of note that will be added}
}
\value{
The modified \verb{bbi_\{.model_type\}_model} object
}
\description{
Add, replace, or remove notes on a model object and corresponding YAML.
Notes can be modified at any time (i.e. before \emph{or} after a model is submitted).
}
\details{
The \code{notes} field on a \verb{bbi_\{.model_type\}_model} object contains a character
vector of brief notes about the model or modeling process.

These are often useful when looking at a \code{bbi_run_log_df} (the tibble output
from \code{\link[=run_log]{run_log()}}); for example to remind yourself of decisions you made along
the way, or of a particular thing you noticed when looking at model
diagnostics.

Like tags, notes can also be collapsed using \code{\link[=collapse_to_string]{collapse_to_string()}}
See the \href{https://metrumresearchgroup.github.io/rbabylon/articles/getting-started.html#viewing-tags-example}{"Getting Started" vignette}
for an example of this.
}
\section{Functions}{
\itemize{
\item \code{add_notes}: Add notes to a model object and corresponding YAML.

\item \code{replace_note}: Replaces a specific \code{.old_note} with \code{.new_note} on a model object and corresponding YAML.
Warns and does nothing if \code{.old_note} is not present.

\item \code{replace_all_notes}: Replaces all notes on a model object and corresponding YAML with new notes.

\item \code{remove_notes}: Removes notes from a model object and corresponding YAML.
}}

\seealso{
\code{\link[=run_log]{run_log()}} \code{\link[=collapse_to_string]{collapse_to_string()}} \code{\link[=modify_tags]{modify_tags()}} \code{\link[=modify_based_on]{modify_based_on()}} \code{\link[=modify_description]{modify_description()}} \code{\link[=modify_bbi_args]{modify_bbi_args()}}
}
