% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-records.R
\name{modify_prob_statement}
\alias{modify_prob_statement}
\title{Modify or retrieve the \verb{$PROBLEM} statement from a \code{NONMEM} control stream
file.}
\usage{
modify_prob_statement(mod, prob_text = NULL)
}
\arguments{
\item{mod}{a bbr model object}

\item{prob_text}{If \code{NULL} return the current \verb{$PROB} statement. If a character
string, set the problem statement to that value.}
}
\description{
Modify or retrieve the \verb{$PROBLEM} statement from a \code{NONMEM} control stream
file.
}
\keyword{internal}
