% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{modify_record_opt}
\alias{modify_record_opt}
\title{Modify an \code{nmrec} record option}
\usage{
modify_record_opt(val_recs, new_values, .bounds_opts)
}
\description{
Modify an \code{nmrec} record option
}
\keyword{internal}
