% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{modify_star}
\alias{modify_star}
\alias{add_star}
\alias{remove_star}
\title{Modify star attribute on a model object}
\usage{
add_star(.mod)

remove_star(.mod)
}
\arguments{
\item{.mod}{The \verb{bbi_\{.model_type\}_model} object to modify}
}
\value{
The modified \verb{bbi_\{.model_type\}_model} object
}
\description{
Add, or remove a "star" to a model to indicate special interest level for this model.
This is typically used for highlighting models that are of some importance to the final analysis.
}
\section{Functions}{
\itemize{
\item \code{add_star()}: Adds a star to the model.

\item \code{remove_star()}: Removes a star from the model.

}}
\seealso{
\code{\link[=run_log]{run_log()}} \code{\link[=collapse_to_string]{collapse_to_string()}} \code{\link[=modify_tags]{modify_tags()}} \code{\link[=modify_notes]{modify_notes()}}
\code{\link[=modify_based_on]{modify_based_on()}} \code{\link[=modify_description]{modify_description()}} \code{\link[=modify_bbi_args]{modify_bbi_args()}}
\code{\link[=modify_star]{modify_star()}}
}
