% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{na_to_null}
\alias{na_to_null}
\title{Convert an object containing entirely NA's to NULL}
\usage{
na_to_null(.obj)
}
\arguments{
\item{.obj}{Object to check for NA}
}
\description{
Convert an object containing entirely NA's to NULL
}
\keyword{internal}
