% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-run-utils.R
\name{new_analysis_run}
\alias{new_analysis_run}
\title{Create a bootstrap or SSE run from an existing model}
\usage{
new_analysis_run(
  .mod,
  .suffix,
  .type = c("nmboot", "nmsse"),
  .add_tags = NULL,
  .inherit_tags = TRUE,
  .overwrite = FALSE,
  remove_cov = TRUE,
  remove_tables = TRUE
)
}
\arguments{
\item{.mod}{A \code{bbr} model object.}

\item{.suffix}{A suffix for the new run directory. Will be prefixed by
the model id of \code{.mod}.}

\item{.type}{Either \code{"nmboot"} or \code{"nmsse"}}

\item{.add_tags}{Character vector with any new tags(s) to be added to
\verb{\{.new_model\}.yaml}}

\item{.inherit_tags}{If \code{TRUE}, the default, inherit any tags from \code{.mod}.}

\item{.overwrite}{If \code{FALSE}, the default,  function will error if a model
file already exists at specified \code{.new_model} path. If \code{TRUE} any existing
file at \code{.new_model} will be overwritten silently.}

\item{remove_cov, remove_tables}{If \code{TRUE}, the default, remove \verb{$COVARIANCE}
and \verb{$TABLE} records respectively, allowing for notably faster run times.}
}
\description{
Create a bootstrap or SSE run from an existing model
}
\keyword{internal}
