% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{new_bootstrap_run}
\alias{new_bootstrap_run}
\title{Create a boostrap model object from an existing model}
\usage{
new_bootstrap_run(
  .mod,
  .suffix = glue("{get_model_id(.mod)}_boot"),
  .inherit_tags = TRUE,
  .overwrite = FALSE,
  increment = FALSE
)
}
\arguments{
\item{.mod}{a \code{bbr} model object}

\item{.suffix}{a prefix for the boostrap model directory. Will be appended by
the boostrap run number for a given model.}

\item{.inherit_tags}{If \code{FALSE}, the default, new model will only have any
tags passed to \code{.add_tags} argument. If \code{TRUE} inherit any tags from
\code{.parent_mod}, with any tags passed to \code{.add_tags} appended.}

\item{.overwrite}{If \code{FALSE}, the default,  function will error if a model
file already exists at specified \code{.new_model} path. If \code{TRUE} any existing
file at \code{.new_model} will be overwritten silently.}

\item{increment}{Logical (T/F). If \code{TRUE}, create a new bootstrap model run.
Will append \verb{_run_[x]} to the provided \code{.suffix}. \code{.overwrite} will have no
impact if incrementing the bootstrap run.}
}
\description{
Create a boostrap model object from an existing model
}
