% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{new_vpc_sim}
\alias{new_vpc_sim}
\title{Create a VPC model object from an existing model}
\usage{
new_vpc_sim(
  .mod,
  n = 100,
  seed = 1234,
  .suffix = "vpc",
  .inherit_tags = TRUE,
  .overwrite = FALSE
)
}
\arguments{
\item{.mod}{a \code{bbr} model object}

\item{n}{number of replicates.}

\item{seed}{a seed for sampling the data.}

\item{.suffix}{a suffix for the VPC simulation directory. Will be prefixed by
the model id of \code{.mod}.}

\item{.inherit_tags}{If \code{FALSE}, the default, new model will only have any
tags passed to \code{.add_tags} argument. If \code{TRUE} inherit any tags from
\code{.parent_mod}, with any tags passed to \code{.add_tags} appended.}

\item{.overwrite}{If \code{FALSE}, the default,  function will error if a model
file already exists at specified \code{.new_model} path. If \code{TRUE} any existing
file at \code{.new_model} will be overwritten silently.}
}
\value{
S3 object of class \code{bbi_vpc_model}.
}
\description{
Create a VPC model object from an existing model
}
