% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-file.R
\name{nm_candidates}
\alias{nm_candidates}
\title{Read in the file that exists}
\usage{
nm_candidates(.func, .candidates)
}
\arguments{
\item{.func}{name of the calling function (for potential error message)}

\item{.candidates}{Files to look for}
}
\description{
Helper for \code{\link[=nm_tab]{nm_tab()}} and \code{\link[=nm_par_tab]{nm_par_tab()}}. Looks for all files passed to
\code{.candidates}. If only one of them exists, it reads that file. Otherwise it
errors informatively.
}
\keyword{internal}
