% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-tran-data.R
\name{nm_data_as_nmtran}
\alias{nm_data_as_nmtran}
\title{Format input data as an \code{NM-TRAN} dataset}
\usage{
nm_data_as_nmtran(.mod, data = nm_data(.mod))
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} object}

\item{data}{a starting dataset}
}
\description{
Format input data as an \code{NM-TRAN} dataset
}
\keyword{internal}
