% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-records.R
\name{nm_data_drop_records}
\alias{nm_data_drop_records}
\title{Drop or skip records based on \verb{$INPUT} record options}
\usage{
nm_data_drop_records(.mod, data = nm_data(.mod))
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} object}

\item{data}{a starting dataset}
}
\description{
Drop or skip records based on \verb{$INPUT} record options
}
\keyword{internal}
