% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-tran-data.R
\name{nm_data_filter}
\alias{nm_data_filter}
\title{Filter \code{NONMEM} input data based on \code{IGNORE} and \code{ACCEPT} record options}
\usage{
nm_data_filter(.mod, data = nm_data(.mod))
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} object}

\item{data}{a starting dataset}
}
\description{
Filter \code{NONMEM} input data based on \code{IGNORE} and \code{ACCEPT} record options
}
\keyword{internal}
