% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-tran-data.R
\name{nm_data_null_map}
\alias{nm_data_null_map}
\title{Apply \code{null} mapping}
\usage{
nm_data_null_map(.mod, data = nm_data(.mod))
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} object}

\item{data}{a starting dataset}
}
\description{
Apply \code{null} mapping
}
\keyword{internal}
