% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-file.R
\name{nm_file}
\alias{nm_file}
\alias{nm_grd}
\alias{nm_ext}
\alias{nm_tab}
\alias{nm_par_tab}
\alias{nm_data}
\title{Read NONMEM files}
\usage{
nm_file(.mod, .suffix = NULL, .est_method = NULL, ...)

nm_grd(.mod, .est_method = NULL)

nm_ext(.mod, .est_method = NULL)

nm_tab(.mod)

nm_par_tab(.mod)

nm_data(.mod, .sep = ",")
}
\arguments{
\item{.mod}{Either a \code{bbi_nonmem_model}, \code{bbi_nonmem_summary}, or a path to a
file to read in. If passing model object to \code{nm_file()}, must also pass \code{.suffix} that
will be passed through to \code{\link[=build_path_from_model]{build_path_from_model()}}.}

\item{.suffix}{Character vector to append the end of the absolute model path.
Will be appended \emph{as is} so, if passing a file extension, be sure to included the leading \code{"."}.
See examples.}

\item{.est_method}{If \code{NULL}, the default, pulls the data from the final
estimation method. If an integer, pulls the data from that estimation
method.}

\item{...}{arguments passed through to methods. (Currently none.)}

\item{.sep}{Single character used to separate fields within a record. Passed
through to \code{readr::read_delim()}. Defaults to \code{","}.}
}
\value{
A tibble with the data from the specified file and estimation method.
}
\description{
Reads in a whitespace-delimited NONMEM output file (for example .grd or .ext
or a table output) or a NONMEM input data file.
}
\section{Functions}{
\itemize{
\item \code{nm_grd}: Reads \code{.grd} file from a \code{bbi_nonmem_model} or
\code{bbi_nonmem_summary} object

\item \code{nm_ext}: Reads \code{.ext} file from a \code{bbi_nonmem_model} or
\code{bbi_nonmem_summary} object

\item \code{nm_tab}: Looks for \verb{\{get_model_id(.mod)\}.tab} or \code{TAB} file in
NONMEM output folder and reads whichever is found.

\item \code{nm_par_tab}: Looks for \verb{\{get_model_id(.mod)\}par.tab} or \code{PAR_TAB} file
in NONMEM output folder and reads whichever is found.

\item \code{nm_data}: Reads the input data file from a \code{bbi_nonmem_model} or
\code{bbi_nonmem_summary} object
}}

