% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-file.R
\name{nm_file_impl}
\alias{nm_file_impl}
\title{Implementation function for reading NONMEM files}
\usage{
nm_file_impl(.path, skip = 1)
}
\arguments{
\item{.path}{a path to a table file.}

\item{skip}{number of rows to skip when reading in table file. Defaults to \code{1}}
}
\description{
Implementation function for reading NONMEM files
}
\keyword{internal}
