% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-file.R
\name{nm_file_multi_tab}
\alias{nm_file_multi_tab}
\title{Read in table file with multiple \code{NONMEM} tables.}
\usage{
nm_file_multi_tab(.path, simplify = TRUE, table_pattern = "TABLE NO")
}
\arguments{
\item{.path}{a path to a table file.}

\item{simplify}{Logical (T/F). If \code{TRUE}, attempt to simplify list of tables into a
single tibble (requires . Table names will show up as a new \code{table_id} column. This is
automatically done if simulation data is detected (same rows \emph{and} columns per
table).}

\item{table_pattern}{character string defining the start of a new
table (regex accepted).}
}
\description{
Read in table file with multiple \code{NONMEM} tables.
}
\note{
table headers are expected
}
