% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-file.R
\name{nm_file_multi_table}
\alias{nm_file_multi_table}
\title{Read in table file with multiple \code{NONMEM} tables.}
\usage{
nm_file_multi_table(
  .path,
  header = TRUE,
  simplify = TRUE,
  sanitize = TRUE,
  ...,
  table_pattern = "^TABLE NO"
)
}
\arguments{
\item{.path}{a path to a table file.}

\item{header}{Logical (T/F). Does the first data line contain column names?
Passed to \code{\link[data.table:fread]{data.table::fread()}}.}

\item{simplify}{Logical (T/F). If \code{TRUE}, simplify list of tables into a
single tibble. Table names will show up as a new \code{table_id} column.}

\item{sanitize}{Logical (T/F). If \code{TRUE}, sanitize columns to ensure no
remaining column headers or other text pollution makes it into the data.}

\item{...}{additional arguments passed to \code{\link[data.table:fread]{data.table::fread()}}}

\item{table_pattern}{character string defining the start of a new
table (regex accepted).}
}
\description{
Read in table file with multiple \code{NONMEM} tables.
}
\note{
This is modified from pmxTools, which is licensed under GPL-2.
}
\seealso{
tab_is_multi_table
}
\keyword{internal}
