% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-join.R
\name{nm_join}
\alias{nm_join}
\title{Return a single data frame with model output and input data}
\usage{
nm_join(
  .mod,
  .join_col = "NUM",
  .files = tabfiles(.mod),
  .more = NULL,
  .verbose = TRUE,
  .superset = FALSE,
  .bbi_args = NULL
)
}
\arguments{
\item{.mod}{the \code{bbi_nonmem_model} object or a path to a NONMEM run}

\item{.join_col}{character column name to use to join tab files}

\item{.files}{absolute paths to table files to try to join}

\item{.more}{absolute paths to other files to try to join}

\item{.verbose}{if TRUE, the default, messages are printed as files are read}

\item{.superset}{if \code{FALSE}, the default, the data will be joined to the
NONMEM output and if \code{TRUE}, the NONMEM output will be joined to the data;
that is, if you use \code{.superset}, you will get the same number of rows as
you have in the input data and NONMEM output columns like \code{PRED} and
\code{CWRES} will be filled with \code{NA}.}

\item{.bbi_args}{list passed to \code{bbr::model_summary(.bbi_args)}. See
\code{\link[=print_bbi_args]{print_bbi_args()}} for valid options.}
}
\description{
Return a single data frame with model output and input data
}
\details{
Note that you can get the absolute path to the model output directory with
\code{\link[=get_output_dir]{get_output_dir()}} or \code{\link[=build_path_from_model]{build_path_from_model()}} if needed to build paths for
passing to \code{.files} or \code{.more}.
}
