% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{nm_join_sim}
\alias{nm_join_sim}
\title{Read in simulation data}
\usage{
nm_join_sim(.mod, .join_col = "NUM", .suffix = "sim.tab")
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} or \code{bbi_nmsim_model} object.}

\item{.join_col}{Character column name to use to join table files. Defaults to
\code{NUM}. See Details.}

\item{.suffix}{Character vector to append the end of the absolute model path.
Will be appended \emph{as is} so, if passing a file extension, be sure to included the leading \code{"."}.
See examples.}
}
\description{
Read in simulation data
}
