% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-tables.R
\name{nm_tables}
\alias{nm_tables}
\alias{nm_table_files}
\title{Read all tables and input data from NONMEM run}
\usage{
nm_table_files(.mod, .check_exists = TRUE)
}
\arguments{
\item{.check_exists}{If \code{TRUE} check if files exist.}
}
\description{
Read all tables and input data from NONMEM run
}
\section{Functions}{
\itemize{
\item \code{nm_table_files}: Extract paths to table output files from NONMEM control
stream, and optionally check if the files exist.
}}

