% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_model_file}
\alias{open_model_file}
\title{Open main model definition file in text editor}
\usage{
open_model_file(.mod)
}
\arguments{
\item{.mod}{Model object or a file path to a model.}
}
\description{
This is a light wrapper around \code{\link[=get_model_path]{get_model_path()}} to enable quickly jumping
to a model definition file in RStudio or any other editor supported by
\code{\link[utils:file.edit]{utils::file.edit()}}.
}
