% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{param_estimates}
\alias{param_estimates}
\alias{param_estimates.bbi_nonmem_summary}
\title{S3 generic for parsing parameter estimate table}
\usage{
param_estimates(.x, ...)

\method{param_estimates}{bbi_nonmem_summary}(.x)
}
\arguments{
\item{.x}{`bbi_nonmem_summary` object}

\item{...}{args passed through}
}
\description{
S3 generic for parsing parameter estimate table

S3 dispatch for parsing `bbi_nonmem_summary` object into parameter estimate table
}
