% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimates-batch.R
\name{param_estimates_batch}
\alias{param_estimates_batch}
\title{Batch Processing of Parameter Estimates}
\usage{
param_estimates_batch(.path, ..., .dry_run = FALSE)
}
\arguments{
\item{.path}{a path to a directory containing model sudirectories for batch parameter processing.}

\item{...}{args passed through to \code{\link[=bbi_exec]{bbi_exec()}}}

\item{.dry_run}{show what the command would be without actually running it}
}
\description{
Calls out to \verb{bbi nonmem params} and returns a tibble of parameter estimates.
The primary reason for this function is for things like bootstraps or
simulations where parameter estimates for a large number of models (hundreds
or thousands) need to be pulled in together. In those cases, this function is
\emph{much} faster than using \code{\link[=model_summaries]{model_summaries()}} and \code{\link[=param_estimates]{param_estimates()}},
because it does \emph{not} parse all of the other information contained in a
\code{bbi_nonmem_summary} object.
}
\details{
The tibble will always have columns \code{absolute_model_path}, \code{run},
\code{error_msg}, and \code{termination_code}. All other columns names are those of the
parameter estimates found in the \code{.ext} files detected in the directory
passed. Note: the passed directory is \emph{not} searched recursively. Passed
directory will be searched for model files and output directories for found
models will be searched for \code{.ext} files.

\code{error_msg}: column will be \code{NA} if error is not detected. Column will contain message if known error is detected. The known errors
are
\itemize{
\item \code{"no ext file contents"}: \code{.ext} file detected, but contains nothing
\item \code{"no estimation output detected"}: \code{.ext} file detected, has contents, but is missing line \code{-1000000000} which are the final parameter estimates
}

\code{termination_code}: parsed from line \code{-1000000007} of the \code{.ext}

\strong{parameter columns:} naming will be the same as names found in \code{.ext} file.
If \code{.ext} files within the directory contain differing parameter names, the
output tibble will show all parameter names across files and NA values for
the files where a given parameter is missing.
}
