% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-labels.R
\name{param_labels}
\alias{param_labels}
\title{Parse parameter labels from the model object control stream comments}
\usage{
param_labels(.mod)
}
\arguments{
\item{.mod}{`bbi_{.model_type}_model` object}
}
\description{
This is using the labeling scheme from "Census", as defined here: https://ghe.metrumrg.com/pages/software/mrgtable/reference/nm_tbl.html
It will _not_ return indices for the parameters, though they can be added with `param_labels() %>% apply_indices()`
}
