% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_args_list}
\alias{parse_args_list}
\title{Combines NONMEM args that were passed into the function call with args that were parsed from a model yaml}
\usage{
parse_args_list(.func_args, .yaml_args)
}
\arguments{
\item{.func_args}{A named list of arguments for bbi, passed into submit_model function call}

\item{.yaml_args}{A named list of arguments for bbi, parsed from user input yaml}
}
\value{
The combination of the two lists, with .func_args overwriting any keys that are shared
}
\description{
Combines NONMEM args that were passed into the function call with args that were parsed from a model yaml
}
