% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-cor.R
\name{parse_cov_cor_full_file}
\alias{parse_cov_cor_full_file}
\title{Helper to pull full matrix from final estimation method from \code{.cov} or \code{.cor} files}
\usage{
parse_cov_cor_full_file(.mod, .suffix)
}
\arguments{
\item{.mod}{the model object}

\item{.suffix}{either \code{.cov} or \code{.cor}}
}
\description{
Helper to pull full matrix from final estimation method from \code{.cov} or \code{.cor} files
}
\keyword{internal}
