% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-tables.R
\name{parse_ctl_to_list}
\alias{parse_ctl_to_list}
\title{Parses NONMEM control stream to a list of $ blocks.
Adapted from mrgsolve.}
\usage{
parse_ctl_to_list(.path)
}
\arguments{
\item{.path}{File path to control stream}
}
\description{
Parses NONMEM control stream to a list of $ blocks.
Adapted from mrgsolve.
}
\keyword{internal}
