% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_description_version_spec}
\alias{parse_description_version_spec}
\title{Parsing specified version number of a package from the DESCRIPTION file}
\usage{
parse_description_version_spec(.file, .field, .package)
}
\arguments{
\item{.file}{path to DESCRIPTION file}

\item{.field}{Field to parse from, i.e. `"SystemRequirements"` or `"Imports"` etc.}

\item{.package}{The name of the package that you are parsing a version for}
}
\description{
Parsing specified version number of a package from the DESCRIPTION file
}
\keyword{internal}
