% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{parse_mod_yaml}
\alias{parse_mod_yaml}
\title{Parses a model yaml file into a list object that contains correctly formatted information from the yaml}
\usage{
parse_mod_yaml(.path)
}
\arguments{
\item{.path}{Path to the yaml file to parse.}
}
\value{
Output list as specified above.
}
\description{
Parses a model yaml file into a list object that contains correctly formatted information from the yaml
}
