% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-model-results.R
\name{parse_model_results}
\alias{parse_model_results}
\title{Reads model results to a list by parsing json output from bbi summary}
\usage{
parse_model_results(.x, ..., file = NULL)
}
\arguments{
\item{.x}{summary output string}

\item{...}{params to pass to jsonlite::fromJSON}

\item{file}{model json summary output file path}
}
\description{
Reads model results to a list by parsing json output from bbi summary
}
