% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{parse_new_model_dots}
\alias{parse_new_model_dots}
\title{Parse ... passed to new_model()}
\usage{
parse_new_model_dots(.model_type, .path, ...)
}
\description{
Parse ... passed to new_model()
}
\keyword{internal}
