% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-nmtran.R
\name{parse_nmtran_args}
\alias{parse_nmtran_args}
\title{Parse \code{.bbi_args} and return the three expected \code{nmfe_options} for \code{NM-TRAN}
in the correct format}
\usage{
parse_nmtran_args(.mod, .bbi_args = NULL, nmfe_options = NULL)
}
\arguments{
\item{.mod}{A \code{bbi_nonmem_model} object.}

\item{.bbi_args}{A named list specifying arguments to pass to \code{NM-TRAN}.
Similar to the \code{.bbi_args} argument defined in \code{\link[=submit_model]{submit_model()}}, though here
only \code{prdefault}, \code{tprdefault}, and \code{maxlim} arguments are passed to \code{NM-TRAN}.
\code{nm_version} is also supported and specifies which \code{NM-TRAN} executable to use.
See \code{\link[=print_bbi_args]{print_bbi_args()}} for more details.}

\item{nmfe_options}{named list of nmfe options defined in a \code{bbi.yaml} file.}
}
\value{
a named character vector
}
\description{
Parse \code{.bbi_args} and return the three expected \code{nmfe_options} for \code{NM-TRAN}
in the correct format
}
\details{
Combines NONMEM submission args and consolidates to \code{NMFE} arguments only
\itemize{
\item The arguments of interest are \code{prdefault}, \code{tprdefault}, and \code{maxlim},
which impact the evaluation of \code{NM-TRAN}
\item Priority: \code{.bbi_args} > model yaml > \code{bbi.yaml}
}
}
\examples{

\dontrun{

 bbi_yaml_path <- get_bbi_yaml_path(.mod)
 bbi_yaml <- yaml::read_yaml(bbi_yaml_path)

 parse_nmtran_args(
   .mod,
   .bbi_args = list(nm_version = "nm74gf", prdefault = TRUE),
   nmfe_options = bbi_yaml$nmfe_options
 )

}
}
\keyword{internal}
