% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{parse_psn_submission_info}
\alias{parse_psn_submission_info}
\title{Parse \code{version_and_option_info.txt} submission file}
\usage{
parse_psn_submission_info(.submission_file, .psn_model_file)
}
\arguments{
\item{.submission_file}{file path to a \code{"version_and_option_info.txt"} file.}

\item{.psn_model_file}{Model file used during PsN execution. Note that the model file must have a \code{.mod} or \code{.ctl} file extension.}
}
\description{
Parse \code{version_and_option_info.txt} submission file
}
\keyword{internal}
