% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan.R
\name{parse_stanargs}
\alias{parse_stanargs}
\title{Parse args to be passed to cmdstanr}
\usage{
parse_stanargs(.mod, valid_stanargs, ...)
}
\arguments{
\item{.mod}{the \code{bbi_stan_model} object}

\item{...}{The arguments to capture and check}

\item{.valid_stanargs}{A character vector of valid arguments to pass
through to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}}
}
\value{
the named list of parsed and checked args
}
\description{
Any arguments destined for \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}} will be passed in
via \code{...}. This function checks that those are valid and then writes
the resulting list of args to a file for reproducibility checking.
}
\keyword{internal}
