% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-log.R
\name{parse_summary_col}
\alias{parse_summary_col}
\alias{extract_ofv}
\alias{extract_param_count}
\alias{extract_details}
\alias{extract_heuristics}
\title{Parses the list column with the bbi_nonmem_summary object into separate (wide-format) columns}
\usage{
parse_summary_col(.df)

extract_ofv(.s)

extract_param_count(.s)

extract_details(.s)

extract_heuristics(.s)
}
\arguments{
\item{.df}{A tibble with a list column called "summary" containing bbi_nonmem_summary objects}

\item{.s}{`bbi_nonmem_summary` object to parse}
}
\description{
Parses the list column with the bbi_nonmem_summary object into separate (wide-format) columns
}
