% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_user_data}
\alias{parse_user_data}
\title{DOES SOMETHING WITH USER DATA FROM YAML. NOT SURE WHAT THIS WILL BE YET.}
\usage{
parse_user_data(.user_data)
}
\arguments{
\item{.user_data}{A list, parsed from user input yaml, that will be persisted}
}
\description{
DOES SOMETHING WITH USER DATA FROM YAML. NOT SURE WHAT THIS WILL BE YET.
}
