% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{plot_nonmem_table_df}
\alias{plot_nonmem_table_df}
\title{Creates a line plot of the wide-format tibble output from check_nonmem_table_output}
\usage{
plot_nonmem_table_df(.df, .x_var, .stat_name)
}
\arguments{
\item{.df}{the wide-format tibble output from check_nonmem_table_output}

\item{.x_var}{Character scaler for the variable name to use on the X-axis}

\item{.stat_name}{Character scaler for the name of the stat the other columns represents (like "gradient" or "theta").}
}
\description{
Creates a line plot of the wide-format tibble output from check_nonmem_table_output
}
