% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{print.model_tree_static}
\alias{print.model_tree_static}
\title{Draw model tree as a static plot}
\usage{
\method{print}{model_tree_static}(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
\item{x}{plot to display}

\item{newpage}{Logical (T/F). If \code{TRUE}, draw new (empty) page first.}

\item{vp}{viewport to draw plot in}

\item{...}{other arguments not used by this method}
}
\description{
Draw model tree as a static plot
}
\examples{
\dontrun{
pl_tree <- model_tree(MODEL_DIR, static = TRUE)

print(pl_tree)
print(pl_tree, vp = grid::viewport(width=0.5, height=0.5))
print(pl_tree, newpage = TRUE, vp = grid::viewport(width=0.5, height=0.5))
}
}
