% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_model_files}
\alias{print_model_files}
\title{Print a list of files.}
\usage{
print_model_files(.mod, print_fn)
}
\arguments{
\item{.mod}{Model object.}

\item{print_fn}{A function that this method should call to print each file.}
}
\description{
\code{\link[=print.bbi_model]{print.bbi_model()}} calls this after printing the YAML and model path to
allow specific model types to display additional files.
}
