% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_nonmem_args}
\alias{print_nonmem_args}
\title{Print valid .bbi_args}
\usage{
print_nonmem_args()
}
\description{
Prints all valid arguments to pass in to `.bbi_args=list()` argument of `submit_model()` or `model_summary()`
}
