% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{rbabylon}
\alias{rbabylon}
\title{rbabylon}
\description{
rbabylon is a library written in R that facilitates modeling and
simulation by exposing functionality from the
\href{https://github.com/metrumresearchgroup/babylon}{babylon} project in R.
From Babylon's docs: Babylon is a solution for managing projects involving
modeling and simulation with a number of software solutions used in
pharmaceutical sciences. This is a fork of the \code{nonmemutils} project that
is broader in scope. Initial support encompasses NONMEM however the api is
designed in a way to be flexible to handle other software.
}
\details{
RBabylon uses several fundamental objects, each of which has a number of S3
methods associated with it:
\itemize{
\item \verb{bbi_\{.model_type\}_model} -- Represents a single model. Contains paths to
any elements of this model that are stored on disk (e.g. a NONMEM control
stream) as well as various fields containing metadata about the model. You
can read more about these fields, and how to interact with them, in the
following help pages: \code{\link[=modify_tags]{modify_tags()}} \code{\link[=modify_notes]{modify_notes()}} \code{\link[=modify_based_on]{modify_based_on()}}
\code{\link[=modify_description]{modify_description()}} \code{\link[=modify_bbi_args]{modify_bbi_args()}}
\item \verb{bbi_\{.model_type\}_summary} -- Contains information and diagnostics about
a model that has been run. See \code{\link[=model_summary]{model_summary()}} for more information about
contents and usage.
\item \code{bbi_log_df} -- A tibble containing information about multiple models.
There are several sub-classes of \code{bbi_log_df} tibbles, variously containing
metadata, information, and diagnostics about the models. See the following
help pages for more information: \code{\link[=run_log]{run_log()}} \code{\link[=summary_log]{summary_log()}} \code{\link[=config_log]{config_log()}}
}
}
