% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bbi-path.R
\name{read_bbi_path}
\alias{read_bbi_path}
\title{Set the path to the bbi executable}
\usage{
read_bbi_path(
  path = Sys.getenv("BBI_EXE_PATH", NA),
  default = BBI_DEFAULT_PATH
)
}
\arguments{
\item{path}{path to the bbi executable}

\item{default}{a default path if \code{path} is \code{NA}}
}
\value{
\code{path} if not \code{NA}, else \code{default}
}
\description{
\code{\link[withr:with_options]{withr::with_options()}} is used throughout package tests to set the path to
the bbi executable. The test suite will fail if this path does not exist,
which limits flexibility across environments, e.g., in continuous
integration. Any solution that decouples the \code{with_options()} pattern from a
particular default path must work with \code{\link[devtools:check]{devtools::check()}}, which runs in a
separate process. We can inject environment variables into \code{check()}, so the
default behavior is to rely on such a variable to set the path, falling back
to a default path if unset.
}
\examples{
read_bbi_path()
}
\keyword{internal}
