% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{read_model}
\alias{read_model}
\title{Creates a model object from a YAML model file}
\usage{
read_model(.path)
}
\arguments{
\item{.path}{Path to the model to read, in the sense of absolute model path.}
}
\value{
S3 object of class \verb{bbi_\{.model_type\}_model}
}
\description{
Parses a model YAML file into a list object that contains correctly formatted
information from the YAML and is an S3 object of class
\verb{bbi_\{.model_type\}_model} that can be passed to \code{\link[=submit_model]{submit_model()}},
\code{\link[=model_summary]{model_summary()}}, etc.
}
\seealso{
\code{\link[=copy_model_from]{copy_model_from()}}, \code{\link[=new_model]{new_model()}}
}
