% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-log.R
\name{read_model_from_config}
\alias{read_model_from_config}
\title{Create a model object from a bbi_config.json path}
\usage{
read_model_from_config(.config_path)
}
\arguments{
\item{.config_path}{The absolute path a \code{bbi_config.json} file}
}
\value{
a \verb{bbi_\{.model_type\}_model} object
}
\description{
This is non-trivial because, while configs
always sit in the ouput directory, for NONMEM models
this is one dir deeper than the YAML and for Stan
models this is two dirs deeper than the YAML.
Hence, this is abstracted into an obnoxious private
helper function.
}
\keyword{internal}
