% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-model-results.R
\name{read_model_results}
\alias{read_model_results}
\title{read the model results}
\usage{
read_model_results(.x, ...)
}
\arguments{
\item{.x}{file_name}

\item{...}{params to pass to jsonlite::read_json}
}
\description{
read the model results
}
