% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{read_sim_data}
\alias{read_sim_data}
\title{Read in simulation data}
\usage{
read_sim_data(.mod, .suffix = "sim.tab")
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} or \code{bbi_nmsim_model} object.}
}
\description{
Read in simulation data
}
\keyword{internal}
