% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{read_translation_file}
\alias{read_translation_file}
\title{Read \code{model_NMrun_translation.txt} text file from \code{NM_run} directory}
\usage{
read_translation_file(.modelfit_dir)
}
\arguments{
\item{.modelfit_dir}{Directory containing where the model was run (i.e. the generated \code{modelfit_dir}).}
}
\description{
Read \code{model_NMrun_translation.txt} text file from \code{NM_run} directory
}
\keyword{internal}
