% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{reconcile_yaml}
\alias{reconcile_yaml}
\title{Reconcile model object with YAML file}
\usage{
reconcile_yaml(.mod)
}
\arguments{
\item{.mod}{`bbi_{.model_type}_model` object}
}
\description{
Extracts YAML path from model object and pulls in YAML file.
Any shared keys are overwritten with the values from the YAML and new keys in YAML are added to the model object.
}
