% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-register.R
\name{register_model_type}
\alias{register_model_type}
\title{Register a model type to class mapping}
\usage{
register_model_type(model_type, classes, class_type = c("model", "summary"))
}
\arguments{
\item{model_type}{Model type, as recorded in the model YAML file.}

\item{classes}{One or more classes to use for \code{model_type}.}

\item{class_type}{The kind of object, "model" or "summary".}
}
\description{
Define how to map a model type to model or summary object names when the
default mapping of \code{{type}} to \verb{bbi_\{type\}_model} or \verb{bbi_\{type\}_summary}
won't do. The primary use for this function is registering custom parent
types.
}
\examples{
\dontrun{
register_model_type("foo",
                    c("bbi_foo_model", "bbi_nonmem_model"),
                    "model")
register_model_type("foo",
                    c("bbi_foo_summary", "bbi_nonmem_summary"),
                    "summary")
}

}
\keyword{internal}
