% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{remove_star}
\alias{remove_star}
\title{remove_star}
\usage{
remove_star(.mod)
}
\arguments{
\item{.mod}{list object of model data}

\item{.star}{boolean variable to indicate significance}
}
\description{
remove_star
}
\section{Functions}{
\itemize{
\item \code{remove_star}: star on a model object and corresponding YAML.
}}

