% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_dup_cols}
\alias{remove_dup_cols}
\title{Removes duplicate col names}
\usage{
remove_dup_cols(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
Removes duplicate col names
}
\keyword{internal}
