% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{replace_bbi_args}
\alias{replace_bbi_args}
\title{Replaces babylon args to model}
\usage{
replace_bbi_args(.mod, .bbi_args)
}
\arguments{
\item{.mod}{A `bbi_{.model_type}_model` object}

\item{.bbi_args}{named list of arguments to add to the model}
}
\description{
Modifies model object and corresponding YAML by replacing .bbi_args with new .bbi_args
Use `print_nonmem_args()` to see a list list of valid babylon arguments.
}
