% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-threads.R
\name{replace_est_opt}
\alias{replace_est_opt}
\title{Helper function for replacing estimation options}
\usage{
replace_est_opt(.est_line, .match, .cap_iterations)
}
\arguments{
\item{.est_line}{the estimation line containing the \code{.match}}

\item{.match}{regex for the estimation option to be updated}

\item{.cap_iterations}{}
}
\description{
Helper function for replacing estimation options
}
\details{
This is specifically for updating the number of iterations for
parameters such as \code{MAXEVAL}, \code{NITER}, etc.
}
\keyword{internal}
