% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\docType{data}
\name{REQUIRED_TREE_PKGS}
\alias{REQUIRED_TREE_PKGS}
\title{Required packages for running \code{\link[=model_tree]{model_tree()}}}
\format{
An object of class \code{character} of length 7.
}
\usage{
REQUIRED_TREE_PKGS
}
\description{
Required packages for running \code{\link[=model_tree]{model_tree()}}
}
\details{
The packages below are required for both interactive and static viewing:
\itemize{
\item \code{collapsibleTree} for the core plot
\item \code{kableExtra} and \code{manipulateWidget} for creating and adding a table
respectively
}

The packages below are all used for rendering the model tree as a static image:
\itemize{
\item \code{htmlwidgets} is a dependency of \code{collapsibleTree}, so it's a 'free' import
\item \code{grid} is a built-in R package
\item \code{webshot} is needed for taking a screenshot
\item \code{png} is needed for reading in a PNG file in order to plot it in Rstudio
}
}
\keyword{internal}
