% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{resample_df}
\alias{resample_df}
\title{Resample data}
\usage{
resample_df(
  df,
  key_cols,
  strat_cols = NULL,
  n = NULL,
  key_col_name = "KEY",
  replace = TRUE
)
}
\arguments{
\item{df}{data frame}

\item{key_cols}{key columns to resample on}

\item{strat_cols}{columns to maintain proportion for stratification}

\item{n}{number of unique sampled keys, defaults to match dataset}

\item{key_col_name}{name of outputted key column. Default to "KEY"}

\item{replace}{whether to stratify with replacement}
}
\description{
Resample data
}
\keyword{internal}
