% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{safe_based_on}
\alias{safe_based_on}
\title{Checks for yaml files relative to a starting location}
\usage{
safe_based_on(.start, .based_on)
}
\arguments{
\item{.start}{The directory of the model (i.e. the YAML file) that the `based_on` will be added to.}

\item{.based_on}{Character vector or scaler of paths (with or without extension) to the models that will be added to `based_on`. Paths should be relative to `.start` argument.}
}
\description{
Checks for yaml files relative to a starting location
}
