% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-data.R
\name{safe_call_sourced}
\alias{safe_call_sourced}
\title{Private helper to safely call a function from a sourced file}
\usage{
safe_call_sourced(.func, .args, .file = NULL, .expected_class = NULL)
}
\arguments{
\item{.func}{The function to call}

\item{.args}{named list of args to call the function with (uses \code{do.call()})}

\item{.file}{(optional) file path to file that was sourced (for friendlier error message)}

\item{.expected_class}{(optional) name of class that the result of calling \code{.func} should return}
}
\description{
Private helper to safely call a function from a sourced file
}
\keyword{internal}
