% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-log.R
\name{safe_read_yaml}
\alias{safe_read_yaml}
\title{Read in YAML tags with error handling}
\usage{
safe_read_yaml(.path)
}
\arguments{
\item{.path}{path of yaml file}
}
\value{
dataframe with yaml path and tags
}
\description{
Read in YAML tags with error handling
}
\keyword{internal}
