% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-data.R
\name{safe_source}
\alias{safe_source}
\title{Private helper to safely source a file containing a single helper function}
\usage{
safe_source(.file, .env, .func_name = NULL)
}
\arguments{
\item{.file}{file path to file to source}

\item{.env}{environment into which to source the function. This is passed directly to
the \code{local} argument of \code{source()}.}

\item{.func_name}{(Optional) name of function you are sourcing (for friendlier error message)}
}
\description{
Private helper to safely source a file containing a single helper function
}
\keyword{internal}
