% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{sanitize_file_extension}
\alias{sanitize_file_extension}
\title{Private helper to remove file extensions to match expected input to new model.}
\usage{
sanitize_file_extension(.path)
}
\arguments{
\item{.path}{Path to save the new model. Will be the path to the model file
and YAML file (both without extension), and the path to the output
directory.}
}
\description{
Private helper to remove file extensions to match expected input to new model.
}
\keyword{internal}
