% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{save_mod_yaml}
\alias{save_mod_yaml}
\title{Saves a model spec object to a yaml file}
\usage{
save_mod_yaml(.spec, .out_path = NULL)
}
\arguments{
\item{.spec}{S3 object of class `bbi_{.model_type}_spec`}

\item{.out_path}{Character scaler with path to save out YAML file. By default, sets it to the model file name, with a yaml extension.}
}
\value{
Output list as specified above.
}
\description{
Saves a model spec object to a yaml file
}
