% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{save_model_yaml}
\alias{save_model_yaml}
\title{Saves a model object to a yaml file}
\usage{
save_model_yaml(.mod)
}
\arguments{
\item{.mod}{S3 object of class \verb{bbi_\{.model_type\}_model}}
}
\value{
The input \verb{bbi_\{.model_type\}_model} object, with its YAML md5 hash updated.
}
\description{
Saves the passed model object to its YAML file and updates
the md5 hash after saving.
}
\keyword{internal}
