% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{save_nsim_yaml}
\alias{save_nsim_yaml}
\title{Saves a model object to a yaml file. Intended for the creation of
\code{bbi_nmsim_model} objects.}
\usage{
save_nsim_yaml(.mod)
}
\arguments{
\item{.mod}{S3 object of class \verb{bbi_\{.model_type\}_model}}
}
\value{
The input \verb{bbi_\{.model_type\}_model} object, with its YAML md5 hash updated.
}
\description{
Saves a model object to a yaml file. Intended for the creation of
\code{bbi_nmsim_model} objects.
}
\details{
Pulls all \code{bbi_nmsim_model} specific changes into a separate \code{nmsim_args} list,
and then calls \code{save_model_yaml} normally.
}
\keyword{internal}
