% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-file-to-model-dir.R
\name{scaffold_missing_stan_files}
\alias{scaffold_missing_stan_files}
\title{Private helper to add scaffolds of any missing stan files}
\usage{
scaffold_missing_stan_files(.mod)
}
\arguments{
\item{.mod}{a \code{bbi_stan_model}}
}
\description{
This is used when setting up a new model.
}
\keyword{internal}
