% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_bbi_exe_path}
\alias{set_bbi_exe_path}
\alias{get_bbi_exe_path}
\title{Set global bbi_exe_path option}
\usage{
set_bbi_exe_path(.path)

get_bbi_exe_path()
}
\arguments{
\item{.path}{Path, either from working directory or absolute, that will be set as `options('rbabylon.bbi_exe_path')`}
}
\description{
Sets `options('rbabylon.bbi_exe_path')` to the absolute path passed to `.path`.
Note that the file must exist or this will error.
This is used whenever a function calls out to `bbi` to execute something, for example in `submit_model()` and `model_summary()`.

Gets the path set to `options('rbabylon.bbi_exe_path')` and checks that it is both absolute and exists.
This is used whenever a function calls out to `bbi` to execute something, for example in `submit_model()` and `model_summary()`.
}
