% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_bbi_null}
\alias{set_bbi_null}
\title{Replace BBI_NULL_NUM and BBI_NULL_STR with NA_real_ or NA_character_}
\usage{
set_bbi_null(.vals)
}
\arguments{
\item{.vals}{a numeric or character vector}
}
\description{
Replace BBI_NULL_NUM and BBI_NULL_STR with NA_real_ or NA_character_
}
\keyword{internal}
