% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-run-utils.R
\name{setup_analysis_run}
\alias{setup_analysis_run}
\title{Set up a bootstrap or SSE run}
\usage{
setup_analysis_run(
  .run,
  run_type = c("bootstrap", "sse"),
  n = 200,
  sample_size = NULL,
  strat_cols = NULL,
  replace = FALSE,
  seed = 1234,
  data = NULL,
  sim_col = NULL,
  .bbi_args = list(no_grd_file = TRUE, no_shk_file = TRUE),
  .overwrite = FALSE
)
}
\arguments{
\item{.run}{A \code{bbi_nmboot_model} or \code{bbi_nmsse_model} object.}

\item{n}{Number of datasets and model runs to generate.}

\item{sample_size}{Sample size for each new dataset. Defaults to \code{NULL}, which
matches the length of the original dataset.}

\item{strat_cols}{Columns to maintain proportion for stratification.}

\item{replace}{Logical (T/F). If \code{TRUE}, stratify with replacement.}

\item{seed}{A numeric seed to set prior to resampling the data; use \code{NULL} to
avoid setting a seed.}

\item{data}{Custom dataset to use for setup. Defaults to \code{NULL}.}

\item{sim_col}{Column denoting the simulation run number. This will be \code{"nn"},
if using \code{\link[=add_simulation]{add_simulation()}}, though \code{"IREP"} is also common (e.g., when
using \code{mrgsolve}).}

\item{.bbi_args}{Named list passed to \code{model_summary(orig_mod, .bbi_args)},
where \code{orig_mod} is the model \code{.boot_run} is based on. See
\code{\link[=print_bbi_args]{print_bbi_args()}} for valid options. Defaults to \code{list(no_grd_file = TRUE, no_shk_file = TRUE)} because \code{\link[=model_summary]{model_summary()}} is only called internally to
extract the number of records, so those files are irrelevant. Only used if
the based on model (the model the analysis is being performed on) has been
executed.}

\item{.overwrite}{Logical (T/F) indicating whether or not to overwrite
existing setup.}

\item{.run_type}{Either \code{"bootstrap"} or \code{"sse"}. Determines behavior for dataset handling.}
}
\description{
Set up a bootstrap or SSE run
}
\keyword{internal}
