% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{setup_bootrap_run}
\alias{setup_bootrap_run}
\title{Set up a bootstrap model run. Creates a new model object per run and generates
a new re-sampled dataset per run.}
\usage{
setup_bootrap_run(
  .boot_run,
  n = 10,
  strat_cols = NULL,
  replace = TRUE,
  seed = 1234,
  .overwrite = FALSE
)
}
\arguments{
\item{.boot_run}{a \code{bbi_nmboot_model} object.}

\item{n}{number of unique sampled keys, defaults to match dataset}

\item{strat_cols}{columns to maintain proportion for stratification}

\item{replace}{whether to stratify with replacement}

\item{seed}{a seed for sampling the data}

\item{.overwrite}{logical (T/F) indicating whether or not to overwrite existing
setup for a bootstrap run.}
}
\description{
Set up a bootstrap model run. Creates a new model object per run and generates
a new re-sampled dataset per run.
}
\details{
This setup is automatically done internally as part of submitting a bootstrap
run. However, users can call this function explicitly before submitting a
bootstrap run in order save some initial overhead. This can be useful for
large datasets or a large number of samples \code{n}.
}
