% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{setup_bootstrap_run}
\alias{setup_bootstrap_run}
\title{Set up a bootstrap model run.}
\usage{
setup_bootstrap_run(
  .boot_run,
  n = 100,
  strat_cols = NULL,
  replace = TRUE,
  seed = 1234,
  .overwrite = FALSE
)
}
\arguments{
\item{.boot_run}{a \code{bbi_nmboot_model} object.}

\item{n}{number of unique sampled keys, defaults to match dataset}

\item{strat_cols}{columns to maintain proportion for stratification}

\item{replace}{whether to stratify with replacement}

\item{seed}{a seed for sampling the data}

\item{.overwrite}{logical (T/F) indicating whether or not to overwrite existing
setup for a bootstrap run.}
}
\description{
Creates a new model object and re-sampled dataset per model run.
}
\examples{
\dontrun{

# Setup
.boot_run <- new_bootstrap_run(.mod)
.boot_run <- setup_bootstrap_run(.boot_run, n = 1000, seed = 1234)

# Submit
submit_model(.boot_run)

# Check status of runs during submission
get_model_status(.boot_run)
}
}
\seealso{
new_bootstrap_run
}
