% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{setup_sim_run}
\alias{setup_sim_run}
\title{Set up VPC simulation}
\usage{
setup_sim_run(.sim_run, n = 100, seed = 1234)
}
\arguments{
\item{.sim_run}{a \code{bbi_nmsim_model} object}

\item{n}{number of replicates.}

\item{seed}{a seed for sampling the data.}
}
\description{
Creates new \verb{$SIMULATION} and \verb{$MSFI} records, and modifies the \verb{$PRED} or
\verb{$ERROR} record (whichever exists) to have the additional line: \code{REPI=IREP}.
}
\keyword{internal}
