% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{sig}
\alias{sig}
\title{Format digits}
\usage{
sig(.x, .digits)
}
\arguments{
\item{x}{numeric, value to manipulate}

\item{digits}{numeric, number of significant digits}
}
\value{
character vector of formatted values
}
\description{
Simplified version of \code{pmtables::sig()} for formatting numbers
to a specific number of significant digits.
}
\keyword{internal}
