% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-for-tests.R
\name{skip_if_no_stan}
\alias{skip_if_no_stan}
\title{Skip test if not running Stan tests}
\usage{
skip_if_no_stan(.test_name)
}
\arguments{
\item{.test_name}{Character scalar to identify the test being potentially skipped.
This is printed in the skip message}
}
\description{
First checks if \code{cmdstanr} is installed, then if the \code{SKIP_STAN_TESTS}
environment variable is set.
}
\keyword{internal}
