% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-for-tests.R
\name{skip_if_old_bbi}
\alias{skip_if_old_bbi}
\title{Skip test if bbi version is as old or older than the given version.}
\usage{
skip_if_old_bbi(v)
}
\arguments{
\item{v}{a package version or a string that can be converted to one}
}
\description{
If the version can't be parsed by \code{\link[=package_version]{package_version()}}, the test is never
skipped.
}
\keyword{internal}
