% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-for-tests.R
\name{skip_if_old_nmrec}
\alias{skip_if_old_nmrec}
\title{Skip test if current nmrec version is below specified one.}
\usage{
skip_if_old_nmrec(v)
}
\arguments{
\item{v}{a package version (or a string that can be converted to one)}
}
\description{
Skip test if current nmrec version is below specified one.
}
\keyword{internal}
