% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-for-tests.R
\name{skip_if_over_rate_limit}
\alias{skip_if_over_rate_limit}
\title{Skip test if over Github API rate limit}
\usage{
skip_if_over_rate_limit(by = 5)
}
\arguments{
\item{by}{if less than this number of requests are left before hitting the rate limit, skip the test}
}
\description{
This is based on a function of the same name from the \code{remotes} package.
Note, this uses no token, so it is checking the public rate limit.
This is appropriate because this package has no interface for passing in a token, so it will always be hitting the public API.
}
\keyword{internal}
