% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{skip_if_tree_missing_deps}
\alias{skip_if_tree_missing_deps}
\title{Skip tests if missing \code{\link[=model_tree]{model_tree()}} dependencies}
\usage{
skip_if_tree_missing_deps()
}
\description{
Skip tests if missing \code{\link[=model_tree]{model_tree()}} dependencies
}
\keyword{internal}
