% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-for-tests.R
\name{skip_long_tests}
\alias{skip_long_tests}
\title{Skip long-running tests}
\usage{
skip_long_tests(.explanation = "Skipping long running tests")
}
\arguments{
\item{.explanation}{Reason for skipping tests, or description of tests being skipped}
}
\description{
For example, tests that actual submit models to be run.
}
\keyword{internal}
