% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{stan_files_from_brms}
\alias{stan_files_from_brms}
\title{Write necessary stan files to disk from brms args}
\usage{
stan_files_from_brms(.path, args)
}
\arguments{
\item{.path}{absolute model path (files will be created in this dir)}

\item{args}{named list of arguments to pass to \link[brms:make_stancode]{brms::make_stancode} and \link[brms:make_standata]{brms::make_standata}}
}
\description{
Write necessary stan files to disk from brms args
}
\keyword{internal}
