% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-stan.R
\docType{data}
\name{STANDATA_SCAFFOLD_STRING}
\alias{STANDATA_SCAFFOLD_STRING}
\title{Create Stan data}
\format{
An object of class \code{character} of length 1.
}
\usage{
STANDATA_SCAFFOLD_STRING
}
\arguments{
\item{.dir}{The absolute path to the directory containing this file.
This is useful for building file paths to the input files you will load.
Note: you \emph{don't} need to pass anything to this argument, you only use it
within the function. \code{bbr} will pass in the correct path when it calls
\code{make_standata()} under the hood.}
}
\value{
Must return the list that will be passed to \code{data} argument of
\code{cmdstanr::sample()}
}
\description{
Create Stan data
}
\keyword{datasets}
