% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan.R
\name{standata_to_json}
\alias{standata_to_json}
\title{Write Stan input data to json}
\usage{
standata_to_json(.mod, .out_path = NULL)
}
\arguments{
\item{.mod}{a \code{bbi_stan_model} object}

\item{.out_path}{The path to write the json to on disk. If \code{NULL}, the
default, it will be written to \code{build_path_from_model(.mod, STANDATA_JSON_SUFFIX)}. This argument is primarily used to write to a temp
path when testing whether the json in an output folder is the same as what
is produced by calling this.}
}
\value{
Invisibly returns the list object returned
from make_standata()
}
\description{
Runs the \code{-standata.R} file associated with a model
and writes the returned data list to json with
\link[cmdstanr:write_stan_json]{cmdstanr::write_stan_json}.
}
\keyword{internal}
