% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-stan.R
\docType{data}
\name{STANINIT_SCAFFOLD_STRING}
\alias{STANINIT_SCAFFOLD_STRING}
\title{Create Stan initial values}
\format{
An object of class \code{character} of length 1.
}
\usage{
STANINIT_SCAFFOLD_STRING
}
\arguments{
\item{.data}{The data list returned from \code{make_standata()} for this model.
This is provided in case any of your initial values are dependant on some
aspect of the data (e.g. the number of rows). Note: you \emph{don't} need to pass
anything to this argument, you only use it within the function. \code{bbr} will
pass in the correct data when it calls \code{make_init()} under the hood.}
}
\value{
Must return something that can be passed to the \code{init} argument of
\code{cmdstanr::sample()}. There are several options; see \code{?cmdstanr::sample} for
details.
}
\description{
Create Stan initial values
}
\keyword{datasets}
