% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_get_fail_msg}
\alias{stop_get_fail_msg}
\title{Build error message and throw error for generic failure fo get_... functions}
\usage{
stop_get_fail_msg(.bbi_object, .key, .msg = "")
}
\arguments{
\item{.bbi_object}{The object that something is attempting to be extracted from}

\item{.key}{The name of the field that is attempting to be extracted}

\item{.msg}{Character scaler or vector of more specific error messages to include at the end}
}
\description{
Build error message and throw error for generic failure fo get_... functions
}
