% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_get_scaler_msg}
\alias{stop_get_scaler_msg}
\title{Build error message and throw error for passing character vector to get_... functions}
\usage{
stop_get_scaler_msg(.len)
}
\arguments{
\item{.len}{The length of the vector that was passed.}
}
\description{
Build error message and throw error for passing character vector to get_... functions
}
