% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{stop_if_tree_missing_deps}
\alias{stop_if_tree_missing_deps}
\title{Error if missing \code{\link[=model_tree]{model_tree()}} dependencies}
\usage{
stop_if_tree_missing_deps()
}
\description{
Error if missing \code{\link[=model_tree]{model_tree()}} dependencies
}
\keyword{internal}
