% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strict_mode_error}
\alias{strict_mode_error}
\title{Raise error in strict mode}
\usage{
strict_mode_error(err_msg)
}
\description{
Raises an error if \code{options("rbabylon.strict" = TRUE)} (recommended). Otherwise raises a warning.
These errors are used for things like type-checking which guarantees safe, predictable behavior of functions,
but theoretically advanced users or developers could want to turn them off.
}
\keyword{internal}
