% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{style_html}
\alias{style_html}
\title{Helper for coloring text and applying other styles}
\usage{
style_html(txt, color = "black", ..., br_before = FALSE, br_after = FALSE)
}
\arguments{
\item{txt}{text to format}

\item{color}{color added to the text}

\item{...}{Other \code{HTML} styling to add to the \code{span} call}

\item{br_before}{Logical (\code{T}/\code{F}). If \code{TRUE}, add a new line before the text}

\item{br_after}{Logical (\code{T}/\code{F}). If \code{TRUE}, add a new line after the text}
}
\description{
Helper for coloring text and applying other styles
}
\keyword{internal}
