% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{submit_nonmem_analysis}
\alias{submit_nonmem_analysis}
\title{Submit a NONMEM analysis}
\usage{
submit_nonmem_analysis(
  .mod,
  .bbi_args = NULL,
  .mode = "sge",
  ...,
  .overwrite = FALSE,
  .config_path = NULL,
  .wait = FALSE,
  .dry_run = FALSE,
  .batch_size = 100
)
}
\arguments{
\item{.mod}{The model object to submit.}

\item{.bbi_args}{A named list specifying arguments to pass to bbi
formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}. Run \code{\link[=print_bbi_args]{print_bbi_args()}} to see valid arguments. Note that bbr does
not support changing the output directory (including through the model or
global YAML files).}

\item{.mode}{Mode for model submission: "local", "sge", or "slurm". If
unspecified, the value is set to the value of the \code{bbr.bbi_exe_mode}
option. This option defaults to "sge" on Linux and "local" otherwise.}

\item{...}{args passed through to \code{bbi_exec()}}

\item{.overwrite}{Logical to specify whether or not to overwrite existing
model output from a previous run. If \code{NULL}, the default, will defer to
setting in \code{.bbi_args} or \code{bbi.yaml}. If \emph{not} \code{NULL} will override any
settings in \code{.bbi_args} or \code{bbi.yaml}. \strong{The exception to this are analysis
runs (\code{bbi_nmboot_model} and \code{bbi_nmsse_model} objects).} For analysis
runs (bootstrap and SSE), this defaults to \code{FALSE} and does \emph{not} respect
any setting passed via \code{.bbi_args} or a \code{bbi.yaml} config file. To
overwrite an existing analysis output, a user must pass \code{TRUE} through this
argument.}

\item{.config_path}{Path to a bbi configuration file. If \code{NULL}, the
default, will attempt to use a \code{bbi.yaml} in the same directory as the
model.}

\item{.wait}{If \code{TRUE}, the default, wait for the bbi process to return
before this function call returns. If \code{FALSE} function will return while
bbi process runs in the background.}

\item{.dry_run}{Returns an object detailing the command that would be run,
insted of running it. This is primarily for testing but also a debugging
tool.}

\item{.batch_size}{Number of models to submit to run concurrently as a
"batch." Passing \code{NULL} (or a number larger than the number of submitted
models) will bypass this and submit all models concurrently. This will
launch a background process to manage the batch submission. Details from
this process are logged in the \code{OUTPUT} file in top-level bootstrap model
directory.}
}
\description{
Private implementation function for submitting bootstrap and SSE runs in
batches.
}
\keyword{internal}
