% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-nonmem-model.R
\name{submit_nonmem_model}
\alias{submit_nonmem_model}
\title{Submit a NONMEM model via babylon}
\usage{
submit_nonmem_model(
  .path,
  .type = c("sge", "local"),
  .args = NULL,
  ...,
  .config_path = NULL,
  .dry_run = FALSE
)
}
\arguments{
\item{.path}{Full path to a model file(s) that should be run. Path MUST be either an absolute path or relative to the R working directory.}

\item{.type}{Either "local" for local execution or "sge" to submit model(s) to the grid}

\item{.args}{A named list specifying arguments to pass to babylon formatted like `list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)` All available arguments are in `NONMEM_ARGS`.}

\item{...}{args passed through to `bbi_exec()`}

\item{.config_path}{Optionally specify a path to a babylon.yml config. If not specified, the config in the model directory will be used by default. Path MUST be either an absolute path or relative to the model directory.}

\item{.dry_run}{Returns character string of command that would be run, insted of running it. This is primarily for testing but also a debugging tool.}
}
\value{
output from the model run (?)
}
\description{
Submit a NONMEM model via babylon
}
