% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{submit_nonmem_models}
\alias{submit_nonmem_models}
\title{Submit multiple NONMEM models in batch via babylon}
\usage{
submit_nonmem_models(
  .mods,
  .bbi_args = NULL,
  .mode = c("sge", "local"),
  ...,
  .config_path = file.path(get_model_directory() \%||\% ".", "babylon.yaml"),
  .wait = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{.mods}{A list of S3 objects of class `bbi_nonmem_model`}
}
\value{
A list of S3 objects of class `babylon_process`
}
\description{
Private implementation function called by `submit_models()` dispatches.
}
\keyword{internal}
