% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{submit_stan_model_cmdstanr}
\alias{submit_stan_model_cmdstanr}
\title{Submit a Stan model via cmdstanr}
\usage{
submit_stan_model_cmdstanr(.mod, .mode = c("local"), ..., .overwrite = NULL)
}
\arguments{
\item{.mod}{An S3 object of class \code{bbi_stan_model}}
}
\value{
The object returned from \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}
}
\description{
Private implementation function called by \code{submit_model()} dispatches.
}
\keyword{internal}
