% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-run-utils.R
\name{summarize_analysis_run}
\alias{summarize_analysis_run}
\title{Summarize an analysis (bootstrap or SSE) run}
\usage{
summarize_analysis_run(.run)
}
\arguments{
\item{.run}{A \code{bbi_nmboot_model} or \code{bbi_nmsse_model} object.}
}
\description{
Summarize an analysis (bootstrap or SSE) run
}
\keyword{internal}
