% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{summarize_bootstrap}
\alias{summarize_bootstrap}
\alias{summarize_bootstrap_run}
\alias{bootstrap_estimates}
\alias{get_boot_models}
\title{Summarize a bootstrap run}
\usage{
summarize_bootstrap_run(.boot_run, force_resummarize = FALSE)

bootstrap_estimates(.boot_run, format_long = FALSE, force_resummarize = FALSE)

get_boot_models(.boot_run)
}
\arguments{
\item{.boot_run}{a \code{bbi_nmboot_model} object.}

\item{force_resummarize}{logical (T/F). If \code{TRUE}, force re-summarization.
Will \emph{only} update the saved out \code{RDS} file when specified via
\code{summarize_bootstrap_run()}. See details for more information.}

\item{format_long}{logical (T/F). If \code{TRUE}, format data as a long table,
making the data more portable for plotting.}
}
\description{
Summarize the parameter estimates, run details, and any heuristics of a
bootstrap run, saving the results to a \code{boot_summary.RDS} data file within the
bootstrap run directory.
}
\details{
\itemize{
\item \code{bootstrap_estimates()} \emph{quickly} extracts and formats the parameter estimates
from each model run. If the data was previously summarized, the data will be
read in instead of re-executing (this can be overridden via
\code{force_resummarize = TRUE}).
\item \code{summarize_bootstrap_run()} does the following things:
\itemize{
\item Tabulates run details and heuristics.
\item Calls \code{summary_log()} and binds the results to the parameter estimates.
\itemize{
\item \code{bootstrap_estimates()} will include this appended model summary
information if a \code{boot_summary.RDS} data file exists.
}
\item Either saves this data out to \code{boot_summary.RDS}, or reads it in if it
already exists (see section below).
\item Formats the returned object as a \code{bbi_nmboot_summary} S3 object, and
displays key summary information when printed to the console.
}
}
\subsection{Saved out data file:}{

The first time \code{summarize_bootstrap_run()} is called (or if
\code{force_resummarize = TRUE}), it will save the results to a \code{boot_summary.RDS}
data file within the bootstrap run directory. If one already exists, that data
set will be read in by default instead of being re-summarized.
\itemize{
\item The purpose of this is functionality two fold. For one, it helps avoid the
need of re-executing \code{model_summary()} calls for a large number of runs. It
also helps to reduce the number of files you need to commit via version
control (see \code{cleanup_bootstrap_run()}).
}
}
}
\section{Functions}{
\itemize{
\item \code{summarize_bootstrap_run()}: Summarize a bootstrap run and store results

\item \code{bootstrap_estimates()}: Tabulate parameter estimates for each model
submission in a bootstrap run

\item \code{get_boot_models()}: Read in all bootstrap run model objects

}}
\examples{
\dontrun{

.boot_run <- read_model(file.path(MODEL_DIR, "1-boot"))
boot_sum <- summarize_bootstrap_run(.boot_run)

# Optionally compare to original estimates
param_estimates_compare(boot_sum)


# Long format is helpful for plotting estimates:
bootstrap_estimates(.boot_run, format_long = TRUE) \%>\%
dplyr::filter(grepl("THETA", parameter_names)) \%>\%
  ggplot(aes(x = estimate)) +
  facet_wrap(~parameter_names, scales = "free") +
  geom_histogram(color = "white", alpha = 0.7) +
  theme_bw()

}

}
\seealso{
\code{\link[=param_estimates_compare]{param_estimates_compare()}} \code{\link[=cleanup_bootstrap_run]{cleanup_bootstrap_run()}}
}
