% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{summarize_bootstrap_run}
\alias{summarize_bootstrap_run}
\title{Summarize a bootstrap run}
\usage{
summarize_bootstrap_run(
  .boot_run,
  estimates_only = FALSE,
  include_based_on = FALSE
)
}
\arguments{
\item{.boot_run}{a \code{bbi_nmboot_model} object.}

\item{estimates_only}{logical (T/F). Set to \code{TRUE} to only include parameter
estimates (much faster for a large number of runs). If \code{FALSE}, appends
additional model summary information.}

\item{include_based_on}{logical (T/F). If \code{TRUE}, include the model the
bootstrap run was based on. \strong{Only use this} if that model hasn't changed
since the latest bootstrap run.}
}
\description{
Summarize a bootstrap run
}
