% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse-model.R
\name{summarize_sse}
\alias{summarize_sse}
\alias{summarize_sse_run}
\alias{sse_estimates}
\alias{get_sse_models}
\title{Summarize an SSE run}
\usage{
summarize_sse_run(.sse_run, force_resummarize = FALSE)

sse_estimates(.sse_run, force_resummarize = FALSE)

get_sse_models(.sse_run)
}
\arguments{
\item{.sse_run}{A \code{bbi_nmsse_model} object.}

\item{force_resummarize}{Logical (T/F). If \code{TRUE}, force re-summarization.
Will \emph{only} update the saved out \code{RDS} file when specified via
\code{summarize_sse_run()}. See details for more information.}
}
\description{
Summarize the parameter estimates, run details, and any heuristics of a
SSE run, saving the results to a \code{sse_summary.RDS} data file within the
SSE run directory.
}
\details{
\itemize{
\item \code{summarize_sse_run()} does the following things:
\itemize{
\item Tabulates run details and heuristics.
\item Calls \code{summary_log()} and binds the results to the parameter estimates.
information if a \code{sse_summary.RDS} data file exists.
\item Either saves this data out to \code{sse_summary.RDS}, or reads it in if it
already exists (see section below).
\item Formats the returned object as a \code{bbi_nmsse_summary} S3 object, and
displays key summary information when printed to the console.
}
}
\subsection{Saved out data file:}{

The first time \code{summarize_sse_run()} is called (or if
\code{force_resummarize = TRUE}), it will save the results to a \code{sse_summary.RDS}
data file within the SSE run directory. If one already exists, that data
set will be read in by default instead of being re-summarized.
\itemize{
\item The purpose of this is functionality two fold. For one, it helps avoid the
need of re-executing \code{model_summary()} calls for a large number of runs. It
also helps to reduce the number of files you need to commit via version
control (see \code{cleanup_sse_run()}).
}
}
}
\section{Functions}{
\itemize{
\item \code{summarize_sse_run()}: Summarize an SSE run and store results

\item \code{sse_estimates()}: Tabulate parameter estimates for each model
submission in an SSE run

\item \code{get_sse_models()}: Read in all SSE run model objects

}}
\examples{
\dontrun{

.sse_run <- read_model(file.path(MODEL_DIR, "1-sse"))
boot_sum <- summarize_sse_run(.sse_run)


}

}
\seealso{
\code{\link[=cleanup_sse_run]{cleanup_sse_run()}} \code{\link[=new_sse_run]{new_sse_run()}} \code{\link[=setup_sse_run]{setup_sse_run()}}
}
