% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{suppressSpecificWarning}
\alias{suppressSpecificWarning}
\title{Suppress a warning that matches `.regexpr`}
\usage{
suppressSpecificWarning(.expr, .regexpr)
}
\arguments{
\item{.expr}{Expression to run}

\item{.regexpr}{Regex to match against any generated warning. Warning will be suppressed if this matches the warning message.}
}
\description{
Suppress a warning that matches `.regexpr`
}
