% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-join.R
\name{tabfiles}
\alias{tabfiles}
\title{Return a vector of absolute paths to table files to look for}
\usage{
tabfiles(.mod)
}
\arguments{
\item{.mod}{the \code{bbi_nonmem_model} object}
}
\description{
Return a vector of absolute paths to table files to look for
}
\details{
Called by redataset
}
\keyword{internal}
