% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags-diff.R
\name{tags_diff}
\alias{tags_diff}
\alias{add_tags_diff}
\title{Compare tags between models}
\usage{
tags_diff(.bbi_object, .mod2 = NULL, .print = TRUE, ...)

add_tags_diff(.log_df)
}
\arguments{
\item{.bbi_object}{The object to compare. Could be a
\verb{bbi_\{.model_type\}_model} object or a
tibble of class \code{bbi_run_log_df}.}

\item{.mod2}{If a \code{bbi_model} object is passed, compare \code{.bbi_object} to
\code{.mod2}. If \code{.mod2 = NULL}, the default, compare the \code{.bbi_object} model to
any models in its \code{based_on} field. \strong{Only valid for \code{bbi_model}
method.}}

\item{.print}{If \code{TRUE}, the default, will print a nicely formatted
version of the returned list to the console. \strong{Only valid for \code{bbi_model}
method.}}

\item{...}{arguments passed through to methods. (Currently none.)}

\item{.log_df}{a \code{bbi_run_log_df} tibble (the output of \code{\link[=run_log]{run_log()}})}
}
\value{
\strong{\code{tags_diff.bbi_model()}} invisibly returns a list with two
elements, the first named \code{tags_added} and the second named \code{tags_removed}.
The list is returned invisibly because, by default, a nicely formatted
version of the same information is printed to the console. User can pass
\code{.print = FALSE} to turn this off.

\strong{\code{tags_diff.bbi_run_log_df()}} returns a named list of lists, with one
element for each row in the input tibble, with the name corresponding to the
value in the \code{run} column for that row. Each element of the list will contain
the two-element list returned from \code{tags_diff.bbi_model()} (described above)
for the relevant model.

\strong{\code{add_tags_diff()}} returns the same tibble that was passed to
it, but with two additional columns named \code{tags_added} and \code{tags_removed}.

\strong{In all cases:}
\itemize{
\item \code{tags_added} contains any tags that are on the relevant model, but \emph{not} on
any of the models it is based on.
\item \code{tags_removed} contains any tags that are on at least one of the models it
is based on, but \emph{not} on the relevant model.
}
}
\description{
By default, this compares a model's tags to the tags on any of its "parent"
models. A parent model is any model referenced in the relevant model's
\code{based_on} field.  The \code{bbi_model} dispatch can optionally take a second
model, in which case it will compare the first model to the second model,
ignoring the \code{based_on} field entirely.
}
