% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbr.R
\name{test_nmrec_version}
\alias{test_nmrec_version}
\title{Check that nmrec satisfies a version constraint}
\usage{
test_nmrec_version(.min_version)
}
\arguments{
\item{.min_version}{The minimum allowed version.}
}
\value{
returns \code{TRUE} if the installed nmrec version satisfies the
constraint. Otherwise error out.
}
\description{
Check that nmrec satisfies a version constraint
}
\keyword{internal}
