% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toString.R
\name{toString.bbi_log_df}
\alias{toString.bbi_log_df}
\title{Collapse \code{bbi_log_df} list columns to character}
\usage{
\method{toString}{bbi_log_df}(x, ..., .sep = ", ")
}
\arguments{
\item{x}{Input tibble to modify}

\item{...}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> One or more unquoted
expressions separated by commas. Variable names can be used as if they
were positions in the data frame, so expressions like \code{x:y} can
be used to select a range of variables.}

\item{.sep}{Character scalar to use a separator when collapsing vectors. Defaults to \code{", "}.}
}
\value{
Returns the same tibble as \code{x}, but any list columns named in \code{...} will be collapsed to a character column, with one scalar value (or \code{NA}) for each row.
See "Details" section for caveats.
}
\description{
Collapses a list column in a \code{bbi_log_df} tibble into a column of character scalars (aka strings).
}
\details{
Any non-list columns passed to \code{...} will be ignored and will trigger a warning notifying the user that only list columns can be collapsed.

Any list columns passed to \code{...} which do \emph{not} contain either character, numeric, or logical vectors (i.e. lists of lists) will be silently ignored.
}
\examples{
df <- tibble::tibble(
  absolute_model_path   = c(1, 2, 3),
  char_list = list(c("foo", "bar"), "baz", c("naw", "dawg")),
  num_list  = list(c(23, 34, 45), c(-1, -2, -3), NULL)
)
class(df) <- c("bbi_log_df", class(df))

toString(df, char_list, num_list)

}
