% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak-initial-estimates.R
\name{tweak_matrix}
\alias{tweak_matrix}
\title{Tweak an \code{OMEGA} or \code{SIGMA} record}
\usage{
tweak_matrix(init_mat, .p, digits)
}
\arguments{
\item{init_mat}{matrix of initial \code{OMEGA} or \code{SIGMA} estimates. Should include
several attributes, including \code{nmrec_flags} and \code{nmrec_record_size}.}
}
\description{
Tweak an \code{OMEGA} or \code{SIGMA} record
}
\keyword{internal}
