% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak-initial-estimates.R
\name{tweak_thetas}
\alias{tweak_thetas}
\title{Tweak theta estimates while respecting bounds}
\usage{
tweak_thetas(init_thetas, .p, digits)
}
\arguments{
\item{init_thetas}{table of theta estimates and bounds}

\item{.p}{Percent to tweak the initial parameter estimates by. Represented as
a decimal.}

\item{digits}{Number of significant digits to round estimates to.}
}
\description{
Tweak theta estimates while respecting bounds
}
\keyword{internal}
