% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak-initial-estimates.R
\name{tweak_values}
\alias{tweak_values}
\title{Tweak values}
\usage{
tweak_values(values, .p)
}
\arguments{
\item{values}{vector of values to tweak}

\item{.p}{Percent to tweak the initial parameter estimates by. Represented as
a decimal.}
}
\description{
Tweak values
}
