% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-tables.R
\name{unquote_filename}
\alias{unquote_filename}
\title{Unquote file paths for inclusion within a NONMEM control file}
\usage{
unquote_filename(.files)
}
\arguments{
\item{.files}{vector of file paths}
}
\description{
Unquote file paths for inclusion within a NONMEM control file
}
\keyword{internal}
