% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{use_bbi}
\alias{use_bbi}
\title{Installs most current release of bbi}
\usage{
use_bbi(.path = NULL, .version = "latest", .force = FALSE, .quiet = NULL)
}
\arguments{
\item{.path}{absolute path to install bbi to. See Details section for
defaults, if nothing is passed. Note that this should be the path where you
would like the bbi executable to be installed, \emph{not} the path to the
directory in which you want to install it. For example, you should pass
\code{"/some/dir/bbi"} and \emph{not} \code{"/some/dir"}.}

\item{.version}{version of bbi to install. Must pass a character scalar
corresponding to a tag found in
\verb{https://github.com/metrumresearchgroup/bbi/releases}}

\item{.force}{If \code{FALSE}, the default, skips installation if requested
version and local version are the same. If \code{TRUE} forces installation if it
will be the same version.}

\item{.quiet}{\strong{Deprecated.} Use \code{options("bbr.verbose")} instead to control
printing. Defaults to \code{NULL}, which reads \code{!getOption("bbr.verbose")}. If
\code{TRUE}, suppresses output printed to the console.}
}
\value{
character
}
\description{
Identifies system running and pulls the relevant tarball for the
current release of bbi from GitHub, and then installs it at \code{.path} (see
Details section for defaults). If used in an interactive session, will open
an installation menu confirming the installed version. This function will
print information about the installed version. This \strong{printing can be
suppressed} by setting \code{options(bbr.verbose = FALSE)}.
}
\details{
If nothing is passed to the \code{.path} argument, \code{use_bbi()} will
attempt to find a valid path for installation. The following decision
waterfall is used:
\itemize{
\item First, check \code{getOption("bbr.bbi_exe_path")}. If this is anything \emph{other
than} \code{"bbi"} (the default value) then attempt to install to that path.
\item Second, check \code{Sys.which("bbi")} which will look for a \code{bbi} installation
in the user's \verb{$PATH}. If one is found, ask the user if they wish to
overwrite it and, if they confirm, install to that path.
\item Third, attempt to install to a "default location" and add this location to
the user's \verb{$PATH}. If \code{Sys.getenv("XDG_DATA_HOME")} is found, install to
\code{{Sys.getenv("XDG_DATA_HOME")}/bbi/bbi} (per \href{https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html}{XGD specification}).
Otherwise, install to OS dependent defaults:
\itemize{
\item Linux: \code{{Sys.getenv("HOME")}/.local/share/bbi/bbi}
\item Mac: \verb{/usr/local/bin/bbi}
\item Windows: \verb{\{Sys.getenv("APPDATA")\}\\bbi\\bbi}
}
}

If none of these are successful, the user will be prompted to set
\code{options("bbr.bbi_exe_path")} and try again.
}
