% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-estimates.R
\name{using_old_priors}
\alias{using_old_priors}
\title{Check if \emph{non-specific} priors are being used.}
\usage{
using_old_priors(ctl)
}
\arguments{
\item{ctl}{An \code{nmrec_ctl_records} object.}
}
\description{
The \emph{non-specific} method of priors means using \verb{$THETA} instead of \verb{$THETAPV}
to specify \code{prior} records.
}
\keyword{internal}
