% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-handling.R
\name{validate_matrix_pd}
\alias{validate_matrix_pd}
\title{Check and modify a full matrix for positive-definiteness}
\usage{
validate_matrix_pd(full_mat, digits)
}
\arguments{
\item{full_mat}{initial estimates object as returned by \code{nmrec::extract_omega}
or \code{nmrec::extract_sigma}. Should be a matrix with \code{nmrec_record_size} attributes.}

\item{digits}{Number of significant digits to round estimates to.}
}
\description{
Splits a full matrix into sub-matrices per record, checks for positive-definiteness,
and attempts to find a near positive-definite matrix if not. If one cant be found,
the initial estimates shouldnt be tweaked in an automated fasion.
}
\keyword{internal}
