% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{validate_parent_mod}
\alias{validate_parent_mod}
\title{Confirm parent model path is valid, and error informatively if not}
\usage{
validate_parent_mod(.parent_mod)
}
\arguments{
\item{.parent_mod}{Path to a parent model to inherit properties from.}
}
\description{
Confirm parent model path is valid, and error informatively if not
}
\keyword{internal}
