% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verbose_msg}
\alias{verbose_msg}
\title{Prints message(s) if \code{isTRUE(getOption("bbr.verbose"))}}
\usage{
verbose_msg(.msg)
}
\description{
Prints message(s) if \code{isTRUE(getOption("bbr.verbose"))}
}
\keyword{internal}
