% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{verify_model_yaml_integrity}
\alias{verify_model_yaml_integrity}
\alias{reconcile_yaml}
\alias{check_yaml_in_sync}
\title{Verify YAML file integrity}
\usage{
reconcile_yaml(.mod)

check_yaml_in_sync(.mod)
}
\arguments{
\item{.mod}{\verb{bbi_\{.model_type\}_model} object}
}
\description{
These functions use an md5 digest of the model YAML file, stored each time a model is loaded into memory,
to keep track of any changes made to the YAML that are \emph{not} reflected in the object held in memory.
}
\section{Functions}{
\itemize{
\item \code{reconcile_yaml()}: Use to manually reconcile model object in memory with its YAML file.
Extracts YAML path from model object and pulls in YAML file.
Any shared keys are overwritten with the values from the YAML and new keys in YAML are added to the model object.
The md5 digest is then updated to reflect the new state.

\item \code{check_yaml_in_sync()}: Checks that model YAML file is the same as when it was last read into the model object.
Errors if the md5 digests are not identical. This is called internally in most functions that interact with a model object.

}}
