% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{version_message}
\alias{version_message}
\title{Private helper to construct version comparison message}
\usage{
version_message(local_v = bbi_version(), release_v = bbi_current_release())
}
\description{
Private helper to construct version comparison message
}
\keyword{internal}
