% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wait_for_nonmem}
\alias{wait_for_nonmem}
\title{Wait for NONMEM models to finish}
\usage{
wait_for_nonmem(model_list, time_limit, interval = 5)
}
\arguments{
\item{model_list}{a list of \code{bbi_nonmem_model} objects}

\item{time_limit}{integer for maximum number of seconds in total to wait before continuing (exiting function)}

\item{interval}{integer for number of seconds to wait between each check}
}
\description{
Calling \code{wait_for_nonmem()} will freeze the user's console until the model(s) have finished running.
}
