% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{windows_install_commands}
\alias{windows_install_commands}
\title{Private helper function for building commands to install bbi on Windows}
\usage{
windows_install_commands(.path, .bbi_url)
}
\arguments{
\item{.path}{Absolute path to install bbi to}

\item{.bbi_url}{Full url to download tarball from}
}
\description{
THIS HAS NOT BEEN TESTED!!! Honestly, I have not idea if the file.path()
will work or not, but passing backslashes around in strings in R is
generally very shaky ground.
}
\keyword{internal}
