% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{replace_star}
\alias{replace_star}
\title{replace_star}
\usage{
replace_star(.mod, .starred)
}
\arguments{
\item{.starred}{boolean variable to indicate significance}
}
\description{
replace_star
}
\section{Functions}{
\itemize{
\item \code{replace_star}: or remove star on a model object and corresponding YAML.
Tags can be modified at any time (i.e. before \emph{or} after a model is submitted).
}}

