% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{set_star_status}
\alias{set_star_status}
\title{set_star_status}
\usage{
set_star_status(.mod, .star)
}
\arguments{
\item{.mod}{list object of model data}

\item{.star}{boolean variable to indicate significance}
}
\description{
set_star_status
}
\section{Functions}{
\itemize{
\item \code{set_star_status}: or remove star on a model object and corresponding YAML.
}}

