% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-log.R
\name{add_dofv}
\alias{add_dofv}
\title{Calculate the difference in objective function value (ofv) between the
current and based_on model.}
\usage{
add_dofv(.log_df)
}
\arguments{
\item{.log_df}{The output of \code{\link[=summary_log]{summary_log()}}) or \code{run_log() \%>\% add_summary()}}
}
\description{
Calculate the difference in objective function value (ofv) between the
current and based_on model.
}
\keyword{internal}
