% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-file-to-model-dir.R
\name{add_file_to_model_dir}
\alias{add_file_to_model_dir}
\alias{add_stan_file}
\alias{add_standata_file}
\alias{add_stan_init}
\title{Attaches a file to a model}
\usage{
add_stan_file(.mod, .source_file = NULL)

add_standata_file(.mod, .source_file = NULL)

add_stan_init(.mod, .source_file = NULL)
}
\arguments{
\item{.mod}{a \verb{bbi_\{.model_type\}_model} object}

\item{.source_file}{If \code{NULL}, the default, create an empty scaffold file
at the destination path. If not \code{NULL}, pass a path to a file that
will be copied to the destination path. Use this if you have a
file elsewhere on disk that you would like to use for this model.}
}
\description{
These functions take a model object and create the appropriate file in the
required location. If \code{.source_file} argument is used, this file will be
copied to the location. Otherwise, a template "scaffold" of the required file
is created in that location. \strong{Note, this primarily intended for Stan
models} which require several necessary files to run. (See \verb{?}\link{bbi_stan_model}
for details about these files.) Users can call \code{check_stan_model(.mod)} to
see if any of these files are missing.
}
\section{Functions}{
\itemize{
\item \code{add_stan_file}: Adds a \verb{<run>.stan} model file.

\item \code{add_standata_file}: Adds a \verb{<run>-standata.R} file for building the data. See \verb{?}\link{bbi_stan_model} for details.

\item \code{add_stan_init}: Adds a \verb{<run>-init.R} file for building the initial values. See \verb{?}\link{bbi_stan_model} for details.
}}

