% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-log.R
\name{add_log_impl}
\alias{add_log_impl}
\title{Add columns to log df}
\usage{
add_log_impl(.log_df, .impl_func, ...)
}
\arguments{
\item{.log_df}{a \code{bbi_run_log_df} tibble (the output of \code{\link[=run_log]{run_log()}})}

\item{.impl_func}{Implementation function to extract the appropriate log tibble that will be joined against the input tibble.}

\item{...}{Arguments passed through to \code{.impl_func}}
}
\value{
The input \code{bbi_run_log_df} tibble, with any columns from the tibble output by \code{.impl_func} left joined onto it.
}
\description{
Private helper to extact columns from another log tibble and join them onto a \code{bbi_run_log_df}
}
\keyword{internal}
