% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-threads.R
\name{adjust_estimation_options}
\alias{adjust_estimation_options}
\title{Parse model files and overwrite estimation iterations}
\usage{
adjust_estimation_options(.mods, .cap_iterations)
}
\arguments{
\item{.mods}{a list of model objects. Generally created by \code{test_threads()}.}

\item{.cap_iterations}{Max number of iterations for NONMEM to run.}
}
\description{
Parse model files and overwrite estimation iterations
}
\keyword{internal}
