% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-run-utils.R
\name{analysis_estimates}
\alias{analysis_estimates}
\title{Tabulate parameter estimates for each model submission in an analysis run}
\usage{
analysis_estimates(.run, force_resummarize = FALSE, .bbi_args = NULL)
}
\arguments{
\item{.run}{A \code{bbi_nmboot_model} or \code{bbi_nmsse_model} object.}

\item{force_resummarize}{Logical (T/F). If \code{TRUE}, force re-summarization.}

\item{.bbi_args}{Named list passed to \code{model_summary(orig_mod, .bbi_args)},
where \code{orig_mod} is the model the analysis is based on. See
\code{\link[=print_bbi_args]{print_bbi_args()}} for valid options. Defaults to \code{list(no_grd_file = TRUE, no_shk_file = TRUE)} because \code{\link[=model_summary]{model_summary()}} is only called internally to
extract the number of records, so those files are irrelevant. Only used if
the based on model (the model the analysis is being performed on) has been
executed.}
}
\description{
Tabulate parameter estimates for each model submission in an analysis run
}
\keyword{internal}
