% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{BBI_ARGS}
\alias{BBI_ARGS}
\title{List of valid arguments bbi arguments}
\format{
An object of class \code{list} of length 33.
}
\usage{
BBI_ARGS
}
\description{
These values can be passed via the \code{.bbi_args} argument of \code{\link[=submit_model]{submit_model()}}
or \code{\link[=model_summary]{model_summary()}}. This is exported for the benefit of RStudio users that
may prefer to inspect \code{View(bbr::BBI_ARGS)} instead of the \code{\link[=print_bbi_args]{print_bbi_args()}}
output.
}
\keyword{internal}
