% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{bbi_current_release}
\alias{bbi_current_release}
\title{Get version number of babylon current release}
\usage{
bbi_current_release(os = "linux")
}
\arguments{
\item{os}{operating system, Default: \code{'linux'}}
}
\description{
Helper function to get version number of most recent release of babylon from GitHub.
}
