% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{bbi_current_release}
\alias{bbi_current_release}
\title{Get version number of bbi current release}
\usage{
bbi_current_release()
}
\description{
Helper function to get version number of most recent release of bbi from GitHub.
}
