% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{bbi_current_release}
\alias{bbi_current_release}
\title{current release}
\usage{
bbi_current_release(os = "linux")
}
\arguments{
\item{os}{operating system, Default: 'linux'}
}
\description{
current release
}
