% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{bbi_dry_run}
\alias{bbi_dry_run}
\title{Creates a \code{babylon_process} object with all the required keys, without actually running the command}
\usage{
bbi_dry_run(.cmd_args, .dir)
}
\arguments{
\item{.cmd_args}{A character vector of command line arguments for the execution call}

\item{.dir}{The working directory to run command in. Defaults to "."}
}
\description{
Returns an S3 object of class \code{babylon_process} but the \code{process} and \code{stdout} elements containing only the string "DRY_RUN".
Also contains the element \code{call} with a string representing the command that could be called on the command line.
}
