% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{bbi_dry_run}
\alias{bbi_dry_run}
\title{Babylon dry_run}
\usage{
bbi_dry_run(.cmd_args, .dir)
}
\arguments{
\item{.cmd_args}{A character vector of command line arguments for the execution call}

\item{.dir}{The working directory to run command in. Defaults to "."}
}
\description{
Creates a `babylon_process` object with all the required keys, without actually running the command.
Returns an S3 object of class `babylon_process` but the `process` and `stdout` elements containing only the string "DRY_RUN".
Also contains the element `call` with a string representing the command that could be called on the command line.
}
