% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{bbi_help}
\alias{bbi_help}
\title{Executes (`bbi --help`) with bbi_exec and prints the output string}
\usage{
bbi_help(.cmd_args = NULL)
}
\arguments{
\item{.cmd_args}{You can optionally pass a vector of args to get help about a specific call}
}
\description{
Executes (`bbi --help`) with bbi_exec and prints the output string
}
