% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbr.R
\name{bbi_init}
\alias{bbi_init}
\title{Initialize bbi}
\usage{
bbi_init(
  .dir,
  .nonmem_dir,
  .nonmem_version = NULL,
  .bbi_args = NULL,
  .no_default_version = FALSE
)
}
\arguments{
\item{.dir}{Path to directory to run \code{init} in (and put the resulting
\code{bbi.yaml} file)}

\item{.nonmem_dir}{Path to directory with the NONMEM installation.}

\item{.nonmem_version}{Character scalar for default version of NONMEM to use.
If left NULL, function will exit and tell you which versions were found in
\code{.nonmem_dir}}

\item{.bbi_args}{A named list specifying options to set in the newly created
\code{bbi.yaml} file, formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}. Run \code{\link[=print_bbi_args]{print_bbi_args()}} to see valid options. Note
that bbr does not support changing the output directory (including through
the model or global YAML files).}

\item{.no_default_version}{If \code{TRUE}, force creation of bbi.yaml with
\strong{no default NONMEM version}. \code{FALSE} by default, and using \code{TRUE} is
\emph{not} encouraged.}
}
\description{
Executes \verb{bbi init ...} in specified directory. This creates a \code{bbi.yaml}
file, which contains defaults for many configurable \code{bbi} settings, in
that directory.
}
\details{
For \code{bbr} to make any calls out to \code{bbi} (for example in
\code{\link[=submit_model]{submit_model()}} or \code{\link[=model_summary]{model_summary()}}) it must find a \code{bbi.yaml} file.
The default behavior is to look for this file in the same directory as the
model. \code{submit_model()} and \code{submit_models()} also support passing a
configuration file via the \code{.config_path} argument.
}
