% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{bbi_init}
\alias{bbi_init}
\title{Executes (`bbi init`) with processx::run in specified directory}
\usage{
bbi_init(.dir, .nonmem_dir = NULL)
}
\arguments{
\item{.dir}{Path to directory to run `init` in (and put the resulting `babylon.yml` file)}

\item{.nonmem_dir}{Path to directory with the NONMEM installation. Defaults to `getOption("rbabylon.nonmem_dir")`}
}
\description{
Executes (`bbi init`) with processx::run in specified directory
}
