% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{bbi_version}
\alias{bbi_version}
\title{bbi version}
\usage{
bbi_version(.bbi_exe_path = getOption("rbabylon.bbi_exe_path"))
}
\arguments{
\item{.bbi_exe_path}{Path to bbi exe file that will be checked}
}
\value{
character
}
\description{
Returns string of installed bbi cli version
}
\examples{
\dontrun{
bbi_version()
}
}
