% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{bbi_version}
\alias{bbi_version}
\title{Get version number of a babylon installation}
\usage{
bbi_version(.bbi_exe_path = get_bbi_exe_path())
}
\arguments{
\item{.bbi_exe_path}{Path to bbi exe file that will be checked. Defaults to `getOption('rbabylon.bbi_exe_path')`.}
}
\value{
character
}
\description{
Returns version number of bbi binary installed at path passed to `.bbi_exe_path`
}
\examples{
\dontrun{
bbi_version()
}
}
