% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-model-from.R
\name{build_new_model_path}
\alias{build_new_model_path}
\title{Private helper to build absolute path for \code{\link[=copy_model_from]{copy_model_from()}}.
Importantly, if the input \code{.new_model} is \emph{not} absolute, it will
be treated as relative to the working directory of \code{.parent_mod}.}
\usage{
build_new_model_path(.parent_mod, .new_model)
}
\arguments{
\item{.parent_mod}{Model to copy from}

\item{.new_model}{Path to the new model, either absolute or relative to the
path to \code{.parent_model}. Represents an absolute model path, which is the
path to the YAML file and model file, both without extension, and the
output directory (once the model is run). Numeric values will be coerced to
character. See examples for usage.}
}
\value{
absolute file path to save new model to (without file extension)
}
\description{
Private helper to build absolute path for \code{\link[=copy_model_from]{copy_model_from()}}.
Importantly, if the input \code{.new_model} is \emph{not} absolute, it will
be treated as relative to the working directory of \code{.parent_mod}.
}
\keyword{internal}
