% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{build_path_from_model}
\alias{build_path_from_model}
\alias{build_path_from_model.bbi_model}
\title{Build path to a model or output file}
\usage{
build_path_from_model(.mod, .suffix, ...)

\method{build_path_from_model}{bbi_model}(.mod, .suffix, ...)
}
\arguments{
\item{.mod}{Model to use, either a \verb{bbi_\{.model_type\}_model} or \verb{bbi_\{.model_type\}_summary} object.}

\item{.suffix}{Character vector to append the end of the absolute model path.
Will be appended \emph{as is} so, if passing a file extension, be sure to included the leading \code{"."}.
See examples.}

\item{...}{arguments passed through to methods. (Currently none.)}
}
\value{
Returns an absolute path to \verb{\{.mod$absolute_model_path\}/\{get_model_id(.mod)\}\{.suffix\}}.
Does \emph{not} check whether the file exists.
}
\description{
Builds the absolute path to a file in the model directory from components of
the \verb{bbi_\{.model_type\}_model} object. For NONMEM models this is used to
construct paths to files in the \emph{output} directory (i.e. the \code{.lst} or \code{.grd}
files). For Stan models this is used to construct paths to the files that
define the model (i.e. \code{-standata.R} or \code{-init.R}).
}
\examples{
.mod <- read_model(
  system.file("model", "nonmem", "basic", "1", package = "bbr")
)
build_path_from_model(.mod, ".lst")
build_path_from_model(.mod, "-standata.R")

}
