% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{build_path_from_model}
\alias{build_path_from_model}
\alias{build_path_from_model.bbi_nonmem_model}
\alias{build_path_from_model.bbi_nonmem_summary}
\title{Build path to output file}
\usage{
build_path_from_model(.mod, .extension, ...)

\method{build_path_from_model}{bbi_nonmem_model}(.mod, .extension, ...)

\method{build_path_from_model}{bbi_nonmem_summary}(.mod, .extension, ...)
}
\arguments{
\item{.mod}{Model to use, either a \verb{bbi_\{.model_type\}_model} or \verb{bbi_\{.model_type\}_summary} object.}

\item{.extension}{file extension to append (for example \code{lst}, \code{ext}, \code{grd}, etc.).
Appended with \code{\link[fs:path_file]{fs::path_ext_set()}} so can either begin with a \code{"."} or not. See \code{fs} docs for more details.}

\item{...}{arguments passed through to methods. (Currently none.)}
}
\description{
Builds the absolute path to a file in the output directory from components of the \verb{bbi_\{.model_type\}_model} object
}
\section{Methods (by class)}{
\itemize{
\item \code{bbi_nonmem_model}: Returns an absolute path to \verb{\{output_dir\}/\{model_id\}.\{.extension\}}.
Does \emph{not} check whether the file exists.

\item \code{bbi_nonmem_summary}: Same as \code{bbi_nonmem_model} method.
}}

