% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{build_path_from_model}
\alias{build_path_from_model}
\alias{build_path_from_model.bbi_model}
\title{Build path to output file}
\usage{
build_path_from_model(.mod, .suffix, ...)

\method{build_path_from_model}{bbi_model}(.mod, .suffix, ...)
}
\arguments{
\item{.mod}{Model to use, either a \verb{bbi_\{.model_type\}_model} or \verb{bbi_\{.model_type\}_summary} object.}

\item{.suffix}{Character vector to append the end of the absolute model path.
Will be appended \emph{as is} so, if passing a file extension, be sure to included the leading \code{"."}.
See examples.}

\item{...}{arguments passed through to methods. (Currently none.)}
}
\value{
Returns an absolute path to \verb{\{output_dir\}/\{model_id\}\{.suffix\}}.
Does \emph{not} check whether the file exists.
}
\description{
Builds the absolute path to a file in the output directory from components of the \verb{bbi_\{.model_type\}_model} object
}
\examples{
.mod <- read_model(
  system.file("model", "nonmem", "basic", "1", package = "bbr")
)
build_path_from_model(.mod, ".lst")
build_path_from_model(.mod, "-standata.R")

}
