% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbr.R
\name{check_bbi_exe}
\alias{check_bbi_exe}
\title{Check that bbi is installed}
\usage{
check_bbi_exe(.bbi_exe_path)
}
\arguments{
\item{.bbi_exe_path}{Path to bbi executable}
}
\value{
\code{NULL}, invisibly
}
\description{
Raises an error if a bbi executable is not found at \code{.bbi_exe_path}.
}
\keyword{internal}
