% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{check_bbi_exe}
\alias{check_bbi_exe}
\title{Checks that a bbi binary is present at the path passed to .bbi_exe_path}
\usage{
check_bbi_exe(.bbi_exe_path)
}
\arguments{
\item{.bbi_exe_path}{Path to bbi exe file that will be checked}
}
\description{
Checks that a bbi binary is present at the path passed to .bbi_exe_path
}
