% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{check_bbi_version_constraint}
\alias{check_bbi_version_constraint}
\title{Check if bbi_version is below minimum allowed version}
\usage{
check_bbi_version_constraint(.bbi_exe_path = get_bbi_exe_path())
}
\arguments{
\item{.bbi_exe_path}{Path to bbi exe file that will be checked}
}
\description{
Check if bbi_version is below minimum allowed version
}
