% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_file}
\alias{check_file}
\alias{tail_output}
\alias{tail_output.character}
\alias{tail_output.bbi_nonmem_model}
\alias{tail_lst}
\alias{tail_lst.character}
\alias{tail_lst.bbi_nonmem_model}
\title{Check an output file}
\usage{
check_file(.file, .head = 3, .tail = 5, .print = TRUE, .return = FALSE)

tail_output(.mod, ...)

\method{tail_output}{character}(.mod, .head = 3, .tail = 5, .print = TRUE, .return = FALSE, ...)

\method{tail_output}{bbi_nonmem_model}(.mod, .head = 3, .tail = 5, .print = TRUE, .return = FALSE, ...)

tail_lst(.mod, ...)

\method{tail_lst}{character}(.mod, .head = 3, .tail = 5, .print = TRUE, .return = FALSE, ...)

\method{tail_lst}{bbi_nonmem_model}(.mod, .head = 3, .tail = 5, .print = TRUE, .return = FALSE, ...)
}
\arguments{
\item{.file}{Character scalar of path to file to read}

\item{.head}{Integer for number of lines to read from the top of the file}

\item{.tail}{Integer for number of lines to read from the bottom of the file}

\item{.print}{If \code{TRUE}, the default, print resulting head and tail to console.}

\item{.return}{If \code{FALSE}, the default, returns \code{NULL} invisibly. If \code{TRUE} returns resulting head and tail as a character vector.}

\item{.mod}{Model object or path associated with file.}

\item{...}{arguments passed through to \code{check_file()}}
}
\description{
Reads the head and tail of specified file and prints it the console and/or returns it as a character vector.
This is called internally in several S3 methods described below.
}
\section{Methods (by generic)}{
\itemize{
\item \code{tail_output}: Tail the OUTPUT file from a file path to a model.

\item \code{tail_output}: Tail the OUTPUT file from a \code{bbi_nonmem_model} object.

\item \code{tail_lst}: Tail the .lst file from a file path to a model.

\item \code{tail_lst}: Tail the .lst file from a \code{bbi_nonmem_model} object.
}}

