% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-summary.R
\name{check_lst_file}
\alias{check_lst_file}
\title{Private helper function to look for .lst function in a directory}
\usage{
check_lst_file(.x)
}
\arguments{
\item{.x}{The directory path to look in for the lst file}
}
\description{
Private helper function to look for .lst function in a directory
}
\keyword{internal}
