% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{check_model_for_sim}
\alias{check_model_for_sim}
\title{Check that a \code{bbr} model can be used for simulation}
\usage{
check_model_for_sim(.mod, .join_col = "NUM")
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} object.}

\item{.join_col}{Character column name(s) used to join table files post
execution. Gets appended to the generated \verb{$TABLE} record. See
\code{\link[=nm_join_sim]{nm_join_sim()}} documentation for details. Defaults to \code{'NUM'}.}
}
\value{
\code{TRUE} invisibly
}
\description{
Checks that a \code{bbr} model object is in the correct format to generate a
\code{bbi_nmsim_model} object.
}
\keyword{internal}
