% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_model_object}
\alias{check_model_object}
\title{Private helper to check if an object inherits a model class and error if not}
\usage{
check_model_object(.mod, .mod_types = VALID_MOD_CLASSES)
}
\arguments{
\item{.mod}{The object to check}

\item{.mod_types}{Character vector of acceptable classes, defaulting to \code{VALID_MOD_CLASSES}}
}
\description{
Private helper to check if an object inherits a model class and error if not
}
\keyword{internal}
