% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_nonmem_finished}
\alias{check_nonmem_finished}
\alias{check_nonmem_finished.character}
\alias{check_nonmem_finished.bbi_nonmem_model}
\title{Checks if NONMEM run is done by looking for "Stop Time" in .lst file}
\usage{
check_nonmem_finished(mod)

\method{check_nonmem_finished}{character}(mod)

\method{check_nonmem_finished}{bbi_nonmem_model}(mod)
}
\arguments{
\item{mod}{bbi model object}
}
\description{
Returns \code{TRUE} if the model appears to be finished running and \code{FALSE} otherwise.
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: takes a file path (character string) of \code{NONMEM} \code{.lst} file.

\item \code{bbi_nonmem_model}: takes a \code{bbi_nonmem_model} object.
}}

