% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_nonmem_table_output}
\alias{check_nonmem_table_output}
\title{Checks nonmem output file that's a whitespace-delimited file (for instance .grd or .ext)}
\usage{
check_nonmem_table_output(.path, .x_var = NULL, .x_floor = NULL)
}
\arguments{
\item{.path}{Character scalar path to the gradient file}

\item{.x_var}{name of variable to filter with `.x_floor`}

\item{.x_floor}{Filters file to only rows with `.x_var` GREATER THAN this value.}
}
\description{
Checks nonmem output file that's a whitespace-delimited file (for instance .grd or .ext)
}
