% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_nonmem_table_output}
\alias{check_nonmem_table_output}
\alias{check_grd}
\alias{check_grd.character}
\alias{check_grd.bbi_nonmem_model}
\alias{check_grd.bbi_nonmem_summary}
\alias{check_ext}
\alias{check_ext.character}
\alias{check_ext.bbi_nonmem_model}
\alias{check_ext.bbi_nonmem_summary}
\title{Check NONMEM output files}
\usage{
check_nonmem_table_output(.path, .x_var = NULL, .x_floor = NULL)

check_grd(.mod, .iter_floor = 0)

\method{check_grd}{character}(.mod, .iter_floor = 0)

\method{check_grd}{bbi_nonmem_model}(.mod, .iter_floor = 0)

\method{check_grd}{bbi_nonmem_summary}(.mod, .iter_floor = 0)

check_ext(.mod, .iter_floor = 0)

\method{check_ext}{character}(.mod, .iter_floor = 0)

\method{check_ext}{bbi_nonmem_model}(.mod, .iter_floor = 0)

\method{check_ext}{bbi_nonmem_summary}(.mod, .iter_floor = 0)
}
\arguments{
\item{.path}{Character scalar path to the file}

\item{.x_var}{name of variable to filter with \code{.x_floor}}

\item{.x_floor}{Filters file to only rows with \code{.x_var} GREATER THAN this value.}

\item{.mod}{Model to check. Either a \code{bbi_nonmem_model} object or a file path.}

\item{.iter_floor}{Filters file to only rows with \code{ITERATION} GREATER THAN this value.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Checks a NONMEM output file that's a whitespace-delimited file (for instance .grd or .ext)
}
\section{Functions}{
\itemize{
\item \code{check_grd(character)}: Checks .grd file from a file path

\item \code{check_grd(bbi_nonmem_model)}: Checks .grd file from a \code{bbi_nonmem_model}

\item \code{check_grd(bbi_nonmem_summary)}: Checks .grd file from a \code{bbi_nonmem_summary}

\item \code{check_ext(character)}: Checks .ext file from a file path

\item \code{check_ext(bbi_nonmem_model)}: Checks .ext file from a \code{bbi_nonmem_model} object

\item \code{check_ext(bbi_nonmem_summary)}: Checks .ext file from a \code{bbi_nonmem_summary} object

}}
