% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_output_dir}
\alias{check_output_dir}
\alias{check_output_dir.character}
\alias{check_output_dir.bbi_nonmem_model}
\title{List files in the output directory to glance at where the process is}
\usage{
check_output_dir(.mod, ...)

\method{check_output_dir}{character}(.output_dir, .filter = NULL)

\method{check_output_dir}{bbi_nonmem_model}(.mod, .filter = NULL)
}
\arguments{
\item{.mod}{The `bbi_{.model_type}_model` object}

\item{.output_dir}{Character scaler of path to output directory}

\item{.filter}{Optional Character scaler of regex to filter filenames on and only return matches}
}
\description{
List files in the output directory to glance at where the process is

S3 dispatch to list output directory files from a directory path

S3 dispatch to list output directory files from a `bbi_{.model_type}_model` object
}
