% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{check_output_dir}
\alias{check_output_dir}
\alias{check_output_dir.character}
\alias{check_output_dir.bbi_nonmem_model}
\title{List files in the output directory to glance at where the process is}
\usage{
check_output_dir(.mod, ...)

\method{check_output_dir}{character}(.mod, ...)

\method{check_output_dir}{bbi_nonmem_model}(.mod, ...)
}
\arguments{
\item{.mod}{The `bbi_{.model_type}_model` object}

\item{...}{values to pass to `fs::dir_ls`}
}
\description{
List files in the output directory to glance at where the process is

S3 dispatch to list output directory files from a directory path

S3 dispatch to list output directory files from a `bbi_{.model_type}_model` object
}
