% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{check_status_code}
\alias{check_status_code}
\title{Checks status code from processx::run output}
\usage{
check_status_code(.output, .cmd_args)
}
\arguments{
\item{.model_type}{Type of model to summarize. Currently only supports "nonmem"}
}
\description{
Checks status code from processx::run output
}
