% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{check_status_code}
\alias{check_status_code}
\title{Checks status code from processx process}
\usage{
check_status_code(.status_code, .output, .cmd_args)
}
\arguments{
\item{.status_code}{numerical status code from the process}

\item{.output}{character scalar of output from the process. Used only for error printing.}

\item{.cmd_args}{character vector of args passed to process. Used only for error printing.}
}
\description{
Checks status code from processx process
}
