% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-up-to-date.R
\name{check_up_to_date}
\alias{check_up_to_date}
\alias{check_up_to_date.bbi_base_model}
\alias{check_up_to_date.bbi_nmboot_model}
\alias{check_up_to_date.bbi_nmsse_model}
\alias{check_up_to_date.bbi_nonmem_summary}
\title{Check model and data up to date with outputs}
\usage{
check_up_to_date(.bbi_object, ...)

\method{check_up_to_date}{bbi_base_model}(.bbi_object, ...)

\method{check_up_to_date}{bbi_nmboot_model}(.bbi_object, ...)

\method{check_up_to_date}{bbi_nmsse_model}(.bbi_object, ...)

\method{check_up_to_date}{bbi_nonmem_summary}(.bbi_object, ...)
}
\arguments{
\item{.bbi_object}{the object to check. Could be
a \verb{bbi_\{.model_type\}_model} object,
a \verb{bbi_\{.model_type\}_summary} object,
or a \code{bbi_log_df} tibble.}

\item{...}{Arguments passed through (currently none).}
}
\value{
\strong{The returned value is invisible because a message is printed} alerting the
user of the specific files that have changed, if any. This facilitates
calling the function for this side effect without explicitly handling the
returned value.

\strong{\code{bbi_model}} method invisibly returns a logical vector of length 2. The
first element (named \code{"model"}) refers to the model files mentioned in Details.
The second element (named \code{"data"}) refers to the data files mentioned in Details.
For both elements, they will be \code{TRUE} if nothing has changed, \code{FALSE} if
anything has changed. Note: \emph{if no file exists} at the specified path,
\code{FALSE} will be returned because that is technically a "change." The file
used to exist and now it does not.

\strong{\code{bbi_log_df}} method invisibly returns a named list of lists, with one
element for each row in the input tibble, with the name corresponding to the
value in the \code{run} column for that row. Each element of the list will contain
the two-element list returned from the \code{bbi_model} method (described above)
for the relevant model.

There is no \code{add_up_to_date()} function because \strong{if you would like to add
these columns to a \code{bbi_log_df} tibble} you can use \code{\link[=add_config]{add_config()}}, which
contains \code{model_has_changed} and \code{data_has_changed} columns. Please note:
these contain the opposite boolean values (\code{check_up_to_date()} returns
\code{TRUE} if up to date, \verb{*_has_changed} returns \code{TRUE} if \emph{changed}).
}
\description{
Functions for checking that the model outputs on disk match the md5 hashes
stored in \code{bbi_config.json} at run time. In other words, checking that the
model and data files have \emph{not} changed since the model was last run. See
Details section for what specific files are checked.
}
\details{
Different files are checked depending on what type of model is being checked.

\strong{For NONMEM models}
\itemize{
\item The model file (control stream)
\item The data file (referenced in \verb{$DATA} within the control stream)
}

\strong{For bootstrap runs}
\itemize{
\item The model file (control stream)
\item The original data file used in the setup of the bootstrap run (referenced
in \verb{$DATA} within the control stream)
}
}
