% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{cleanup_bootstrap_run}
\alias{cleanup_bootstrap_run}
\title{Cleanup bootstrap run directory}
\usage{
cleanup_bootstrap_run(.boot_run)
}
\arguments{
\item{.boot_run}{a \code{bbi_nmboot_model} object.}
}
\description{
This will delete all child models, and only keep the information
you need to read in estimates or summary information
}
