% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-threads.R
\name{cleanup_mods}
\alias{cleanup_mods}
\title{Remove model files associated with the specified tags}
\usage{
cleanup_mods(.mods, .tags = "test threads")
}
\arguments{
\item{.mods}{a bbi model object or list of model objects.}

\item{.tags}{a character vector identifying the tags of the models you want to delete.
If set to NULL, delete all associated model files.}
}
\description{
Remove model files associated with the specified tags
}
