% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_list_objects}
\alias{combine_list_objects}
\title{Combines two named lists. By default, shared keys are overwritten by that key in .new_list
Any S3 classes will be inherited with those from .new_list given precedence.}
\usage{
combine_list_objects(.new_list, .old_list, .append = FALSE)
}
\arguments{
\item{.new_list}{A named list}

\item{.old_list}{A named list}

\item{.append}{Shared keys will have their contents concatenated instead of overwriting from .new_list}
}
\value{
The combined list
}
\description{
Combines two named lists. By default, shared keys are overwritten by that key in .new_list
Any S3 classes will be inherited with those from .new_list given precedence.
}
