% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model-stan-helpers.R
\name{compile_stanmod}
\alias{compile_stanmod}
\title{Private helper to compile a stan model and save a gitignore that ignores the
binary and posterior csv's}
\usage{
compile_stanmod(.mod)
}
\description{
Private helper to compile a stan model and save a gitignore that ignores the
binary and posterior csv's
}
\keyword{internal}
