% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-log.R
\name{config_log}
\alias{config_log}
\alias{add_config}
\title{Parse bbi configs to log}
\usage{
config_log(.base_dir, .recurse = FALSE, .include = NULL)

add_config(.log_df)
}
\arguments{
\item{.base_dir}{Base directory to look in for models.}

\item{.recurse}{If \code{FALSE}, the default, only include models in \code{.base_dir}
but \emph{not} subdirectories. This is passed through to \code{fs::dir_ls()} --
If \code{TRUE} recurse fully, if a positive number the number of levels to recurse.}

\item{.include}{A character vector specifying which runs or model tags to include in the run log.}

\item{.log_df}{A \code{bbi_run_log_df} tibble (the output of \code{\link[=run_log]{run_log()}}).}
}
\value{
An object of class \code{bbi_config_log_df}, which includes the fields described below.
If \emph{no} \code{bbi_config.json} files are found, the returned tibble will only contain the
\code{absolute_model_path} and \code{run} columns, and will have 0 rows.

\code{config_log()} creates a new tibble with one row per \code{bbi_config.json}
found in \code{.base_dir} (and subdirectories, if \code{.recurse = TRUE}).

\code{add_config()} adds these fields to the tibble passed to \code{.log_df}.
}
\description{
Extracts selected fields from \code{bbi_config.json}, which is created by bbi
in the model output folder to store metadata about the execution of a model
run.
}
\details{
The following fields from are extracted from \code{bbi_config.json}.
\itemize{
\item \code{absolute_model_path}: the path to the model file, excluding the file
extension
\item \code{model_md5}: the MD5 sum of the model file
\item \code{data_path}: the path to the data file, relative to the model's output directory
(which can be extracted via \code{\link[=get_output_dir]{get_output_dir()}})
\item \code{data_md5}: the MD5 sum of the data file
\item \code{bbi_version}: the version of bbi last used to run the model
\item \code{nm_version}: the version of NONMEM last used to run the model
\item \code{model_has_changed}: a logical indicating whether the model file has
changed since it was last run
\item \code{data_has_changed}: a logical indicating whether the data file has
changed since the model was last run
}
}
\seealso{
\code{\link[=run_log]{run_log()}}, \code{\link[=summary_log]{summary_log()}}
}
