% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{convert_psn_data_path}
\alias{convert_psn_data_path}
\title{Convert data path during conversion of PsN to bbr model submission}
\usage{
convert_psn_data_path(.mod_path, .bbr_run_dir)
}
\arguments{
\item{.mod_path}{Path to a NONMEM control stream file.}

\item{.bbr_run_dir}{Directory to copy the \strong{generated bbr run files} to. Subdirectory of \code{.bbr_dir}}
}
\description{
Convert data path during conversion of PsN to bbr model submission
}
\keyword{internal}
