% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-model-from.R
\name{copy_control_stream}
\alias{copy_control_stream}
\title{Copy a NONMEM control stream file}
\usage{
copy_control_stream(
  .parent_model_path,
  .new_model_path,
  .update_model_file = FALSE,
  .description = NULL
)
}
\arguments{
\item{.parent_model_path}{Path to the control stream to copy}

\item{.new_model_path}{Path to copy the new control stream to}

\item{.update_model_file}{If \code{TRUE}, the default, update the \verb{$PROBLEM} line in the new control stream. If \code{FALSE}, \verb{\{.new_model\}.[mod|ctl]} will be an exact copy of its parent control stream.}

\item{.description}{Description of new model run. This will be passed into the \verb{$PROBLEM} of the new control stream (if \code{.update_model_file=TRUE}).}
}
\description{
Helper function to copy a NONMEM control stream file and optionally update the description, etc. in the new file.
Note that any file existing at \code{.new_model_path} will be overwritten.
}
\keyword{internal}
