% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{copy_model_from}
\alias{copy_model_from}
\alias{copy_model_from.bbi_nonmem_model}
\alias{copy_model_from.character}
\alias{copy_nonmem_model_from}
\title{Generic S3 method from iterating on models.}
\usage{
copy_model_from(.x, ...)

\method{copy_model_from}{bbi_nonmem_model}(.parent_mod, .new_model, .description, ...)

\method{copy_model_from}{character}(.parent_path, .new_model, .description, ...)

copy_nonmem_model_from(
  .parent_mod,
  .new_model,
  .description,
  .based_on_additional = NULL,
  .add_tags = NULL,
  .inherit_tags = FALSE,
  .update_mod_file = TRUE
)
}
\arguments{
\item{.x}{Object to copy from}

\item{...}{arguments to pass through to implementation method}

\item{.parent_mod}{S3 object of class `bbi_nonmem_model` to be used as the basis for copy.}

\item{.new_model}{Path to write new model files to WITHOUT FILE EXTENSION. Function will create both `{.new_model}.yaml` and `{.new_model}.[mod|ctl]` based on this path.}

\item{.description}{Description of new model run. This will be stored in the yaml (to be used later in `create_run_log()`) and optionally passed into the `$PROBLEM` of the new control stream.}

\item{.parent_path}{Path to YAML file to copy model from}

\item{.based_on_additional}{The run id for the `.parent_model` will automatically be added to the `based_on` field but this argument can contain a character scaler or vector of additional run id's (model names) that this model was "based on." These are used to reconstuct model developement and ancestry.}

\item{.add_tags}{A character scaler or vector with any new tags to be added to `{.new_model}.yaml`}

\item{.inherit_tags}{Boolean for whether to inherit any tags from `.parent_model.yaml`}

\item{.update_mod_file}{Boolean for whether to update the `$PROBLEM` line in the new control stream. By default it is TRUE, but if FALSE is passed `{.new_model}.[mod|ctl]` will be an exact copy of it's parent control stream.}
}
\value{
S3 object of class `bbi_nonmem_model` that can be passed to `submit_nonmem_model()`
}
\description{
Create new .mod/ctl and new .yaml files based on a previous model. Used for iterating on model development.
Also fills in necessary YAML fields for using `create_run_log()` later.
}
