% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{copy_model_from}
\alias{copy_model_from}
\title{Create new .mod/ctl and new .yaml files based on a previous model. Used for iterating on model development.
Also fills in necessary YAML fields for using `create_run_log()` later.}
\usage{
copy_model_from(
  .parent_model,
  .new_model,
  .description,
  .based_on_additional = NULL,
  .inherit_tags = T,
  .add_tags = NULL
)
}
\value{
Character scaler with the name of the model .yaml file that can be passed to `submit_nonmem_model()`
}
\description{
Create new .mod/ctl and new .yaml files based on a previous model. Used for iterating on model development.
Also fills in necessary YAML fields for using `create_run_log()` later.
}
