% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{copy_psn_run_files}
\alias{copy_psn_run_files}
\title{Copy PsN run files from control from model and run directory}
\usage{
copy_psn_run_files(.psn_run_dir, .bbr_run_dir, .overwrite = FALSE)
}
\arguments{
\item{.psn_run_dir}{PsN execution directory where the model was run (e.g., \code{modelfit_dir/NM_run1}).}

\item{.bbr_run_dir}{Directory to copy the \strong{generated bbr run files} to. Subdirectory of \code{.bbr_dir}}

\item{.overwrite}{Logical (T/F). Whether to overwrite files if they already exist.}
}
\description{
Copy PsN run files from control from model and run directory
}
\details{
Following a PsN run, some run files are stored in the upper level "model" directory (\code{.psn_mod_dir}), such
as \code{run.cov}, \code{run.cor}, etc. However, some important run files only exist in the \code{NM_run} subdirectory, such as
\code{psn.shk} or \code{psn.grd}.

All these files exist in \code{NM_run}, so we only need that directory. These files will be copied to \code{.bbr_run_dir},
and will be renamed to have the model name be part of the file path.
}
\keyword{internal}
