% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{copy_psn_tables}
\alias{copy_psn_tables}
\title{Get NONMEM table files from control file}
\usage{
copy_psn_tables(.mod_path, .psn_mod_dir, .bbr_run_dir, .overwrite = FALSE)
}
\arguments{
\item{.mod_path}{Path to a NONMEM control stream file.}

\item{.psn_mod_dir}{Directory containing where the model and table files exist.
Note that the model file must have a \code{.mod} or \code{.ctl} file extension.}

\item{.bbr_run_dir}{Directory to copy the generated bbr run files to.}

\item{.overwrite}{logical (T/F). Whether to overwrite files if they already exist.}
}
\value{
a vector of file paths
}
\description{
Get NONMEM table files from control file
}
\keyword{internal}
