% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{copy_sigmas}
\alias{copy_sigmas}
\title{Copy sigma records from one model to another}
\usage{
copy_sigmas(.mod_lines, .new_sigmas, .bounds_opts)
}
\arguments{
\item{.mod_lines}{lines of ctl file. Must be an \code{nmrec} \code{nmrec_ctl_records} object.}

\item{.new_sigmas}{list of new sigma matrices, or a single sigma matrix}
}
\description{
Copy sigma records from one model to another
}
\keyword{internal}
