% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-model-from.R
\name{copy_stan_files}
\alias{copy_stan_files}
\title{Copy necessary Stan files}
\usage{
copy_stan_files(.parent_mod, .new_model, .overwrite)
}
\arguments{
\item{.parent_mod}{a \code{bbi_stan_model} object to copy from}

\item{.new_model}{Path to new model directory}

\item{.overwrite}{If \code{TRUE}, overwrite existing directory at \code{.new_model}. If \code{FALSE} and directory exists at \code{.new_model} error.}
}
\description{
Helper function to copy a necessary Stan files from parent model to new model
and rename them appropriately. Note that any directory existing at
\code{.new_model} will be overwritten by default.
}
\keyword{internal}
