% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{copy_thetas}
\alias{copy_thetas}
\title{Copy theta records from one model to another}
\usage{
copy_thetas(.mod_lines, .new_thetas, .bounds_opts)
}
\arguments{
\item{.mod_lines}{lines of ctl file. Must be an \code{nmrec} \code{nmrec_ctl_records} object.}

\item{.new_thetas}{list of new theta vectors, or a single theta vector}
}
\description{
Copy theta records from one model to another
}
\keyword{internal}
