% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crash_model_run}
\alias{crash_model_run}
\title{Crash a specific job}
\usage{
crash_model_run(run_name)
}
\arguments{
\item{run_name}{character vector. The \code{name} (or names) of the job(s) you want to end. Run qstat in your terminal to confirm.}
}
\description{
Crash a specific job
}
