% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create_model}
\alias{create_model}
\title{Create new model spec from a model yaml}
\usage{
create_model(.yaml_path)
}
\value{
S3 object of class `bbi_nonmem_spec` that can be passed to `submit_nonmem_model()`
}
\description{
Create new model spec from a model yaml
}
