% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{create_model_from_yaml}
\alias{create_model_from_yaml}
\title{Create new model spec from a model yaml}
\usage{
create_model_from_yaml(.yaml_path, .model_type = c("nonmem"))
}
\value{
S3 object of class `bbi_nonmem_spec` that can be passed to `submit_nonmem_model()`
}
\description{
Create new model spec from a model yaml
}
