% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_model_object}
\alias{create_model_object}
\alias{create_summary_object}
\alias{create_process_object}
\alias{create_run_log_object}
\title{Creates S3 object of class `bbi_{.model_type}_model` from list with `MODEL_REQ_INPUT_KEYS`}
\usage{
create_model_object(.mod_list)

create_summary_object(res, .model_type = SUPPORTED_MOD_TYPES)

create_process_object(res)

create_run_log_object(log_df)
}
\arguments{
\item{.mod_list}{List with the required information to create a model object}

\item{res}{List to attempt to assign the class to}

\item{.model_type}{Character scaler of a valid model type (currently either `nonmem` or `stan`)}

\item{log_df}{data.frame or tibble to attempt to assign the class to}
}
\value{
S3 object of class `bbi_{.model_type}_model` that can be passed to `submit_model()`, `model_summary()` etc.
}
\description{
First checks to be sure the required keys are all present

First checks to be sure the required keys are all present

First checks to be sure the required columns are all present
}
