% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert-psn.R
\name{create_psn_json}
\alias{create_psn_json}
\title{Create a fake bbi_config.json file for a previous PsN model submission}
\usage{
create_psn_json(.mod_path, .psn_mod_dir, .bbr_run_dir, .psn_info)
}
\arguments{
\item{.mod_path}{Path to a NONMEM control stream file.}

\item{.psn_mod_dir}{Directory containing where the model and table files exist.
Note that the model file must have a \code{.mod} or \code{.ctl} file extension.}

\item{.bbr_run_dir}{Directory to copy the generated bbr run files to.}
}
\description{
Create a fake bbi_config.json file for a previous PsN model submission
}
\keyword{internal}
