% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{current_release}
\alias{current_release}
\title{Private helper function to most recent release version from repo}
\usage{
current_release(
  owner = "metrumresearchgroup",
  repo = "babylon",
  os = c("linux", "darwin", "windows")
)
}
\arguments{
\item{owner}{Repository owner/organization}

\item{repo}{Repository name}

\item{os}{Operating system user intends to install on}
}
\description{
Private helper function to most recent release version from repo
}
\keyword{internal}
