% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-log.R
\name{enforce_length}
\alias{enforce_length}
\title{Helper to enforce length of a list element}
\usage{
enforce_length(.l, .k, .len = 1)
}
\arguments{
\item{.l}{list to check}

\item{.k}{key to check}

\item{.len}{length to enforce. Defaults to 1. Throws an error if `length(.l[[.k]]) != .len`}
}
\description{
Helper to enforce length of a list element
}
