% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-nmtran.R
\name{execute_nmtran}
\alias{execute_nmtran}
\title{Execute \code{NM-TRAN} in a given directory}
\usage{
execute_nmtran(
  nmtran_exe,
  mod_path,
  cmd_args = c("0", "0", "2"),
  nmtran_presort_exe = NULL,
  dir = "."
)
}
\arguments{
\item{nmtran_exe}{Path to \code{NM-TRAN} executable.}

\item{mod_path}{Path of a model to evaluate. Should be relative to \code{dir}. Nested
directories not supported.}

\item{cmd_args}{A character vector of command line arguments for the \code{NM-TRAN}
execution call}

\item{nmtran_presort_exe}{Path to \code{nmtran_presort} executable. Only available
for \code{NONMEM} versions \code{nm74gf}, \code{nm74gf_nmfe}, and \code{nm75}. If provided, will run
\code{nmtran_presort} before running \code{NM-TRAN}. Set to \code{NULL} to skip this step.}

\item{dir}{Directory in which to execute the command.}
}
\description{
Execute \code{NM-TRAN} in a given directory. Also runs \code{nmtran_presort} if an
executable is found.
}
\keyword{internal}
