% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-nmtran.R
\name{execute_nmtran}
\alias{execute_nmtran}
\title{Execute \code{NM-TRAN} in a given directory}
\usage{
execute_nmtran(nmtran_exe, mod_path, cmd_args = NULL, dir = ".")
}
\arguments{
\item{nmtran_exe}{Path to \code{NM-TRAN} executable.}

\item{mod_path}{Path of a model to evaluate. Should be relative to \code{dir}.}

\item{cmd_args}{A character vector of command line arguments for the \code{NM-TRAN}
execution call}

\item{dir}{Directory in which to execute the command.}
}
\description{
Execute \code{NM-TRAN} in a given directory
}
\keyword{internal}
