% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-log.R
\name{find_models}
\alias{find_models}
\title{Search for model YAML files and read them}
\usage{
find_models(.base_dir, .recurse, .include)
}
\arguments{
\item{.base_dir}{Directory to search for model YAML files.}

\item{.recurse}{If \code{TRUE} search recursively in subdirectories as well.}

\item{.include}{A character vector specifying which runs or model tags to include in the log
Provides filter for runs based on an input vector.}
}
\description{
Private helper function that searches from a base directory for any YAML files (excluding \code{bbi.yaml})
and attempts to read them to a model object with \code{\link[=safe_read_model]{safe_read_model()}}.
}
\keyword{internal}
