% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{find_yaml_file_path}
\alias{find_yaml_file_path}
\title{helper to find valid yaml file}
\usage{
find_yaml_file_path(.path)
}
\arguments{
\item{.path}{File path with any extension or no extension.
Function will search for that path with either a .yaml or .yml extension (preferring .yaml) and error if neither is found.}
}
\description{
helper to find valid yaml file
}
