% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{format_model_type}
\alias{format_model_type}
\title{Format the model type for use in \code{\link[=model_tree]{model_tree()}}}
\usage{
format_model_type(model_type, fmt_html = FALSE, ...)
}
\arguments{
\item{fmt_html}{Logical (\code{T}/\code{F}). If \code{TRUE}, apply coloring and format as \code{HTML}.}

\item{...}{Additional arguments passed to \code{\link[=style_html]{style_html()}}.}

\item{.mod}{a \code{bbi_base_model} object.}

\item{msg}{Character string. Appended to the end of the formatted model type
if provided.}
}
\description{
Format the model type for use in \code{\link[=model_tree]{model_tree()}}
}
\keyword{internal}
