% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_boot_models}
\alias{get_boot_models}
\title{Read in all bootstrap run model objects}
\usage{
get_boot_models(.boot_run)
}
\arguments{
\item{.boot_run}{a \code{bbi_nmboot_model} model object}
}
\description{
Read in all bootstrap run model objects
}
