% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_boot_spec}
\alias{get_boot_spec}
\title{Tabulate all relevant \code{bbi_nonmem_model} model files from a bootstrap control
stream file.}
\usage{
get_boot_spec(.boot_run)
}
\arguments{
\item{.boot_run}{Either a \code{bbi_nmboot_model} or \code{bbi_nmboot_summary} object}
}
\description{
Tabulate all relevant \code{bbi_nonmem_model} model files from a bootstrap control
stream file.
}
\keyword{internal}
