% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_boot_spec_path}
\alias{get_boot_spec_path}
\title{Get the bootstrap specification file path}
\usage{
get_boot_spec_path(.mod, .check_exists = TRUE)
}
\arguments{
\item{.mod}{a \code{bbi_nmboot_model} model object.}

\item{.check_exists}{If \code{TRUE}, the default, will throw an error if the file
does not exist.}
}
\description{
Get the bootstrap specification file path
}
\note{
this is not the configuration file generated by bbi
}
\keyword{internal}
