% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-estimates.R
\name{get_matrix_opts}
\alias{get_matrix_opts}
\title{Get options for matrix-type records}
\usage{
get_matrix_opts(.mod)
}
\arguments{
\item{.mod}{a \code{bbr} model object}
}
\description{
Get options for matrix-type records
}
\details{
Returns a tibble indicating the record type and occurrence (\code{record_number}),
along with the matrix type (diagonal vs block), and value types for both
diagonal (variance vs standard deviation) and off-diagonal (covariance vs
correlation) options.
\subsection{Example:}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{> get_matrix_opts(.mod)
# A tibble: 3 × 5
record_type record_number mat_type diag     off_diag
<chr>       <chr>         <chr>    <chr>    <chr>
1 omega       1             diagonal variance covariance
2 omega       2             block    standard correlation
3 sigma       1             diagonal variance covariance
}\if{html}{\out{</div>}}
}
}
\keyword{internal}
