% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-estimates.R
\name{get_matrix_types}
\alias{get_matrix_types}
\title{Get matrix type for omega and sigma records}
\usage{
get_matrix_types(records)
}
\arguments{
\item{records}{Either a list of records, or a single \code{nmrec_record} object}
}
\description{
Get matrix type for omega and sigma records
}
\keyword{internal}
