% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mod_id}
\alias{get_mod_id}
\title{Helper to strip path and extension from model file to get only model identifier}
\usage{
get_mod_id(.mod_path)
}
\arguments{
\item{.mod_path}{model path to strip}
}
\value{
Character scaler with only model identifier
}
\description{
Helper to strip path and extension from model file to get only model identifier
}
