% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_model_id}
\alias{get_model_id}
\alias{get_model_id.character}
\alias{get_model_id.bbi_nonmem_model}
\alias{get_model_id.bbi_nonmem_summary}
\alias{get_model_id.bbi_stan_model}
\alias{get_model_id.bbi_stan_summary}
\title{Get model identifier}
\usage{
get_model_id(.mod)

\method{get_model_id}{character}(.mod)

\method{get_model_id}{bbi_nonmem_model}(.mod)

\method{get_model_id}{bbi_nonmem_summary}(.mod)

\method{get_model_id}{bbi_stan_model}(.mod)

\method{get_model_id}{bbi_stan_summary}(.mod)
}
\arguments{
\item{.mod}{Model to use, either a \verb{bbi_\{.model_type\}_model} or
\verb{bbi_\{.model_type\}_summary}, or a file path to a model.}
}
\value{
Character scalar with only model identifier
}
\description{
Helper to strip path and extension from model file to get only model identifier
(i.e. the model file name without extension)
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: Takes file path to model.

\item \code{bbi_nonmem_model}: Takes \code{bbi_nonmem_model} object

\item \code{bbi_nonmem_summary}: Takes \code{bbi_nonmem_summary} object

\item \code{bbi_stan_model}: Takes \code{bbi_nonmem_model} object

\item \code{bbi_stan_summary}: Takes \code{bbi_nonmem_summary} object
}}

