% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_model_id}
\alias{get_model_id}
\alias{get_model_id.character}
\alias{get_model_id.bbi_nonmem_model}
\title{Helper to strip path and extension from model file to get only model identifier}
\usage{
get_model_id(.mod)

\method{get_model_id}{character}(.mod)

\method{get_model_id}{bbi_nonmem_model}(.mod)
}
\arguments{
\item{.mod}{The `bbi_{.model_type}_model` object}
}
\value{
Character scaler with only model identifier
}
\description{
Helper to strip path and extension from model file to get only model identifier

S3 dispatch to get model identifier from a `bbi_{.model_type}_model` object
}
