% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-param.R
\name{get_omega}
\alias{get_omega}
\title{Return omega values as a matrix}
\usage{
get_omega(.summary)
}
\arguments{
\item{.summary}{A \code{bbi_nonmem_summary} or \code{bbi_summary_list} object.}
}
\description{
Return omega values as a matrix
}
