% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-param.R
\name{get_param}
\alias{get_param}
\title{Return a formatted object for a given parameter}
\usage{
get_param(.summary, .param = c("omega", "sigma", "theta"))
}
\arguments{
\item{.summary}{a \code{bbi_nonmem_summary} or \code{bbi_summary_list} object.}

\item{.param}{the parameter(s) to retrieve and format. Any subset of \code{c("omega", "sigma", "theta")}}
}
\description{
Return a formatted object for a given parameter
}
\keyword{internal}
