% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_path_from_object}
\alias{get_path_from_object}
\alias{get_model_path}
\alias{get_output_dir}
\title{Get path from bbi object}
\usage{
get_path_from_object(.bbi_object, .key)

get_model_path(.mod)

get_output_dir(.mod)
}
\arguments{
\item{.bbi_object}{The object to query}

\item{.key}{The key that contains the relative path}

\item{.mod}{The `bbi_...` S3 object}
}
\description{
Builds the full path to a file that is stored as part of a `bbi_...` S3 object
All paths saved in the object or accompanying YAML will be relative **to the location of that YAML**
When the object is loaded into memory, the absolute path to the YAML is stored in the object.
These functions simply stitch together that path with the requested relative path.
As long as the YAML has not moved since it was loaded, this will work.
}
