% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_path_from_object}
\alias{get_path_from_object}
\alias{get_model_path}
\alias{get_model_path.bbi_nonmem_model}
\alias{get_model_path.bbi_log_df}
\alias{get_output_dir}
\alias{get_output_dir.bbi_nonmem_model}
\alias{get_output_dir.bbi_log_df}
\alias{get_yaml_path}
\alias{get_yaml_path.bbi_nonmem_model}
\alias{get_yaml_path.bbi_log_df}
\title{Get path from bbi object}
\usage{
get_model_path(.bbi_object, .check_exists = TRUE)

\method{get_model_path}{bbi_nonmem_model}(.bbi_object, .check_exists = TRUE)

\method{get_model_path}{bbi_log_df}(.bbi_object, .check_exists = TRUE)

get_output_dir(.bbi_object, .check_exists = TRUE)

\method{get_output_dir}{bbi_nonmem_model}(.bbi_object, .check_exists = TRUE)

\method{get_output_dir}{bbi_log_df}(.bbi_object, .check_exists = TRUE)

get_yaml_path(.bbi_object, .check_exists = TRUE)

\method{get_yaml_path}{bbi_nonmem_model}(.bbi_object, .check_exists = TRUE)

\method{get_yaml_path}{bbi_log_df}(.bbi_object, .check_exists = TRUE)
}
\arguments{
\item{.bbi_object}{The object to query. Could be
a \verb{bbi_\{.model_type\}_model} object,
or a tibble of class \code{bbi_log_df}.}

\item{.check_exists}{If \code{TRUE}, the default, will throw an error if the file does not exist}
}
\description{
\verb{bbi_\{.model_type\}_model} objects correspond to multiple files on disk. These functions
provide an easy way to retrieve the absolute path to these files.
}
