% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_path_from_object}
\alias{get_path_from_object}
\alias{get_model_path}
\alias{get_model_path.bbi_nonmem_model}
\alias{get_model_path.bbi_nonmem_summary}
\alias{get_model_path.bbi_stan_model}
\alias{get_model_path.bbi_stan_summary}
\alias{get_model_path.bbi_log_df}
\alias{get_output_dir}
\alias{get_output_dir.bbi_nonmem_model}
\alias{get_output_dir.bbi_nonmem_summary}
\alias{get_output_dir.bbi_stan_model}
\alias{get_output_dir.bbi_stan_summary}
\alias{get_output_dir.bbi_log_df}
\alias{get_yaml_path}
\alias{get_yaml_path.bbi_model}
\alias{get_yaml_path.bbi_log_df}
\alias{get_config_path}
\alias{get_config_path.bbi_model}
\alias{get_config_path.bbi_log_df}
\title{Get path from bbi object}
\usage{
get_model_path(.bbi_object, .check_exists = TRUE)

\method{get_model_path}{bbi_nonmem_model}(.bbi_object, .check_exists = TRUE)

\method{get_model_path}{bbi_nonmem_summary}(.bbi_object, .check_exists = TRUE)

\method{get_model_path}{bbi_stan_model}(.bbi_object, .check_exists = TRUE)

\method{get_model_path}{bbi_stan_summary}(.bbi_object, .check_exists = TRUE)

\method{get_model_path}{bbi_log_df}(.bbi_object, .check_exists = TRUE)

get_output_dir(.bbi_object, .check_exists = TRUE)

\method{get_output_dir}{bbi_nonmem_model}(.bbi_object, .check_exists = TRUE)

\method{get_output_dir}{bbi_nonmem_summary}(.bbi_object, .check_exists = TRUE)

\method{get_output_dir}{bbi_stan_model}(.bbi_object, .check_exists = TRUE)

\method{get_output_dir}{bbi_stan_summary}(.bbi_object, .check_exists = TRUE)

\method{get_output_dir}{bbi_log_df}(.bbi_object, .check_exists = TRUE)

get_yaml_path(.bbi_object, .check_exists = TRUE)

\method{get_yaml_path}{bbi_model}(.bbi_object, .check_exists = TRUE)

\method{get_yaml_path}{bbi_log_df}(.bbi_object, .check_exists = TRUE)

get_config_path(.bbi_object, .check_exists = TRUE)

\method{get_config_path}{bbi_model}(.bbi_object, .check_exists = TRUE)

\method{get_config_path}{bbi_log_df}(.bbi_object, .check_exists = TRUE)
}
\arguments{
\item{.bbi_object}{The object to query. Could be
a \verb{bbi_\{.model_type\}_model} object,
\verb{bbi_\{.model_type\}_summary} object,
or a tibble of class \code{bbi_log_df}.}

\item{.check_exists}{If \code{TRUE}, the default, will throw an error if the file does not exist}
}
\description{
\verb{bbi_\{.model_type\}_model} objects correspond to multiple files on disk. These functions
provide an easy way to retrieve the absolute path to these files.
}
\details{
\strong{\code{get_model_path()}} returns the path to the model definition file.
For NONMEM models, this is the control stream. For Stan models,
this is the \code{.stan} file.

\strong{\code{get_output_dir()}} returns the path to the directory containing
output files created when the model is run.

\strong{\code{get_yaml_path()}} returns the path to the YAML file created by bbr.
This file contains metadata like tags, etc. and should, generally speaking,
\emph{not} be interacted with directly. Use the helper functions mentioned in the
\code{\link[=modify_model_field]{modify_model_field()}} help page to modify this file for you.

Note that \code{\link[=build_path_from_model]{build_path_from_model()}} and \code{\link[=get_data_path]{get_data_path()}} provide similar functionality, but have a slightly
different interface and implementation and so are documented separately.
}
\seealso{
\code{\link[=get_model_id]{get_model_id()}} \code{\link[=get_data_path]{get_data_path()}} \code{\link[=build_path_from_model]{build_path_from_model()}}
}
