% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{get_spec_path}
\alias{get_spec_path}
\alias{get_spec_path.bbi_nmboot_model}
\alias{get_spec_path.bbi_base_model}
\title{Get the relevant specification file path}
\usage{
get_spec_path(.mod, .check_exists = TRUE)

\method{get_spec_path}{bbi_nmboot_model}(.mod, .check_exists = TRUE)

\method{get_spec_path}{bbi_base_model}(.mod, .check_exists = TRUE)
}
\arguments{
\item{.mod}{a \verb{bbi_\{.model_type\}_model} object.}

\item{.check_exists}{If \code{TRUE}, the default, will throw an error if the file
does not exist.}
}
\value{
a file path
}
\description{
Get the bootstrap specification path from a \code{bbi_nmboot_model} object or the
simulation specification path from a \code{bbi_nmsim_model} \emph{or} \code{bbi_nonmem_model}
object.
}
\section{Methods (by class)}{
\itemize{
\item \code{get_spec_path(bbi_nmboot_model)}: Get the bootstrap specification file path from a
\code{bbi_nmboot_model} object

\item \code{get_spec_path(bbi_base_model)}: Get the simulation specification file path from a
\code{bbi_nmsim_model} or \code{bbi_nonmem_model} object

}}
\note{
this is not the configuration file generated by bbi
}
\keyword{internal}
