% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-initial-est.R
\name{get_theta_inits}
\alias{get_theta_inits}
\title{Helper function for getting the initial theta values, including bounds and
the presence of \code{FIXED} parameters.}
\usage{
get_theta_inits(ctl, mark_flags)
}
\arguments{
\item{ctl}{an \code{nmrec} ctl object.}

\item{mark_flags}{TODO}
}
\description{
Helper function for getting the initial theta values, including bounds and
the presence of \code{FIXED} parameters.
}
\keyword{internal}
