% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_yaml_path}
\alias{get_yaml_path}
\alias{get_yaml_path.bbi_nonmem_model}
\alias{get_yaml_path.character}
\title{S3 generic to return the path to the YAML file}
\usage{
get_yaml_path(.path, ...)

\method{get_yaml_path}{bbi_nonmem_model}(.mod, .check_exists = TRUE)

\method{get_yaml_path}{character}(.path, .check_exists = TRUE)
}
\arguments{
\item{.path}{The file path to convert to a YAML}

\item{.mod}{The `bbi_nonmem_model` S3 object}

\item{.check_exists}{Boolean for whether it will check if the file exists and error if it does not. True by default.}
}
\description{
S3 generic to return the path to the YAML file

S3 dispatch to return the path to the YAML file from a `bbi_nonmem_model` object

S3 dispatch to return the path to the YAML file from a model path or output directory
}
