% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model-stan-helpers.R
\name{import_stan_init}
\alias{import_stan_init}
\title{Pull in Stan init file}
\usage{
import_stan_init(.mod, .standata)
}
\arguments{
\item{.mod}{a \code{bbi_stan_model} object}

\item{.standata}{the data list that is returned from \code{\link[=build_data]{build_data()}} for this model}
}
\description{
Sources the \code{-init.R} file associated with a model
and returns whatever is returned by the the \code{make_init()}
function it contains.
}
\keyword{internal}
