% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{inherit_param_estimates}
\alias{inherit_param_estimates}
\title{Inherit parameter estimates}
\usage{
inherit_param_estimates(
  .mod,
  .mod_inherit_path = get_based_on(.mod),
  .inherit = c("theta", "sigma"),
  .bounds_opts = c("maintain_bounds", "single_value"),
  .digits = 3
)
}
\arguments{
\item{.mod}{new model object to overwrite.}

\item{.mod_inherit_path}{model path to inherit properties from.}

\item{.inherit}{type of estimates to inherit from parent model.}
}
\description{
Inherit parameter estimates
}
