% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{inherit_param_estimates}
\alias{inherit_param_estimates}
\title{Inherit parameter estimates}
\usage{
inherit_param_estimates(
  .mod,
  .parent_mod = get_based_on(.mod)[1],
  inherit = c("theta", "sigma", "omega"),
  bounds = c("keep", "discard"),
  digits = 3
)
}
\arguments{
\item{.mod}{model object to update.}

\item{.parent_mod}{Either a model object, or path to a model to inherit
properties from. Defaults to the "parent" model of \code{.mod} (i.e. the first
entry in \code{get_based_on(.mod)})}

\item{inherit}{type of estimates to inherit from parent model. Defaults to
replacing all of THETA, SIGMA, and OMEGA}

\item{bounds}{Whether to keep or discard the existing bounds when setting
the initial estimates in THETA records.}

\item{digits}{Number of significant digits to round estimates to.}
}
\description{
Set the initial parameter estimates of a model using the estimates of a
previously executed model. Often this is used to carry forward the final
estimates of a "parent" model to be used as the initial estimates of the
"child" model, for example a model created with \code{\link[=copy_model_from]{copy_model_from()}}.
}
\details{
\subsection{Constraints and limitations}{
\itemize{
\item These functions update initial estimates only if they are \strong{explicitly}
defined in the control stream.

For example, consider the update of \verb{$THETA (1)x4}. That defines four
initial estimates, but only the first explicitly appears. Calling
\code{inherit_param_estimates()} with theta estimates \code{c(5, 6, 7, 8)} would
return a result of \verb{(5)x4}.
\item Using additional parameter records for priors is not supported and will
lead to a size mismatch between the parameter and its records. Instead use
informative prior record names (such as \code{THETAP} and \code{THETAPV}).
}
}
}
\examples{

\dontrun{
base_mod <- read_model(file.path(MODEL_DIR, "1"))

mod2 <- copy_model_from(base_mod, "mod2") \%>\%
  inherit_param_estimates()

}
}
