% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{inherit_param_estimates}
\alias{inherit_param_estimates}
\title{Inherit parameter estimates}
\usage{
inherit_param_estimates(.mod)
}
\arguments{
\item{.mod}{}
}
\description{
Inherit parameter estimates
}
