% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-initial-est.R
\name{initial_estimates}
\alias{initial_estimates}
\title{Retrieve and format initial parameter estimates}
\usage{
initial_estimates(.mod, flag_fixed = FALSE)
}
\arguments{
\item{.mod}{a \code{bbr} model object}

\item{flag_fixed}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, return which
parameters are fixed.}
}
\description{
Retrieve and format initial parameter estimates
}
\details{
\code{NA} values indicate that they were not specified in the control stream file.
This is true for \code{THETA} bounds and whether a given parameter is \code{FIXED} or not.

If you would like to format \code{OMEGA} or \code{SIGMA} records as \emph{full matrices},
they are stored as attributes:

\if{html}{\out{<div class="sourceCode">}}\preformatted{ initial_est <- initial_estimates(.mod)
 attr(initial_est, "omega_mat")
 attr(initial_est, "sigma_mat")
}\if{html}{\out{</div>}}
}
\examples{
mod1 <- read_model(
  system.file("model", "nonmem", "basic", "1", package = "bbr")
)

initial_estimates(mod1)

}
