% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse-model.R
\name{initial_estimates_compare}
\alias{initial_estimates_compare}
\title{Compare SSE results to initial (or "true") estimates.}
\usage{
initial_estimates_compare(
  .sse_sum,
  .orig_mod = NULL,
  probs = c(0.5, 0.025, 0.975),
  na.rm = FALSE
)
}
\description{
Compare SSE results to initial (or "true") estimates.
}
