% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{install_menu}
\alias{install_menu}
\title{Private implementation function for installing bbi with interactive menu}
\usage{
install_menu(.path, .version, .force, .quiet)
}
\arguments{
\item{.path}{path to install bbi to. See Details section for defaults, if
nothing is passed. Note that this should be the path where you would like
the bbi executable to be installed, \emph{not} the path to the directory in
which you want to install it. For example, you should pass
\code{"/some/dir/bbi"} and \emph{not} \code{"/some/dir"}.}

\item{.version}{version of bbi to install. Must pass a character scalar
corresponding to a tag found in
\verb{https://github.com/metrumresearchgroup/bbi/releases}}

\item{.force}{If \code{FALSE}, the default, skips installation if requested
version and local version are the same. If \code{TRUE} forces installation if it
will be the same version.}

\item{.quiet}{\strong{Deprecated.} Use \code{options("bbr.verbose")} instead to control
printing. Defaults to \code{NULL}, which reads \code{!getOption("bbr.verbose")}. If
\code{TRUE}, suppresses output printed to the console.}
}
\description{
Private implementation function for installing bbi with interactive menu
}
\keyword{internal}
