% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-summaries.R
\name{model_summaries}
\alias{model_summaries}
\alias{model_summaries.list}
\alias{model_summaries.bbi_run_log_df}
\title{Summarize multiple models in batch}
\usage{
model_summaries(
  .mods,
  .bbi_args = NULL,
  .fail_flags = NULL,
  ...,
  .dry_run = FALSE
)

\method{model_summaries}{list}(
  .mods,
  .bbi_args = NULL,
  .fail_flags = NULL,
  ...,
  .dry_run = FALSE
)

\method{model_summaries}{bbi_run_log_df}(
  .mods,
  .bbi_args = NULL,
  .fail_flags = NULL,
  ...,
  .dry_run = FALSE
)
}
\arguments{
\item{.mods}{The model objects to summarize. Could be
a \code{bbi_run_log_df} tibble, or
a list of \verb{bbi_\{.model_type\}_model} objects.}

\item{.bbi_args}{A named list specifying arguments to pass to babylon formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}.
See \code{\link[=print_bbi_args]{print_bbi_args()}} for full list of options.}

\item{.fail_flags}{Same as \code{.bbi_args} except these are used \emph{only} when a \code{\link[=model_summary]{model_summary()}} call fails.
In that case, flags are appended to anything in \code{.bbi_args} and the summary is tried again.
See details section for more info on these flags.}

\item{...}{args passed through to \code{bbi_exec()}}

\item{.dry_run}{show what the command would be without actually running it}
}
\value{
All dispatches will return a list of \verb{bbi_\{.model_type\}_summary} objects.
}
\description{
Run multiple \code{\link[=model_summary]{model_summary()}} calls in batch.
}
\details{
The summary call will error if it does not find certain files in the output folder.
However, you can override this behavior with the following file-specific flags:
\itemize{
\item \code{no_ext_file}
\item \code{no_grd_file}
\item \code{no_shk_file}
}

If some of your runs are using an estimation method that does not produce any of the following files,
or they are missing for some other legitimate reason, pass the appropriate flags through the \code{.fail_flags} argument.

Additionally, \strong{if you have renamed your \code{.ext} file, you will need to pass the new name through}, to \code{.bbi_args} or \code{.fail_flags} like so:
\code{model_summaries(..., .bbi_args = list(ext_file = "whatever_you_named_it"))}
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Summarize a list of \verb{bbi_\{.model_type\}_model} objects.

\item \code{bbi_run_log_df}: Takes a \code{bbi_run_log_df} tibble and summarizes all models in it.
}}

