% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{model_summary}
\alias{model_summary}
\alias{model_summary.bbi_nonmem_model}
\alias{model_summary.character}
\alias{nonmem_summary}
\title{S3 generic for getting model summary}
\usage{
model_summary(.mod, ...)

\method{model_summary}{bbi_nonmem_model}(.mod, ...)

\method{model_summary}{character}(.path, .model_type = c("nonmem", "stan"), ...)

nonmem_summary(.mod, .args = NULL, ..., .dry_run = FALSE)
}
\arguments{
\item{.mod}{`bbi_nonmem_model` object for summary}

\item{...}{args passed through to `bbi_exec()`}

\item{.path}{Full path to one of the following:
1) path to a model file in the output directory
2) path to a nonmem output directory}

\item{.model_type}{Character scaler specifying the type of model, either 'nonmem' or 'stan'}

\item{.args}{A named list specifying arguments to pass to babylon formatted like `list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)`. Run `print_nonmem_args()` to see valid arguments.}

\item{.dry_run}{show what the command would be without actually running it}
}
\value{
List of S3 Class "bbi_nonmem_summary" with all summary information
}
\description{
S3 generic for getting model summary

Get model summary from `bbi_nonmem_model` object and optionally check if model is done before erroring

S3 dispatch for getting model summary from output directory path

Run `bbi nonmem summary` and parse the output to a list
}
