% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-model-results.R
\name{model_summary}
\alias{model_summary}
\title{Execute `bbi summary` to generate json output of model results}
\usage{
model_summary(.path, .model_type = c("nonmem"))
}
\arguments{
\item{.path}{Full path to a model that has been run that user wants to summarize}

\item{.model_type}{Type of model to summarize. Currently only supports "nonmem"}
}
\description{
Execute `bbi summary` to generate json output of model results
}
