% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-summary.R
\name{model_summary}
\alias{model_summary}
\alias{model_summary.bbi_nonmem_model}
\alias{model_summary.character}
\alias{model_summary.numeric}
\title{Summarize model outputs}
\usage{
model_summary(.mod, .bbi_args = NULL, ..., .dry_run = FALSE, .directory = NULL)

\method{model_summary}{bbi_nonmem_model}(.mod, .bbi_args = NULL, ..., .dry_run = FALSE, .directory = NULL)

\method{model_summary}{character}(
  .mod,
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = get_model_directory()
)

\method{model_summary}{numeric}(
  .mod,
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = get_model_directory()
)
}
\arguments{
\item{.mod}{Model to summarize. Can be a \verb{bbi_\{.model_type\}_model} object, a file path, or an integer corresponding to a file path.}

\item{.bbi_args}{A named list specifying arguments to pass to babylon formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}.
See \code{\link[=print_bbi_args]{print_bbi_args()}} for full list of options.}

\item{...}{args passed through to \code{bbi_exec()}}

\item{.dry_run}{show what the command would be without actually running it}

\item{.directory}{Model directory which \code{.mod} path is relative to. Defaults to \code{options('rbabylon.model_directory')}, which can be set globally with \code{set_model_directory()}. \strong{Only used when passing a path for \code{.mod} instead of a \verb{bbi_\{.model_type\}_model} object.}}
}
\description{
Calls out to bbi and returns a named list of class \verb{bbi_\{.model_type\}_summary} with model outputs and diagnostics.
}
\details{
\strong{NONMEM}

The returned list for a NONMEM model will contain the following top-level elements:
\itemize{
\item run_details
\item run_heuristics
\item parameters_data
\item parameter_names
\item ofv
\item shrinkage_details
\item covariance_theta
\item correlation_theta
}

The summary call will error if it does not find certain files in the output folder.
However, you can override this behavior with the following file-specific flags:
\itemize{
\item \code{no_cor_file}
\item \code{no_cov_file}
\item \code{no_ext_file}
\item \code{no_grd_file}
\item \code{no_shk_file}
}

If you are using an estimation method that does not produce any of the following files,
or they are missing for some other legitimate reason, pass the appropriate flags through the \code{.bbi_args} argument.
For example, if have asked to skip the \verb{$COV} step, you would call \code{model_summary(..., .bbi_args = list(no_cov_file = TRUE))}.
}
\section{Methods (by class)}{
\itemize{
\item \code{bbi_nonmem_model}: Get model summary from \code{bbi_nonmem_model} object

\item \code{character}: Get model summary from output directory path

\item \code{numeric}: Get model summary from numeric input.
This will only work if you are calling from the same directory as the models, or if you have set \code{options('rbabylon.model_directory')} to the directory constaining the relevant model.
}}

