% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{model_summary}
\alias{model_summary}
\alias{model_summary.bbi_nonmem_model}
\alias{model_summary.character}
\alias{model_summary.numeric}
\title{Summarize model outputs}
\usage{
model_summary(
  .mod,
  .model_type = NULL,
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = NULL
)

\method{model_summary}{bbi_nonmem_model}(
  .mod,
  .model_type = NULL,
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = NULL
)

\method{model_summary}{character}(
  .mod,
  .model_type = c("nonmem", "stan"),
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = get_model_directory()
)

\method{model_summary}{numeric}(
  .mod,
  .model_type = c("nonmem", "stan"),
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = get_model_directory()
)
}
\arguments{
\item{.mod}{Integer that corresponds to the name of a YAML, model file, and output directory}

\item{.model_type}{Character scaler specifying the type of model, either 'nonmem' or 'stan'}

\item{.bbi_args}{A named list specifying arguments to pass to babylon formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}.}

\item{...}{args passed through to \code{bbi_exec()}}

\item{.dry_run}{show what the command would be without actually running it}

\item{.directory}{Model directory containing the files referenced by \code{.mod}. Defaults to \code{options('rbabylon.model_directory')}, which can be set globally with \code{set_model_directory()}.}
}
\description{
Returns a named list with model outputs and diagnostics.
}
