% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-summary.R
\name{model_summary}
\alias{model_summary}
\alias{model_summary.bbi_nonmem_model}
\title{Summarize model outputs}
\usage{
model_summary(.mod, .bbi_args = NULL, ..., .dry_run = FALSE)

\method{model_summary}{bbi_nonmem_model}(.mod, .bbi_args = NULL, ..., .dry_run = FALSE)
}
\arguments{
\item{.mod}{Model to summarize.}

\item{.bbi_args}{A named list specifying arguments to pass to babylon formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}.
See \code{\link[=print_bbi_args]{print_bbi_args()}} for full list of options.}

\item{...}{args passed through to \code{bbi_exec()}}

\item{.dry_run}{show what the command would be without actually running it}
}
\description{
Calls out to bbi and returns a named list of class \verb{bbi_\{.model_type\}_summary} with model outputs and diagnostics.
}
\details{
\strong{NONMEM}

The returned list for a NONMEM model will contain the following top-level elements:
\itemize{
\item run_details
\item run_heuristics
\item parameters_data
\item parameter_names
\item ofv
\item condition_number
\item shrinkage_details
}

The summary call will error if it does not find certain files in the output folder.
However, you can override this behavior with the following file-specific flags:
\itemize{
\item \code{no_ext_file}
\item \code{no_grd_file}
\item \code{no_shk_file}
}

If you are using an estimation method that does not produce any of the following files,
or they are missing for some other legitimate reason, pass the appropriate flags through the \code{.bbi_args} argument.
For example, if have asked to skip the \verb{$COV} step, you would call \code{model_summary(..., .bbi_args = list(no_cov_file = TRUE))}.

Additionally, if you have renamed the \code{.ext} file from its default of \verb{<root>.ext} you will need to pass
\code{ext_file = "NEWNAME"} to \code{.bbi_args}.
}
\section{Methods (by class)}{
\itemize{
\item \code{bbi_nonmem_model}: Get model summary from \code{bbi_nonmem_model} object
}}

