% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-summary.R
\name{model_summary}
\alias{model_summary}
\alias{model_summary.bbi_nonmem_model}
\title{Summarize model outputs}
\usage{
model_summary(.mod, .bbi_args = NULL, ..., .dry_run = FALSE)

\method{model_summary}{bbi_nonmem_model}(.mod, .bbi_args = NULL, ..., .dry_run = FALSE)
}
\arguments{
\item{.mod}{Model to summarize.}

\item{.bbi_args}{A named list specifying arguments to pass to bbi formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}.
See \code{\link[=print_bbi_args]{print_bbi_args()}} for full list of options.}

\item{...}{args passed through to \code{\link[=bbi_exec]{bbi_exec()}}}

\item{.dry_run}{show what the command would be without actually running it}
}
\description{
Calls out to bbi and returns a named list of class \verb{bbi_\{.model_type\}_summary} with model outputs and diagnostics.
}
\details{
\strong{NONMEM}

The returned list for a NONMEM model will contain the following top-level elements:
\itemize{
\item \strong{absolute_model_path} -- Absolute path to the model that generated this summary.
Note, using this directly is discouraged in favor of using the "getters"
described in the \code{\link{get_path_from_object}} help page.
\item \strong{run_details} -- General details about the run including estimation
method, numbers of patients and records, significant digits, run time, and
more.
\item \strong{run_heuristics} -- Boolean flags indicating potential issues with the
run. For example, if there was a large condition number or a parameter near
boundary. \code{TRUE} for a given element indicates that issue was detected.
\code{FALSE} can either mean it was not detected or it was not applicable for
that run.
\item \strong{parameters_data} -- Data about parameter estimates. This can be
accessed directly, but is much easier to look at with the
\code{\link[=param_estimates]{param_estimates()}} function (which parses this, and other related elements,
into a tibble).
\item \strong{parameter_names} -- Names of parameters (\code{THETA1}, \code{THETA2}, etc.).
Parsed into tibble by \code{\link[=param_estimates]{param_estimates()}}.
\item \strong{ofv} -- Objective function value, with and without constant. List with
one element for each estimation method.
\item \strong{condition_number} -- Condition number, if eigenvalues were returned
from NONMEM. List with one element for each estimation method.
\item \strong{shrinkage_details} -- Information about shrinkage. The shrinkage values
(using SD parameterization) are parsed into the tibble output from
\code{\link[=param_estimates]{param_estimates()}}. List of lists with one element for each estimation
method, and one element per sub-population within that.
}

The summary call will error if it does not find certain files in the output folder.
However, you can override this behavior with the following file-specific flags:
\itemize{
\item \code{no_ext_file}
\item \code{no_grd_file}
\item \code{no_shk_file}
}

If you are using an estimation method that does not produce any of the following files,
or they are missing for some other legitimate reason, pass the appropriate flags through the \code{.bbi_args} argument.
For example, if have asked to skip the \verb{$COV} step, you would call \code{model_summary(..., .bbi_args = list(no_cov_file = TRUE))}.

Additionally, if you have renamed the \code{.ext} file from its default of \verb{<root>.ext} you will need to pass
\code{ext_file = "NEWNAME"} to \code{.bbi_args}.
}
\section{Methods (by class)}{
\itemize{
\item \code{bbi_nonmem_model}: Get model summary from \code{bbi_nonmem_model} object
}}

\seealso{
\code{\link[=model_summaries]{model_summaries()}}
}
