% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{model_summary}
\alias{model_summary}
\alias{model_summary.bbi_nonmem_model}
\alias{model_summary.character}
\alias{model_summary.numeric}
\alias{nonmem_summary}
\title{S3 generic for getting model summary}
\usage{
model_summary(
  .mod,
  .model_type = NULL,
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = NULL
)

\method{model_summary}{bbi_nonmem_model}(
  .mod,
  .model_type = NULL,
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = NULL
)

\method{model_summary}{character}(
  .mod,
  .model_type = c("nonmem", "stan"),
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = get_model_directory()
)

\method{model_summary}{numeric}(
  .mod,
  .model_type = c("nonmem", "stan"),
  .bbi_args = NULL,
  ...,
  .dry_run = FALSE,
  .directory = get_model_directory()
)

nonmem_summary(.mod, .bbi_args = NULL, ..., .dry_run = FALSE)
}
\arguments{
\item{.mod}{`bbi_nonmem_model` object for summary}

\item{.model_type}{Character scaler specifying the type of model, either 'nonmem' or 'stan'}

\item{.bbi_args}{A named list specifying arguments to pass to babylon formatted like `list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)`. Run `print_nonmem_args()` to see valid arguments.}

\item{...}{args passed through to `bbi_exec()`}

\item{.dry_run}{show what the command would be without actually running it}

\item{.directory}{Model directory containing the files referenced by `.mod`. Defaults to `options('rbabylon.model_directory')`, which can be set globally with `set_model_directory()`.}
}
\value{
List of S3 Class "bbi_nonmem_summary" with all summary information
}
\description{
S3 generic for getting model summary

Get model summary from `bbi_nonmem_model` object and optionally check if model is done before erroring

S3 dispatch for getting model summary from output directory path

S3 dispatch for getting model summary from numeric input
This will only work if you are calling from the same directory as the models, or if you have set the model directory with `set_model_directory()`

Run `bbi nonmem summary` and parse the output to a list
}
