% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{model_tree}
\alias{model_tree}
\title{Create a tree diagram of a modeling directory}
\usage{
model_tree(
  .base_dir,
  .recurse = FALSE,
  .include = NULL,
  .add_summary = TRUE,
  .static_plot = FALSE
)
}
\arguments{
\item{.base_dir}{Base directory to look in for models.}

\item{.recurse}{If \code{FALSE}, the default, only include models in \code{.base_dir}
but \emph{not} subdirectories. This is passed through to \code{fs::dir_ls()} --
If \code{TRUE} recurse fully, if a positive number the number of levels to recurse.}

\item{.include}{A character vector specifying which runs or model tags to include in the run log.}
}
\description{
Create a tree diagram of a modeling directory
}
