% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{model_tree_png}
\alias{model_tree_png}
\title{Save an interactive \code{\link[=model_tree]{model_tree()}} as a PNG and load in Rstudio viewer}
\usage{
model_tree_png(widget)
}
\arguments{
\item{widget}{an \code{HTML} widget}
}
\description{
Save an interactive \code{\link[=model_tree]{model_tree()}} as a PNG and load in Rstudio viewer
}
\details{
This function first saves and renders widget as HTML. It then screenshots
the webpage and saves the file to a temporary PNG. The PNG is loaded in as
an array and all data is re-packaged to be print like a \code{ggplot} object.
}
\keyword{internal}
