% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{modify_model_field}
\alias{modify_model_field}
\alias{add_tags}
\alias{replace_tags}
\alias{add_decisions}
\alias{replace_decisions}
\alias{add_based_on}
\alias{replace_based_on}
\alias{replace_description}
\title{Modify field in model object}
\usage{
modify_model_field(.mod, .field, .value, .append = TRUE, .unique = TRUE)

add_tags(.mod, .tags)

replace_tags(.mod, .tags)

add_decisions(.mod, .decisions)

replace_decisions(.mod, .decisions)

add_based_on(.mod, .based_on)

replace_based_on(.mod, .based_on)

replace_description(.mod, .description)
}
\arguments{
\item{.mod}{The `bbi_{.model_type}_model` object to modify}

\item{.field}{Character scaler of the name of the component to modify}

\item{.value}{Whatever is to be added to `.mod[[.field]]`, typically a character scaler or vector}

\item{.append}{Boolean for whether to concatenate new values with currently present values. TRUE by default. If FALSE, new values will overwrite old values.}

\item{.unique}{Boolean for whether to de-duplicate `.mod[[.field]]` after adding new values. TRUE by default.}

\item{.tags}{Character scaler or vector of tags use as replacement}

\item{.decisions}{Character scaler or vector to use as replacement}

\item{.based_on}{Character scaler or vector to use as replacement}

\item{.description}{Character scaler to use as replacement}
}
\description{
Implementation function for updating fields in a `bbi_{.model_type}_model` object
Also checks the object against the corresponding YAML before modifying (and errors if they are out of sync) and then writes the modified object back to the YAML.
}
