% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{modify_tags}
\alias{modify_tags}
\alias{add_tags}
\alias{replace_tag}
\alias{replace_all_tags}
\alias{remove_tags}
\title{Modify tags on a model object}
\usage{
add_tags(.mod, .tags)

replace_tag(.mod, .old_tag, .new_tag)

replace_all_tags(.mod, .tags)

remove_tags(.mod, .tags)
}
\arguments{
\item{.mod}{The \verb{bbi_\{.model_type\}_model} object to modify}

\item{.tags}{Character vector of tags to add or remove}

\item{.old_tag}{Character scalar of tag to be replaced}

\item{.new_tag}{Character scalar of tag that will be added}
}
\value{
The modified \verb{bbi_\{.model_type\}_model} object
}
\description{
Add, replace, or remove tags on a model object and corresponding YAML.
Tags can be modified at any time (i.e. before \emph{or} after a model is submitted).
}
\details{
The \code{tags} fields on a \verb{bbi_\{.model_type\}_model} object contains a character
vector of brief descriptors about the model. Often this is used to keep track
of the model structure, such as the covariates or error stucture that was
used.

One of the more useful things to do with tags is to use them for filtering a
\code{bbi_run_log_df} (the tibble output from \code{\link[=run_log]{run_log()}}); for example to look at
all models that used a particular random effect or covariate. To facilitate
this, it is recommended to keep individual tags consistent and short, and/or
to use a "glossary" of acceptable tags, stored elsewhere in your project
(i.e. in a \code{.yaml} or \code{.csv} or a named list that you can source). In other
words, if you tag one model with \code{"CLAGE"} and another with
\code{"ClearanceAgeCov"}, etc. then it will be very difficult to make any use of
those tags later, for instance when you want to filter to models with that
covariate.

Tags can also be collapsed using \code{\link[=collapse_to_string]{collapse_to_string()}} to create a compact description of the
model stucture. See the \href{https://metrumresearchgroup.github.io/bbr/articles/getting-started.html#viewing-tags-example}{"Getting Started" vignette}
for an example of this.
}
\section{Functions}{
\itemize{
\item \code{add_tags}: Add tags to a model object and corresponding YAML.

\item \code{replace_tag}: Replaces a specific \code{.old_tag} with \code{.new_tag} on a model object and corresponding YAML.
Warns and does nothing if \code{.old_tag} is not present.

\item \code{replace_all_tags}: Replaces all tags on a model object and corresponding YAML with new tags.

\item \code{remove_tags}: Removes tags from a model object and corresponding YAML.
}}

\seealso{
\code{\link[=run_log]{run_log()}} \code{\link[=collapse_to_string]{collapse_to_string()}} \code{\link[=modify_notes]{modify_notes()}} \code{\link[=modify_based_on]{modify_based_on()}} \code{\link[=modify_description]{modify_description()}} \code{\link[=modify_bbi_args]{modify_bbi_args()}}
}
