% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{new_bootstrap_run}
\alias{new_bootstrap_run}
\title{Create a boostrap model object from an existing model}
\usage{
new_bootstrap_run(
  .mod,
  .suffix = "boot",
  .inherit_tags = TRUE,
  .overwrite = FALSE,
  remove_cov = TRUE,
  remove_tables = TRUE
)
}
\arguments{
\item{.mod}{a \code{bbr} model object}

\item{.suffix}{a suffix for the bootstrap run directory. Will be prefixed by
the model id of \code{.mod}.}

\item{.inherit_tags}{If \code{FALSE}, the default, new model will only have any
tags passed to \code{.add_tags} argument. If \code{TRUE} inherit any tags from
\code{.parent_mod}, with any tags passed to \code{.add_tags} appended.}

\item{.overwrite}{If \code{FALSE}, the default,  function will error if a model
file already exists at specified \code{.new_model} path. If \code{TRUE} any existing
file at \code{.new_model} will be overwritten silently.}

\item{remove_cov, remove_tables}{Logical (T/F). Optionally remove \verb{$COVARIANCE}
and \verb{$TABLE} records respectively, allowing for notably faster run times.}
}
\value{
S3 object of class \code{bbi_nmboot_model}.
}
\description{
Create a boostrap model object from an existing model
}
\seealso{
setup_bootstrap_run summarize_bootstrap_run
}
