% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{new_bootstrap_run}
\alias{new_bootstrap_run}
\title{Create a boostrap model object from an existing model}
\usage{
new_bootstrap_run(
  .mod,
  .prefix = glue("{get_model_id(.mod)}_boot"),
  .inherit_tags = TRUE
)
}
\arguments{
\item{.mod}{a \code{bbr} model object}

\item{.prefix}{a prefix for the boostrap model directory. Will be appended by
the boostrap run number for a given model.}

\item{.inherit_tags}{If \code{FALSE}, the default, new model will only have any
tags passed to \code{.add_tags} argument. If \code{TRUE} inherit any tags from
\code{.parent_mod}, with any tags passed to \code{.add_tags} appended.}
}
\description{
Create a boostrap model object from an existing model
}
