% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{new_bootstrap_run}
\alias{new_bootstrap_run}
\title{Create a boostrap model object from an existing model}
\usage{
new_bootstrap_run(
  .mod,
  .prefix = glue("{get_model_id(.mod)}_boot"),
  .inherit_tags = TRUE
)
}
\arguments{
\item{.mod}{a \code{bbr} model object}

\item{.prefix}{a prefix for the boostrap model directory. Will be appended by
the boostrap run number for a given model.}

\item{.overwrite}{Logical to specify whether or not to overwrite existing
model output from a previous run.}
}
\description{
Create a boostrap model object from an existing model
}
