% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{new_bootstrap_run}
\alias{new_bootstrap_run}
\title{Create a bootstrap run from an existing model}
\usage{
new_bootstrap_run(
  .mod,
  .suffix = "boot",
  .inherit_tags = TRUE,
  .overwrite = FALSE,
  remove_cov = TRUE,
  remove_tables = TRUE
)
}
\arguments{
\item{.mod}{A \code{bbr} model object.}

\item{.suffix}{A suffix for the bootstrap run directory. Will be prefixed by
the model id of \code{.mod}.}

\item{.inherit_tags}{If \code{TRUE}, the default, inherit any tags from \code{.mod}.}

\item{.overwrite}{If \code{FALSE}, the default,  function will error if a model
file already exists at specified \code{.new_model} path. If \code{TRUE} any existing
file at \code{.new_model} will be overwritten silently.}

\item{remove_cov, remove_tables}{If \code{TRUE}, the default, remove \verb{$COVARIANCE}
and \verb{$TABLE} records respectively, allowing for notably faster run times.}
}
\value{
S3 object of class \code{bbi_nmboot_model}.
}
\description{
Creates a new \code{bbi_nmboot_model} object, from an existing \code{bbi_nonmem_model}
object. This function creates a new control stream, that is a copy of \code{.mod}
with the \verb{$TABLE} and \verb{$COV} records optionally removed (see \code{remove_cov} and
\code{remove_tables} arguments). The object returned from this must then be passed
to \code{\link[=setup_bootstrap_run]{setup_bootstrap_run()}} before submission (see examples).
}
\examples{
\dontrun{

# Create new bootstrap object
.boot_run <- new_bootstrap_run(.mod)

# Optionally inherit final parameter estimates
.boot_run <- inherit_param_estimates(.boot_run)

# Set up the run
setup_bootstrap_run(.boot_run)
}
}
\seealso{
\code{\link[=setup_bootstrap_run]{setup_bootstrap_run()}} \code{\link[=summarize_bootstrap_run]{summarize_bootstrap_run()}}
}
