% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-path-from-object.R
\name{new_ext}
\alias{new_ext}
\alias{ctl_ext}
\alias{mod_ext}
\alias{yaml_ext}
\alias{yml_ext}
\title{helpers that strip extension then add a diffent type}
\usage{
new_ext(.x, .ext)

ctl_ext(.x)

mod_ext(.x)

yaml_ext(.x)

yml_ext(.x)
}
\arguments{
\item{.x}{file path to modify}

\item{.ext}{new file extension to apply}
}
\description{
helpers that strip extension then add a diffent type
}
