% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{new_sim_model}
\alias{new_sim_model}
\title{Create a simulation model object from an existing model}
\usage{
new_sim_model(
  .mod,
  n = 100,
  seed = 1234,
  replicate_var = "REPI",
  .suffix = "sim",
  .inherit_tags = TRUE,
  .overwrite = FALSE
)
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} object.}

\item{n}{number of replicates.}

\item{seed}{a seed for simulation.}

\item{replicate_var}{a variable name for storing each iteration. Cannot match
a variable previously defined in \verb{$PRED} or \verb{$ERROR}.}

\item{.suffix}{a suffix for the simulation directory. Will be prefixed by
the model id of \code{.mod}.}

\item{.inherit_tags}{If \code{FALSE}, the default, new model will only have any
tags passed to \code{.add_tags} argument. If \code{TRUE} inherit any tags from
\code{.parent_mod}, with any tags passed to \code{.add_tags} appended.}

\item{.overwrite}{If \code{FALSE}, the default,  function will error if a model
file already exists at specified \code{.new_model} path. If \code{TRUE} any existing
file at \code{.new_model} will be overwritten silently.}
}
\value{
S3 object of class \code{bbi_nmsim_model}.
}
\description{
Create a simulation model object from an existing model
}
