% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{new_sim_model}
\alias{new_sim_model}
\title{Create a simulation model object from an existing model}
\usage{
new_sim_model(
  .mod,
  n = 200,
  seed = 1234,
  .join_col = "NUM",
  .suffix = "sim",
  .inherit_tags = TRUE,
  .overwrite = FALSE
)
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} object.}

\item{n}{number of replicates/subproblems. Adds \code{SUBPROBLEMS=n} to a
\verb{$SIMULATION} record.}

\item{seed}{a seed for simulation. Appended to \verb{$SIMULATION} record.}

\item{.join_col}{Character column name(s) used to join table files post
execution. Gets appended to the generated \verb{$TABLE} record. See
\code{\link[=nm_join_sim]{nm_join_sim()}} documentation for details. Defaults to \code{'NUM'}.}

\item{.suffix}{a suffix for the simulation directory. Will be prefixed by
the model id of \code{.mod}.}

\item{.inherit_tags}{If \code{TRUE}, the default, inherit any tags from \code{.mod}.}

\item{.overwrite}{If \code{FALSE}, the default,  function will error if a model
file already exists at specified \code{.new_model} path. If \code{TRUE} any existing
file at \code{.new_model} will be overwritten silently.}
}
\value{
S3 object of class \code{bbi_nmsim_model}.
}
\description{
Create a simulation model object from an existing model
}
\details{
\code{new_sim_model} does the following things:
\itemize{
\item Checks that \code{.mod} was previously executed and saved out an \code{.MSF} file was
generated (i.e. \verb{$EST MSFO=1.MSF}).
\item Performs various checks to confirm the status of \code{.mod} and contents of its
control stream.
\item Creates a new \code{bbi_nmsim_model} object with the following differences from
the original control stream:
\itemize{
\item Removes the following record types for simulation: \verb{$EST}, \verb{$COV},
\verb{$TABLE}, \verb{$SIMULATION}
\item Removes PK and prior records: \verb{$PRIOR}, \verb{$THETA/$THETAP/$THETAPV},
\verb{$OMEGA/$OMEGAP/$OMEGAPD}, \verb{$SIGMA/$SIGMAP/$SIGMAPD}
\item Adds a new custom \verb{$SIMULATION} record using user specified values
\item Adds a new \verb{$TABLE} record tabling out simulated values and \code{.join_col}
column(s)
\item Adds a new \verb{$MSFI} record (run with \code{NOMSFTEST}) pointing to the \code{.MSF}
file of \code{.mod}
}
}
}
\examples{
\dontrun{

# Create new `nmsim` model object
.sim_mod <- new_sim_model(.mod, n = 500, .join_col = c("NUM", "ID"))

# Visualize changes
model_diff(.sim_mod)

# Submit
submit_model(.sim_mod)
}

}
\seealso{
nm_join_sim
}
