% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-file.R
\name{nm_file}
\alias{nm_file}
\alias{nm_grd}
\alias{nm_ext}
\alias{nm_tab}
\alias{nm_par_tab}
\alias{nm_data}
\title{Read NONMEM files}
\usage{
nm_file(.mod, .suffix = NULL, .est_method = NULL, ...)

nm_grd(.mod, .est_method = NULL, .rename = TRUE)

nm_ext(.mod, .est_method = NULL)

nm_tab(.mod)

nm_par_tab(.mod)

nm_data(.mod)
}
\arguments{
\item{.mod}{Either a \code{bbi_nonmem_model}, \code{bbi_nonmem_summary}, or a path to a
file to read in. If passing model object to \code{nm_file()}, must also pass \code{.suffix} that
will be passed through to \code{\link[=build_path_from_model]{build_path_from_model()}}.}

\item{.suffix}{Character vector to append the end of the absolute model path.
Will be appended \emph{as is} so, if passing a file extension, be sure to included the leading \code{"."}.
See examples.}

\item{.est_method}{If \code{NULL}, the default, pulls the data from the final
estimation method. If an integer, pulls the data from that estimation
method. Can also pass \code{"fail"} in which case a warning will be raised
and \code{NULL} returned if more than one table is found in the file. This
is useful for table outputs that should not have multiple tables in a
single file.}

\item{...}{arguments passed through to methods. (Currently none.)}

\item{.rename}{If \code{TRUE}, the default, will rename \code{.grd} columns to the
relevant parameter names. Otherwise will leave column names as is.}
}
\value{
A tibble with the data from the specified file and estimation method.
All column names will be converted to uppercase.
}
\description{
Reads in a whitespace-delimited NONMEM output file (for example .grd or .ext
or a table output) or a NONMEM input data file. Will print the number of rows
and columns when the file is loaded. This \strong{printing can be suppressed} by
setting \code{options(bbr.verbose = FALSE)}.
}
\details{
\code{nm_file()} is called internally by the family of functions in this help doc,
as well as by \code{\link[=nm_tables]{nm_tables()}} and \code{\link[=nm_join]{nm_join()}}.

\code{nm_file()} looks for \verb{TABLE NO} to find the beginning of the requested
table. For this reason, \code{nm_file()} (and family) are \emph{not} compatible with
the \code{NOHEADER} or \code{NOTITLE} options in \verb{$TABLE}.
}
\section{Functions}{
\itemize{
\item \code{nm_grd}: Reads \code{.grd} file from a \code{bbi_nonmem_model} or
\code{bbi_nonmem_summary} object

\item \code{nm_ext}: Reads \code{.ext} file from a \code{bbi_nonmem_model} or
\code{bbi_nonmem_summary} object

\item \code{nm_tab}: Reads \verb{\{get_model_id(.mod)\}.tab} file from a
\code{bbi_nonmem_model} or \code{bbi_nonmem_summary} object

\item \code{nm_par_tab}: Reads \verb{\{get_model_id(.mod)\}par.tab} file from a
\code{bbi_nonmem_model} or \code{bbi_nonmem_summary} object

\item \code{nm_data}: Reads the input data file from a \code{bbi_nonmem_model} or
\code{bbi_nonmem_summary} object
}}

\seealso{
\code{\link[=nm_tables]{nm_tables()}}, \code{\link[=nm_table_files]{nm_table_files()}}, \code{\link[=nm_join]{nm_join()}}
}
