% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-file.R
\name{nm_file}
\alias{nm_file}
\alias{nm_grd}
\alias{nm_tab}
\alias{nm_par_tab}
\alias{nm_data}
\title{Read NONMEM files}
\usage{
nm_file(.mod, .suffix = NULL, ...)

nm_grd(.mod, .rename = TRUE)

nm_tab(.mod)

nm_par_tab(.mod)

nm_data(.mod)
}
\arguments{
\item{.mod}{Either a \code{bbi_nonmem_model}, \code{bbi_nonmem_summary}, or a path to a
file to read in. If passing model object to \code{nm_file()}, must also pass \code{.suffix} that
will be passed through to \code{\link[=build_path_from_model]{build_path_from_model()}}.}

\item{.suffix}{Character vector to append the end of the absolute model path.
Will be appended \emph{as is} so, if passing a file extension, be sure to included the leading \code{"."}.
See examples.}

\item{...}{arguments passed through to methods. (Currently none.)}

\item{.rename}{If \code{TRUE}, the default, will rename \code{.grd} columns to the
relevant parameter names. Otherwise will leave column names as is.}
}
\value{
A tibble with the data from the specified file and estimation method.
}
\description{
Reads in a whitespace-delimited NONMEM output file (for example .grd or .ext
or a table output) or a NONMEM input data file. Will print the number of rows
and columns when the file is loaded. This \strong{printing can be suppressed} by
setting \code{options(bbr.verbose = FALSE)}.
}
\details{
\code{nm_file()} is called internally by the family of functions in this help doc,
as well as by \code{\link[=nm_tables]{nm_tables()}} and \code{\link[=nm_join]{nm_join()}}.

\code{nm_file()} assumes there is only one table per file and therefore
\code{nm_file()} (and family) are \emph{not} compatible with files that have multiple
tables, for example an \code{.ext} file for a model with multiple estimation
methods or a table file from a model using \verb{$SIM}. For these kinds of files,
consider using \code{data.table::fread()} with the \code{skip} and \code{nrows} arguments.
}
\section{Functions}{
\itemize{
\item \code{nm_grd()}: Reads \code{.grd} file from a \code{bbi_nonmem_model} or
\code{bbi_nonmem_summary} object

\item \code{nm_tab()}: Reads \verb{\{get_model_id(.mod)\}.tab} file from a
\code{bbi_nonmem_model} or \code{bbi_nonmem_summary} object

\item \code{nm_par_tab()}: Reads \verb{\{get_model_id(.mod)\}par.tab} file from a
\code{bbi_nonmem_model} or \code{bbi_nonmem_summary} object

\item \code{nm_data()}: Reads the input data file from a \code{bbi_nonmem_model} or
\code{bbi_nonmem_summary} object

}}
\seealso{
\code{\link[=nm_tables]{nm_tables()}}, \code{\link[=nm_table_files]{nm_table_files()}}, \code{\link[=nm_join]{nm_join()}}
}
