% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-file.R
\name{nm_file_impl}
\alias{nm_file_impl}
\title{Implementation function for reading NONMEM files}
\usage{
nm_file_impl(.path, .est_method)
}
\description{
Implementation function for reading NONMEM files
}
\keyword{internal}
