% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-join.R
\name{nm_join}
\alias{nm_join}
\title{Return a single data frame with model output and input data}
\usage{
nm_join(
  .mod,
  .join_col = "NUM",
  .files = nm_table_files(.mod),
  .superset = FALSE,
  .bbi_args = list(no_grd_file = TRUE, no_shk_file = TRUE)
)
}
\arguments{
\item{.mod}{A \code{bbi_nonmem_model} or \code{bbi_nonmem_summary} object, or a path to
a NONMEM run.}

\item{.join_col}{Character column name to use to join table files. Defaults to
\code{NUM}. See Details.}

\item{.files}{Character vector of file paths to table files to read in.
Defaults to calling \code{\link[=nm_table_files]{nm_table_files()}} on \code{.mod}, which will parse all file
names from \verb{$TABLE} blocks in the control stream. If passing manually,
paths should be either absolute, or relative to \code{get_output_dir(.mod)}.}

\item{.superset}{If \code{FALSE}, the default, the data will be joined to the
NONMEM output and if \code{TRUE}, the NONMEM output will be joined to the data;
that is, if you use \code{.superset}, you will get the same number of rows as
you have in the input data and NONMEM output columns like \code{PRED} and
\code{CWRES} will be filled with \code{NA}.}

\item{.bbi_args}{Named list passed to \code{model_summary(.bbi_args)}. See
\code{\link[=print_bbi_args]{print_bbi_args()}} for valid options. Defaults to \code{list(no_grd_file = TRUE, no_shk_file = TRUE)} because \code{\link[=model_summary]{model_summary()}} is only called internally to
extract the number of records and individuals, so those files are
irrelevant.}
}
\description{
For NONMEM models, when a unique row identifier (e.g an integer numbering the
rows) is included in the input data set (i.e. the file in \verb{$DATA}) and
carried into each table output, \code{nm_join()} can read in all output table
files and join back to the input data set. By default, the input data is
joined to the table files so that the number of rows in the result will match
the number of rows in the table files (i.e. the number of rows \emph{not} bypassed via
\verb{$IGNORE}). Use the \code{.superset} argument to join table outputs to the
(complete) input data set. This function will print the number of rows and
columns when each file is loaded, as well as some information about the
joins. This \strong{printing can be suppressed} by setting \code{options(bbr.verbose = FALSE)}.
}
\details{
\strong{Join column}

The \code{.join_col} is the name of a single column that should appear in both the
input data set and any tables you want to join. We recommend you make this
column a simple integer numbering the rows in the input data set (for example
\code{NUM}). When this column is carried into the output table files, there will
be unambiguous matching from the table file back to the input data set.

The one exception to this are \code{FIRSTONLY} tables. If a table file has the
same number of rows as the there are individuals in the input data set
(accounting for any filtering of data in the NONMEM control stream), it will
assumed to be a \code{FIRSTONLY} table. In this case, the table will be joined to
the input data by the \code{ID} column. If \code{ID} is not present in the table, it
will be using \code{.join_col}. Note that if \emph{neither} \code{ID} or the column passed
to \code{.join_col} are present in the table, the join will fail.

Note also that, when \code{.join_col} is carried into table outputs, \strong{there is no
need to table any other columns from the input data} as long as the
\code{nm_join()} approach is used; any column in the input data set, regardless
of whether it is listed in \verb{$INPUT} or not, will be carried through from the
input data and therefore available in the joined result.

\strong{Duplicate columns are dropped}

If a table has columns with the same name as columns in the input data set,
or a table that has already been joined, those columns will be dropped from
the joined data. If \code{getOption(bbr.verbose) == TRUE} a message will be
printed about any columns dropped this way.

The one exception to this is the \code{DV} column. If \code{DV} is present in the input
data \emph{and} at least one of the table files, the \code{DV} column from the input
data will be renamed to \code{DV.DATA} and the column from the table file kept as
\code{DV}.

The origin of each column is attached to the return value via the
"nm_join_origin" attribute, a list that maps each source (as named by
\code{\link[=nm_tables]{nm_tables()}}) to the columns that came from that source.

\strong{Duplicate Rows Warning for Join Column}

If there are duplicate rows found in the specified \code{.join_col}, a warning will be raised specifying a subset of the repeated rows.
Duplicates may be caused by lack of output width. \code{FORMAT} may be need to be stated in control stream to have sufficient
width to avoid truncating \code{.join_col}.

\strong{Multiple tables per file incompatibility}

Because \code{nm_tables()} calls \code{\link[=nm_file]{nm_file()}} internally, it is \emph{not} compatible
with multiple tables written to a single file. See "Details" in \code{\link[=nm_file]{nm_file()}}
for alternatives.
}
\seealso{
\code{\link[=nm_tables]{nm_tables()}}, \code{\link[=nm_table_files]{nm_table_files()}}, \code{\link[=nm_file]{nm_file()}}
}
