% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-join.R
\name{nm_join_impl}
\alias{nm_join_impl}
\title{Implementation function for \code{nm_join()}}
\usage{
nm_join_impl(
  .mod,
  .join_col = "NUM",
  .files = nm_table_files(.mod),
  .superset = FALSE,
  .bbi_args = list(no_grd_file = TRUE, no_shk_file = TRUE)
)
}
\arguments{
\item{.mod}{A \code{bbi_nonmem_model} or \code{bbi_nonmem_summary} object, or a path to
a NONMEM run.}

\item{.join_col}{Character column name to use to join table files. Defaults to
\code{NUM}. See Details.}

\item{.files}{Character vector of file paths to table files to read in.
Defaults to calling \code{\link[=nm_table_files]{nm_table_files()}} on \code{.mod}, which will parse all file
names from \verb{$TABLE} blocks in the control stream. If passing manually,
paths should be either absolute, or relative to \code{get_output_dir(.mod)}.}

\item{.superset}{If \code{FALSE}, the default, the data will be joined to the
NONMEM output and if \code{TRUE}, the NONMEM output will be joined to the data;
that is, if you use \code{.superset}, you will get the same number of rows as
you have in the input data and NONMEM output columns like \code{PRED} and
\code{CWRES} will be filled with \code{NA}.}

\item{.bbi_args}{Named list passed to \code{model_summary(.bbi_args)}. See
\code{\link[=print_bbi_args]{print_bbi_args()}} for valid options. Defaults to \code{list(no_grd_file = TRUE, no_shk_file = TRUE)} because \code{\link[=model_summary]{model_summary()}} is only called internally to
extract the number of records and individuals, so those files are
irrelevant.}
}
\description{
Implementation function for \code{nm_join()}
}
\keyword{internal}
