% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-join.R
\name{nm_join_sim}
\alias{nm_join_sim}
\title{Join simulation and input data}
\usage{
nm_join_sim(.mod, .join_col = "NUM", .files = nm_table_files(.mod))
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} or \code{bbi_nmsim_model} object.}

\item{.join_col}{Character column name to use to join table files. Defaults to
\code{NUM}. See Details.}

\item{.files}{Character vector of file paths to table files to read in.
Defaults to calling \code{\link[=nm_table_files]{nm_table_files()}} on \code{.mod}, which will parse all file
names from \verb{$TABLE} blocks in the control stream. If passing manually,
paths should be either absolute, or relative to \code{get_output_dir(.mod)}.}
}
\value{
a tibble
}
\description{
Join simulation and input data
}
\section{Join column}{


The \code{.join_col} is the name of a single column that should appear in both the
input data set and any tables you want to join. We recommend you make this
column a simple integer numbering the rows in the input data set (for example
\code{NUM}). When this column is carried into the output table files, there will
be unambiguous matching from the table file back to the input data set.
\itemize{
\item The one exception to this are \code{FIRSTONLY} tables. If a table file has the
same number of rows as the there are individuals in the input data set
(accounting for any filtering of data in the NONMEM control stream), it
will assumed to be a \code{FIRSTONLY} table. In this case, the table will be
joined to the input data by the \code{ID} column. If \code{ID} is not present in
the table, it will be using \code{.join_col}. Note that if \emph{neither} \code{ID} or
the column passed to \code{.join_col} are present in the table, the join will
fail.
}

Note also that, when \code{.join_col} is carried into table outputs, \strong{there is no
need to table any other columns from the input data} as long as the
\code{nm_join()} approach is used; any column in the input data set, regardless
of whether it is listed in \verb{$INPUT} or not, will be carried through from the
input data and therefore available in the joined result.
}

\seealso{
\code{\link[=nm_file_multi_table]{nm_file_multi_table()}}, \code{\link[=nm_join]{nm_join()}}
}
