% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-tables.R
\name{nm_tables}
\alias{nm_tables}
\alias{nm_table_files}
\title{Read all tables and input data}
\usage{
nm_tables(.mod, .files = nm_table_files(.mod))

nm_table_files(.mod, .check_exists = TRUE)
}
\arguments{
\item{.mod}{Either a \code{bbi_nonmem_model} or \code{bbi_nonmem_summary} object}

\item{.files}{Character vector of file paths to table files to read in.
Defaults to calling \code{\link[=nm_table_files]{nm_table_files()}} on \code{.mod}, which will parse all file
names from \verb{$TABLE} blocks in the control stream.}

\item{.check_exists}{If \code{TRUE} check if files exist.}
}
\value{
A named list of tibbles. The first element will always be named
\code{data} and will contain the input data set. Subsequent elements will be
named for the file from which they were loaded, with \code{get_model_id(.mod)}
(and \code{.}) removed from the beginning and end, if present. For example, a
model named \verb{001.ctl} that generated a table named \verb{001.tab} will have the
relevant element named \code{tab}.
}
\description{
Reads in the input data set and all table output files from a NONMEM run.
This function will return a named list with all the relevant tibbles. To return
a \emph{single tibble} with the input data joined to the relevant table outputs, use
the related \code{\link[=nm_join]{nm_join()}} function.
}
\section{Functions}{
\itemize{
\item \code{nm_table_files}: Extract paths to table output files from NONMEM control
stream, and optionally check if the files exist.
}}

\seealso{
\code{\link[=nm_join]{nm_join()}}, \code{\link[=nm_file]{nm_file()}}
}
