% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-nmtran.R
\name{nmtran}
\alias{nmtran}
\alias{run_nmtran}
\title{Interface for running \code{NM-TRAN} on model objects}
\usage{
run_nmtran(
  .mod,
  .bbi_args = list(prdefault = FALSE, tprdefault = FALSE, maxlim = 2),
  .config_path = NULL,
  nmtran_exe = NULL,
  delete_on_exit = TRUE
)
}
\arguments{
\item{.mod}{A \code{bbr} model object.}

\item{.bbi_args}{A named list specifying arguments to pass to \code{NM-TRAN}.
Similar to the \code{.bbi_args} argument defined in \code{\link[=submit_model]{submit_model()}}, though here
only \code{prdefault}, \code{tprdefault}, and \code{maxlim} flags are passed to \code{NM-TRAN}.
See details.}

\item{.config_path}{Path to a bbi configuration file. If \code{NULL}, the default,
will attempt to use a \code{bbi.yaml} in the same directory as the model.}

\item{nmtran_exe}{Path to an \code{NM-TRAN} executable. If \code{NULL}, will look for a
\code{bbi.yaml} file in the same directory as the model.}

\item{delete_on_exit}{Logical. If \code{FALSE}, don't delete the temporary folder
containing the \code{NM-TRAN} run, which will be stored in the current working
directory.}
}
\description{
Provides functions to run \code{NM-TRAN} on model objects for validation and to
generate \code{NM-TRAN} datasets (\code{FDATA}).
}
\details{
\code{NM-TRAN} is a preprocessor for \code{NONMEM} that translates user-specified
control stream data and instructions into a form executable by \code{NONMEM}.

\code{run_nmtran()} allows users to test their models ahead of submission to ensure
correct coding.
}
\section{Functions}{
\itemize{
\item \code{run_nmtran()}: Run \code{NM-TRAN} on a model object to validate its control
stream for correct coding before submission.

}}
\section{Supported \code{bbi_args}}{

\code{run_nmtran()} supports passing the following raw \code{NMFE} options through to \code{NM-TRAN}
via the familiar \code{.bbi_args} argument.
\itemize{
\item \strong{\code{maxlim}}: Set the maximum values for the buffers used by \code{NONMEM}
(if 0, don't pass \code{-maxlim} to nmfe) (default is 2)
\itemize{
\item If \code{maxlim = 3}, it is preferred to also set \code{tprdefault = TRUE}, but
\emph{not} \code{prdefault}, as \code{NM-TRAN}'s optional resizing of the \code{PREDPP} size
parameter \code{MAXRECID} may conflict with the \code{-prdefault} option.
}
\item \strong{\code{prdefault}}: If \code{TRUE}, do not recompile any routines other than \code{FSUBS}
\item \strong{\code{tprdefault}}: If \code{TRUE}, test if is okay to do \code{-prdefault}
}
}

\examples{
\dontrun{
mod <- read_model(file.path(MODEL_DIR, 1))
run_nmtran(mod)

# Set the path to an NM-TRAN executable
run_nmtran(mod, nmtran_exe = "/opt/NONMEM/nm75/tr/NMTRAN.exe")

}
}
