% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-nmtran.R
\name{nmtran}
\alias{nmtran}
\alias{run_nmtran}
\title{Interface for running \code{NM-TRAN} on model objects}
\usage{
run_nmtran(.mod, .config_path = NULL, nmtran_exe = NULL, delete_on_exit = TRUE)
}
\arguments{
\item{.mod}{A \code{bbr} model object.}

\item{.config_path}{Path to a bbi configuration file. If \code{NULL}, the default,
will attempt to use a \code{bbi.yaml} in the same directory as the model.}

\item{nmtran_exe}{Path to an \code{NM-TRAN} executable. If \code{NULL}, will look for a
\code{bbi.yaml} file in the same directory as the model.}

\item{delete_on_exit}{Logical. If \code{FALSE}, don't delete the temporary folder
containing the \code{NM-TRAN} run, which will be stored in the current working
directory.}
}
\description{
Provides functions to run \code{NM-TRAN} on model objects for validation and to
generate \code{NM-TRAN} datasets (\code{FDATA}).
}
\details{
\code{NM-TRAN} is a preprocessor for \code{NONMEM} that translates user-specified
control stream data and instructions into a form executable by \code{NONMEM}.

\code{run_nmtran()} allows users to test their models ahead of submission to ensure
correct coding.
}
\section{Functions}{
\itemize{
\item \code{run_nmtran()}: Run \code{NM-TRAN} on a model object to validate its control
stream for correct coding before submission.

}}
\examples{
\dontrun{
mod <- read_model(file.path(MODEL_DIR, 1))
run_nmtran(mod)

# Set the path to an NM-TRAN executable
run_nmtran(mod, nmtran_exe = "/opt/NONMEM/nm75/tr/NMTRAN.exe")

}
}
