% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{nonmem_output}
\alias{nonmem_output}
\title{get a model summary}
\usage{
nonmem_output(
  .path,
  .args = NULL,
  ...,
  .json = TRUE,
  .parse = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{.path}{Full path to one of the following:
1) path to a model file in the output directory
2) path to a nonmem output directory}

\item{.args}{A named list specifying arguments to pass to babylon formatted like `list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)`. Run `print_nonmem_args()` to see valid arguments.}

\item{...}{args passed through to `bbi_exec()`}

\item{.json}{return output as json}

\item{.parse}{parse the output and return the json rather than the stderr/stdout}

\item{.dry_run}{show what the command would be without actually running it}
}
\value{
output from the model run (?)
}
\description{
get a model summary
}
