% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{nonmem_summary}
\alias{nonmem_summary}
\title{Run `bbi summary` and parse the output to a list}
\usage{
nonmem_summary(.path, .args = NULL, ..., .dry_run = FALSE)
}
\arguments{
\item{.path}{Full path to one of the following:
1) path to a model file in the output directory
2) path to a nonmem output directory}

\item{.args}{A named list specifying arguments to pass to babylon formatted like `list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)`. Run `print_nonmem_args()` to see valid arguments.}

\item{...}{args passed through to `bbi_exec()`}

\item{.dry_run}{show what the command would be without actually running it}
}
\value{
List of S3 Class "bbi_nonmem_summary" with all summary information
}
\description{
Run `bbi summary` and parse the output to a list
}
