% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-summary.R
\name{nonmem_summary}
\alias{nonmem_summary}
\title{Run \verb{bbi nonmem summary} and parse the output to a list}
\usage{
nonmem_summary(.mod, .bbi_args = NULL, ..., .dry_run = FALSE)
}
\arguments{
\item{.mod}{\code{bbi_nonmem_model} object for summary}

\item{.bbi_args}{A named list specifying arguments to pass to babylon formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}. Run \code{print_nonmem_args()} to see valid arguments.}
}
\value{
List of S3 Class "bbi_nonmem_summary" with all summary information
}
\description{
Private implementation function called by \code{model_summary()} dispatches.
}
\keyword{internal}
