% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimates.R
\name{param_estimates}
\alias{param_estimates}
\alias{param_estimates.bbi_nonmem_summary}
\title{Parses parameter estimates table}
\usage{
param_estimates(.summary)

\method{param_estimates}{bbi_nonmem_summary}(.summary)
}
\arguments{
\item{.summary}{A \verb{bbi_\{.model_type\}_summary} object.}
}
\value{
Returns a tibble with the following columns:
\itemize{
\item parameter_names -- Parameter name ("THETA1", "THETA2", etc.)
\item estimate -- Parameter estimate
\item stderr -- Standard Error
\item random_effect_sd -- OMEGA and SIGMA elements in standard deviation/correlation format
\item random_effect_sdse -- Standard errors to the OMEGA and SIGMA elements in standard deviation/correlation format
\item fixed -- TRUE if parameter is fixed, FALSE otherwise
\item diag -- TRUE if parameter is a diagonal element in random effect matrix, FALSE if off-diagonal, NA if parameter is a THETA
}
}
\description{
Returns a tibble containing parameter estimates from a \verb{bbi_\{.model_type\}_summary} object.
Details about the tibble that is returned are in the return value section below.
}
\section{Methods (by class)}{
\itemize{
\item \code{bbi_nonmem_summary}: Takes \code{bbi_nonmem_summary} object, the output of \code{\link[=model_summary.bbi_nonmem_model]{model_summary.bbi_nonmem_model()}}.
}}

\seealso{
\code{\link[=param_labels]{param_labels()}} \code{\link[=apply_indices]{apply_indices()}}
}
