% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimates.R
\name{param_estimates}
\alias{param_estimates}
\alias{param_estimates.bbi_nonmem_summary}
\title{Parses parameter estimates table}
\usage{
param_estimates(.summary)

\method{param_estimates}{bbi_nonmem_summary}(.summary)
}
\arguments{
\item{.summary}{\verb{bbi_\{.model_type\}_summary} object}
}
\description{
Returns a tibble containing parameter estimates from a model.
Currently can only take a \verb{bbi_\{.model_type\}_summary} object, as output from \code{model_summary()}.
}
\section{Methods (by class)}{
\itemize{
\item \code{bbi_nonmem_summary}: Takes \code{bbi_nonmem_summary} object.
}}

\seealso{
\code{param_labels()} \code{apply_indices()}
}
