% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimates.R
\name{param_estimates}
\alias{param_estimates}
\alias{param_estimates.bbi_nonmem_summary}
\title{Parses parameter estimates table from model_summary() output}
\usage{
param_estimates(.summary)

\method{param_estimates}{bbi_nonmem_summary}(.summary)
}
\arguments{
\item{.summary}{\code{bbi_nonmem_summary} object}
}
\description{
Parses parameter estimates table from model_summary() output

S3 dispatch for parsing \code{bbi_nonmem_summary} object into parameter estimate table
}
