% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimates-batch.R
\name{param_estimates_compare}
\alias{param_estimates_compare}
\alias{param_estimates_compare.bbi_nmboot_summary}
\alias{param_estimates_compare.default}
\title{Compare parameter estimates}
\usage{
param_estimates_compare(
  .boot_sum,
  .orig_mod = NULL,
  .compare_cols = starts_with(c("THETA", "SIGMA", "OMEGA")),
  probs = c(0.5, 0.025, 0.975),
  na.rm = FALSE
)

\method{param_estimates_compare}{bbi_nmboot_summary}(
  .boot_sum,
  .orig_mod = NULL,
  .compare_cols = NULL,
  probs = c(0.5, 0.025, 0.975),
  na.rm = FALSE
)

\method{param_estimates_compare}{default}(
  .boot_sum,
  .orig_mod = NULL,
  .compare_cols = starts_with(c("THETA", "SIGMA", "OMEGA")),
  probs = c(0.5, 0.025, 0.975),
  na.rm = FALSE
)
}
\arguments{
\item{.boot_sum}{Either a \code{bbi_nmboot_summary} object, or a tibble with
columns for each parameter and rows for each model (like what's returned
from \code{\link[=param_estimates_batch]{param_estimates_batch()}}).}

\item{.orig_mod}{\code{bbi_model} object to compare \code{.boot_sum} against. This will
be automatically set if passing in a \code{bbi_nmboot_summary} object.}

\item{.compare_cols}{An expression that can be passed to \code{\link[dplyr:select]{dplyr::select()}} to
select which columns in \code{.boot_sum} will be pivoted and summarized. See
\href{https://tidyselect.r-lib.org/reference/language.html}{?tidyselect::language}
for more details and options. \strong{Only used} if \code{.boot_sum} is a tibble.}

\item{probs}{Numeric vector with values between 0 and 1 to be passed through to
\code{\link[stats:quantile]{stats::quantile()}}. Represents the quantiles to calculate for parameter
estimates in \code{.boot_sum}.}

\item{na.rm}{Logical scalar, passed through to \code{\link[stats:quantile]{stats::quantile()}}.}
}
\value{
A tibble containing quantiles for each parameter estimate, optionally
compared to the estimates from \code{.orig_mod}.
}
\description{
Summarizes each parameter estimate, showing you the quantiles passed to
\code{probs}, and optionally the estimates of \code{.orig_mod}.
}
\examples{
\dontrun{

# Via a bootstrap run
boot_run <- read_model(file.path(MODEL_DIR, "1_boot"))
boot_sum <- summarize_bootstrap_run(boot_run)
param_estimates_compare(boot_sum)

# Via a custom table
orig_mod <- read_model(file.path(MODEL_DIR, "1"))
param_df <- param_estimates_batch(MODEL_DIR)
param_estimates_compare(param_df, .orig_mod = orig_mod)
}

}
