% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-estimates-batch.R
\name{param_estimates_compare}
\alias{param_estimates_compare}
\title{Compare parameter estimates}
\usage{
param_estimates_compare(.mod, .comp, .probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{.mod}{\code{bbi_model} object}

\item{.comp}{A tibble with columns for each parameter and rows for each model (like what's returned from
\code{\link[=param_estimates_batch]{param_estimates_batch()}})}

\item{.probs}{Numeric vector with between 0 and 1 to be passed through to \code{quantile}.
Represents the quantiles to calculate for parameter estimates in \code{.comp}.}
}
\description{
Let's you compare the parameter estimates from  a single model
against those in tibble like what's returned from
\code{\link[=param_estimates_batch]{param_estimates_batch()}}.
}
