% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-labels.R
\name{param_labels}
\alias{param_labels}
\alias{param_labels.bbi_nonmem_model}
\alias{param_labels.character}
\title{Generic s3 for parsing parameter labels from control stream}
\usage{
param_labels(.mod, ...)

\method{param_labels}{bbi_nonmem_model}(.mod, ...)

\method{param_labels}{character}(.mod, ...)
}
\arguments{
\item{.mod}{Character scaler of the raw control stream with newline character separating lines}

\item{...}{arguments passed through}
}
\description{
This is using the labeling scheme from "Census", as defined here: https://ghe.metrumrg.com/pages/software/mrgtable/reference/nm_tbl.html
It will _not_ return indices for the parameters, though they can be added with `param_labels() %>% apply_indices()`
}
