% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_mod_yaml}
\alias{parse_mod_yaml}
\title{Parses a model yaml file into a list object that contains correctly formatted information from the yaml
Output list will have:
  $model_path -- the path the model file that should be run
  $args_list -- any babylon arguments that should be passed though. The keys must be the same as would be passed to check_nonmem_args()
  $user_data -- any other data the user included, such as metadata about the model, that is NOT a babylon argument}
\usage{
parse_mod_yaml(.path)
}
\arguments{
\item{.path}{Path to the yaml file to parse.}
}
\value{
Output list as specified above.
}
\description{
Parses a model yaml file into a list object that contains correctly formatted information from the yaml
Output list will have:
  $model_path -- the path the model file that should be run
  $args_list -- any babylon arguments that should be passed though. The keys must be the same as would be passed to check_nonmem_args()
  $user_data -- any other data the user included, such as metadata about the model, that is NOT a babylon argument
}
