% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-nmtran.R
\name{parse_nmtran_args}
\alias{parse_nmtran_args}
\title{Parse \code{.bbi_args} and return the three expected \code{nmfe_options} for \code{NM-TRAN}
in the correct format}
\usage{
parse_nmtran_args(.mod, .bbi_args = NULL, nmfe_options = NULL)
}
\arguments{
\item{.mod}{A \code{bbi_nonmem_model} object.}

\item{.bbi_args}{A named list specifying arguments to pass to \code{NM-TRAN}.
Similar to the \code{.bbi_args} argument defined in \code{\link[=submit_model]{submit_model()}}, though here
only \code{prdefault}, \code{tprdefault}, and \code{maxlim} arguments are passed to \code{NM-TRAN}.
\code{nm_version} is also supported and specifies which \code{NM-TRAN} executable to use.
See \code{\link[=print_bbi_args]{print_bbi_args()}} for more details.}

\item{nmfe_options}{named list of nmfe options defined in a \code{bbi.yaml} file.}
}
\value{
a named character vector
}
\description{
Parse \code{.bbi_args} and return the three expected \code{nmfe_options} for \code{NM-TRAN}
in the correct format
}
\details{
Combines NONMEM submission args and consolidates to \code{NMFE} arguments only
\itemize{
\item The arguments of interest are \code{prdefault}, \code{tprdefault}, and \code{maxlim},
which impact the evaluation of \code{NM-TRAN}. Priority is \code{.bbi_args} > model
yaml > \code{bbi.yaml}
\item Note: \code{run_nmtran()} considers only the first three \code{NM-TRAN} arguments
(\code{prdefault}, \code{tprdefault}, and \code{maxlim}), but, starting with \verb{NONMEM 7.5},
nmfe passes a fourth argument (\code{do2test}) to NM-TRAN.
\itemize{
\item \code{bbi} \emph{doesn't} expose nmfe's \code{-do2test} argument, so in the \code{bbi}/\code{bbr}
context this call always passes \code{"0"} to \code{NM-TRAN} for the \code{do2test}
value. If that first call exits with a status of \code{9}, nmfe does a follow-up
call that drops the fourth argument (\code{do2test}).
\item Given the above, \code{parse_nmtran_args()} does not support the additional
\code{do2test} argument for \verb{NONMEM 7.5}. This should be revisited if
\code{run_nmtran()} is called within \code{submit_model()} as a method of validating
the control stream ahead of model execution. See additional discussion at
https://github.com/metrumresearchgroup/bbr/pull/705.
}
}
}
\keyword{internal}
