% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-labels.R
\name{parse_param_comment}
\alias{parse_param_comment}
\title{extract labels from comments in raw ctl string}
\usage{
parse_param_comment(.x, .theta = TRUE)
}
\arguments{
\item{.x}{an object from the list that comes out of clean_ctl() (i.e. .ctl_clean$THETA, .ctl_clean$SIGMA, .ctl_clean$OMEGA)}

\item{.theta}{Boolean for if it's a theta. If TRUE, parses `[{}]` to unit, otherwise parses to type}
}
\description{
Modified tidynm::parse_theta() to parse out the comments into the right columns for all three params.
}
