% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-labels.R
\name{parse_param_comment}
\alias{parse_param_comment}
\alias{parse_same_block}
\title{extract labels from comments in raw ctl string}
\usage{
parse_param_comment(.x, .theta = FALSE)

parse_same_block(.x)
}
\arguments{
\item{.x}{List of character vectors representing either $OMEGA or $SIGMA blocks, as parsed by `clean_ctl()`}

\item{.theta}{Boolean for if it's a theta. If TRUE, parses `[{}]` to unit, otherwise parses to type}
}
\description{
Heavily modified from tidynm::parse_theta() to parse out the comments into the right columns for all three params.

Iterates over a list of $OMEGA or $SIGMA blocks and replaces the SAME blocks with a copy of the block they are referring to.
}
