% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-output.R
\name{plot_nonmem_table_df}
\alias{plot_nonmem_table_df}
\alias{plot_grd}
\alias{plot_ext}
\title{Plot NONMEM output files}
\usage{
plot_nonmem_table_df(.df, .x_var, .stat_name)

plot_grd(.df)

plot_ext(.df)
}
\arguments{
\item{.df}{the wide-format tibble output from \code{check_nonmem_table_output()}}

\item{.x_var}{String of the variable name to use on the X-axis}

\item{.stat_name}{String of the name of the stat the other columns represents (like "gradient" or "theta").}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Creates a line plot of the wide-format tibble output from \code{check_nonmem_table_output()}
}
\section{Functions}{
\itemize{
\item \code{plot_grd}: Plot the .grd file

\item \code{plot_ext}: Plot the .ext file
}}

