% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_bbi}
\alias{print_bbi}
\alias{print.bbi_process}
\alias{print.bbi_model}
\alias{print.bbi_nonmem_summary}
\alias{print.bbi_nmboot_summary}
\alias{print.bbi_nmsse_summary}
\alias{print.nmtran_process}
\alias{print.model_tree_static}
\title{Print methods for bbr objects}
\usage{
\method{print}{bbi_process}(x, ..., .call_limit = 250)

\method{print}{bbi_model}(x, ...)

\method{print}{bbi_nonmem_summary}(x, .digits = 3, .fixed = FALSE, .off_diag = FALSE, .nrow = NULL, ...)

\method{print}{bbi_nmboot_summary}(x, .digits = 3, .nrow = 10, ...)

\method{print}{bbi_nmsse_summary}(x, .digits = 3, .nrow = 10, ...)

\method{print}{nmtran_process}(x, ...)

\method{print}{model_tree_static}(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
\item{x}{plot to display}

\item{...}{other arguments not used by this method}

\item{.call_limit}{Integer scalar for the max number of characters to print
before truncating the call string. This is compared with the entire length,
but only the positional arguments between the executable path and the first
long option will be truncated.}

\item{.digits}{Number of significant digits to use for parameter table. Defaults to 3.}

\item{.fixed}{If \code{FALSE}, the default, omits fixed parameters from the parameter table.}

\item{.off_diag}{If \code{FALSE}, the default, omits off-diagonals of OMEGA and SIGMA matrices from the parameter table.}

\item{.nrow}{If \code{NULL}, the default, print all rows of the parameter table. If \code{0}, don't print table at all.
Otherwise, prints only \code{.nrow} rows.}

\item{newpage}{Logical (T/F). If \code{TRUE}, draw new (empty) page first.}

\item{vp}{viewport to draw plot in}
}
\description{
The various objects defined by \code{bbr} have their own print methods to
allow users to get a quick view of the contents of the object.
\strong{When printing a \code{bbi} object in an \code{.Rmd} file} that is intended to be
knit, consider setting \code{results = 'asis'} in the chunk options. This
will make for prettier formatting, especially of table outputs.
}
\section{Functions}{
\itemize{
\item \code{print(bbi_process)}: Prints the call made to bbi and whether the process is still running or has finished.

\item \code{print(bbi_model)}: Prints the information contained in the model object and whether the model has been run

\item \code{print(bbi_nonmem_summary)}: Prints a high level summary of a model from a \code{bbi_nonmem_summary} object

\item \code{print(bbi_nmboot_summary)}: Prints a high level summary of a model from a \code{bbi_nmboot_summary} object

\item \code{print(bbi_nmsse_summary)}: Prints a high level summary of a model from a \code{bbi_nmsse_summary} object

\item \code{print(nmtran_process)}: Prints the \code{NM-TRAN} evaluation of a \code{bbi_nonmem_model}
object

\item \code{print(model_tree_static)}: Draw model tree as a static plot

}}
\examples{
\dontrun{
pl_tree <- model_tree(MODEL_DIR, static = TRUE)

print(pl_tree)
print(pl_tree, vp = grid::viewport(width=0.5, height=0.5))
print(pl_tree, newpage = TRUE, vp = grid::viewport(width=0.5, height=0.5))
}
}
