% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbabylon.R
\name{rbabylon}
\alias{rbabylon}
\title{RBabylon}
\description{
RBabylon is a library written in R that facilitates modeling and simulation by exposing functionality from the https://github.com/metrumresearchgroup/babylon project in R
From Babylon's docs: Babylon is (will be) a complete solution for managing projects involving modeling and simulation with a number of software solutions used in pharmaceutical sciences.
This is a fork of the `nonmemutils`` project that is broader in scope.
Initial support encompasses NONMEM however the api is designed in a way to be flexible to handle other software.
}
