% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{read_model}
\alias{read_model}
\title{Creates a model object from a YAML model file}
\usage{
read_model(.path)
}
\arguments{
\item{.path}{Path to the YAML file to parse.}
}
\value{
S3 object of class `bbi_{.model_type}_model`
}
\description{
Parses a model YAML file into a list object that contains correctly formatted information from the YAML
and is an S3 object of class `bbi_{.model_type}_model` that can be passed to `submit_model()`, `model_summary()`, etc.
}
