% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{read_model}
\alias{read_model}
\title{Creates a model object from a YAML model file}
\usage{
read_model(.path, .directory = get_model_directory())
}
\arguments{
\item{.path}{Path to the YAML file to parse. MUST be either an absolute path, or a path relative to the `.directory` argument.}

\item{.directory}{Model directory which `.path` is relative to. Defaults to `options('rbabylon.model_directory')`, which can be set globally with `set_model_directory()`.}
}
\value{
S3 object of class `bbi_{.model_type}_model`
}
\description{
Parses a model YAML file into a list object that contains correctly formatted information from the YAML
and is an S3 object of class `bbi_{.model_type}_model` that can be passed to `submit_model()`, `model_summary()`, etc.
}
