% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify-model-field.R
\name{remove_star}
\alias{remove_star}
\title{remove_star}
\usage{
remove_star(.mod, .starred)
}
\arguments{
\item{.mod}{list object of model data}

\item{.starred}{boolean variable to indicate significance}
}
\description{
remove_star removes starring from a model object and corresponding YAML.
}
