% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tree.R
\name{req_tree_pkgs}
\alias{req_tree_pkgs}
\alias{check_for_model_tree_pkgs}
\alias{skip_if_tree_missing_deps}
\alias{stop_if_tree_missing_deps}
\title{Required packages for running \code{\link[=model_tree]{model_tree()}}}
\usage{
req_tree_pkgs(static = FALSE)

check_for_model_tree_pkgs(static = FALSE)

skip_if_tree_missing_deps(static = FALSE)

stop_if_tree_missing_deps(static = FALSE)
}
\arguments{
\item{static}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, check for additional
required packages for rendering the plot as a PNG.}
}
\value{
Returns a vector with the missing packages, or returns NULL if all are
present.
}
\description{
Required packages for running \code{\link[=model_tree]{model_tree()}}
}
\details{
The packages below are required for both interactive and static viewing:
\itemize{
\item \strong{\code{collapsibleTree}} for the core plot
}

The packages below are all used for rendering the model tree as a \strong{static} image:
\itemize{
\item \strong{\code{htmlwidgets}} is a dependency of \code{collapsibleTree}, so it's a 'free'
import. It's used to save the tree diagram out as an \code{HTML} file.
\item \strong{\code{grid}} is a built-in R package and is used to display the PNG in the
Rstudio viewer.
\item \strong{\code{webshot}} is needed for taking a screenshot of the rendered HTML,
saved out via \code{htmlwidgets::saveWidget()}.
\item \strong{\code{png}} is needed for reading in a PNG file in order to plot it in the
Rstudio viewer.
}
}
\section{Functions}{
\itemize{
\item \code{check_for_model_tree_pkgs()}: Checks if all packages needed for \code{\link[=model_tree]{model_tree()}} are present

\item \code{skip_if_tree_missing_deps()}: Skip tests if missing \code{\link[=model_tree]{model_tree()}} dependencies

\item \code{stop_if_tree_missing_deps()}: Error if missing \code{\link[=model_tree]{model_tree()}} dependencies. Called internally
within \code{model_tree()}

}}
\keyword{internal}
