% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-log.R
\name{run_log}
\alias{run_log}
\title{Create tibble summarizing all model runs}
\usage{
run_log(.base_dir, .recurse = TRUE, .filter = 0:0)
}
\arguments{
\item{.base_dir}{Base directory to look in for models.}

\item{.recurse}{If \code{TRUE}, the default, search recursively in all subdirectories. Passed through to \code{fs::dir_ls()} -- If a positive number, the number of levels to recurse.}

\item{.filter}{Provides filter for runs based on numeric labeling. Filters inclusively}
}
\value{
A tibble of class \code{bbi_run_log_df} with information on each model, or an empty tibble if no models are found.
}
\description{
Parses all model YAML files and outputs into a tibble that serves as a run log for the project.
Future releases will incorporate more diagnostics and parameter estimates, etc. from the runs into this log.
Users can also use \code{\link[=add_config]{add_config()}} or \code{\link[=add_summary]{add_summary()}} to append additional output about the model run.
}
\seealso{
\code{\link[=config_log]{config_log()}}, \code{\link[=summary_log]{summary_log()}}
}
