% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{run_log}
\alias{run_log}
\title{Parses model YAML files (and optionally model outputs) into a log tibble}
\usage{
run_log(.base_dir = ".", .recurse = TRUE)
}
\arguments{
\item{.base_dir}{Directory to look in for YAML model files}

\item{.recurse}{Boolean for whether to search recursively in subdirectories}
}
\value{
tibble with information on each run
}
\description{
Parses model YAML files (and optionally model outputs) into a log tibble
}
