% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-log.R
\name{run_log}
\alias{run_log}
\title{Parses model yaml and outputs into a tibble that serves as a run log. Future releases will incorporate more diagnostics and parameter estimates, etc. from the runs into this log.}
\usage{
run_log(.base_dir = get_model_directory(), .recurse = TRUE)
}
\arguments{
\item{.base_dir}{Directory to search for model yaml files. Only runs with a corresponding yaml will be included.}

\item{.recurse}{Boolean for whether to search subdirectories recursively for additional yaml files. Defaults to TRUE.}
}
\value{
tibble with information on each run
}
\description{
Parses model yaml and outputs into a tibble that serves as a run log. Future releases will incorporate more diagnostics and parameter estimates, etc. from the runs into this log.
}
