% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-nmtran.R
\name{run_nmtran}
\alias{run_nmtran}
\title{Interface for running \code{NM-TRAN} on model objects}
\usage{
run_nmtran(
  .mod,
  .bbi_args = list(prdefault = FALSE, tprdefault = FALSE, maxlim = 2),
  .nonmem_version = NULL,
  .config_path = NULL,
  clean = TRUE
)
}
\arguments{
\item{.mod}{A \code{bbi_nonmem_model} object.}

\item{.bbi_args}{A named list specifying arguments to pass to \code{NM-TRAN}.
Similar to the \code{.bbi_args} argument defined in \code{\link[=submit_model]{submit_model()}}, though here
only \code{prdefault}, \code{tprdefault}, and \code{maxlim} flags are passed to \code{NM-TRAN}.
See \code{\link[=print_bbi_args]{print_bbi_args()}} for more details.}

\item{.config_path}{Path to a bbi configuration file. If \code{NULL}, the
default, will attempt to use a \code{bbi.yaml} in the same directory as the
model.}

\item{clean}{Logical. If \code{FALSE}, don't delete the temporary folder
containing the \code{NM-TRAN} run, which will be stored in the current working
directory.}

\item{nmtran_exe}{Path to an \code{NM-TRAN} executable. If \code{NULL}, will look for a
\code{bbi.yaml} file in the same directory as the model.}
}
\description{
Function to run \code{NM-TRAN} on a model object to validate its control stream
for correct coding before submission. The \code{NM-TRAN} dataset (\code{FDATA}) and
other \code{NONMEM} artifacts can be further inspected by keeping the run directory
around.
}
\details{
\code{NM-TRAN} is a preprocessor for \code{NONMEM} that translates user-specified
control stream data and instructions into a form executable by \code{NONMEM}.

\code{run_nmtran()} allows users to test their models ahead of submission to ensure
correct coding.
}
\examples{
\dontrun{
mod <- read_model(file.path(MODEL_DIR, 1))
run_nmtran(mod)

# Set the path to an NM-TRAN executable
run_nmtran(mod, nmtran_exe = "/opt/NONMEM/nm75/tr/NMTRAN.exe")

}

}
