% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-nmtran.R
\name{run_nmtran}
\alias{run_nmtran}
\title{Run NMTRAN on a model object}
\usage{
run_nmtran(.mod, .config_path = NULL, nmtran_exe = NULL, delete_on_exit = TRUE)
}
\arguments{
\item{.mod}{a \code{bbr} model object}

\item{.config_path}{Path to a bbi configuration file. If \code{NULL}, the
default, will attempt to use a \code{bbi.yaml} in the same directory as the
model.}

\item{nmtran_exe}{Path to an \code{NMTRAN} executable. If \code{NULL}, will look for a
\code{bbi.yaml} file in the same directory as the model.}

\item{delete_on_exit}{Logical. If \code{FALSE}, don't delete the temporary folder
containing the \code{NMTRAN} run.}
}
\description{
Run NMTRAN on a model object
}
\examples{
\dontrun{
mod <- read_model(file.path(MODEL_DIR, 1))
run_nmtran(mod)

# Set the path to an NMTRAN executable
run_nmtran(mod, nmtran_exe = "/opt/NONMEM/nm75/tr/NMTRAN.exe")
}

}
