% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-log.R
\name{safe_read_model}
\alias{safe_read_model}
\title{Read in model YAML with error handling}
\usage{
safe_read_model(.path)
}
\arguments{
\item{.path}{Path to the model to read, in the sense of absolute model path.
The absolute model path is the path to the YAML file and model file, both
without extension, and (possibly) the output directory.}
}
\value{
A model object, if \code{.path} represents a valid model.
}
\description{
Private helper function that tries to call \code{\link[=read_model]{read_model()}} on a yaml path and
returns NULL, with no error, if the YAML is not a valid model file. Also
wraps \code{\link[=read_model]{read_model()}} in \code{suppressMessages()} so that things like the
\code{\link[=check_stan_model]{check_stan_model()}} messages will be suppressed. This is done primarily
because this function is primarily used when mapping over a number of models
and these messages can be very annoying and mostly pointless in that context.
}
\keyword{internal}
