% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-log.R
\name{safe_read_model}
\alias{safe_read_model}
\title{Read in model with error handling}
\usage{
safe_read_model(.path)
}
\arguments{
\item{.path}{Path to the model to read, in the sense of absolute model path.
The absolute model path is the path to the YAML file and model file, both
without extension, and (possibly) the output directory.}
}
\value{
A model object, if \code{.path} represents a valid model.
}
\description{
Private helper function that tries to call \code{read_model()} on a model path and returns NULL, with no error, if the YAML is not a valid model file.
}
\keyword{internal}
