% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{save_model_yaml}
\alias{save_model_yaml}
\title{Saves a model object to a yaml file}
\usage{
save_model_yaml(.mod)
}
\arguments{
\item{.mod}{S3 object of class \verb{bbi_\{.model_type\}_model}}
}
\value{
Output list as specified above.
}
\description{
Saves a model object to a yaml file
}
\keyword{internal}
