% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-model.R
\name{save_model_yaml}
\alias{save_model_yaml}
\title{Saves a model model object to a yaml file}
\usage{
save_model_yaml(.mod, .out_path = NULL)
}
\arguments{
\item{.mod}{S3 object of class `bbi_{.model_type}_model`}

\item{.out_path}{Character scaler with path to save out YAML file. By default, sets it to the model file name, with a yaml extension.}
}
\value{
Output list as specified above.
}
\description{
Saves a model model object to a yaml file
}
