% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan.R
\name{scaffold_missing_stan_files}
\alias{scaffold_missing_stan_files}
\title{Adds scaffolds of any missing stan files}
\usage{
scaffold_missing_stan_files(.mod)
}
\arguments{
\item{.mod}{a \code{bbi_stan_model}}
}
\description{
Adds scaffolds of any missing stan files
}
\keyword{internal}
