% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_model_directory}
\alias{set_model_directory}
\alias{get_model_directory}
\title{Get or set global model directory option}
\usage{
set_model_directory(.path)

get_model_directory()
}
\arguments{
\item{.path}{Path, either from working directory or absolute, that will be set as \code{options('rbabylon.model_directory')}}
}
\description{
Gets or sets \code{options('rbabylon.model_directory')},
which is used by default in functions like \code{read_model()}, \code{submit_model()} and \code{model_summary()} so that,
once this is set, those functions can take a path relative to this directory instead of the working/script directory.
}
\details{
\code{set_model_directory()} sets \code{options('rbabylon.model_directory')} to the absolute path of the directory passed to \code{.path}.
Note that the directory must exist or this will error.

\code{get_model_directory()} gets the path set to \code{options('rbabylon.model_directory')} and checks that it is both absolute and exists,
erroring if it is not.
}
