% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{setup_bootstrap_run}
\alias{setup_bootstrap_run}
\title{Set up a bootstrap model run}
\usage{
setup_bootstrap_run(
  .boot_run,
  n = 200,
  strat_cols = NULL,
  seed = 1234,
  data = NULL,
  .bbi_args = list(no_grd_file = TRUE, no_shk_file = TRUE),
  .overwrite = FALSE
)
}
\arguments{
\item{.boot_run}{A \code{bbi_nmboot_model} object.}

\item{n}{Number of data sets and model runs to generate.}

\item{strat_cols}{Columns to maintain proportion for stratification.}

\item{seed}{A numeric seed to set prior to resampling the data; use \code{NULL} to
avoid setting a seed.}

\item{data}{A dataset to resample from. Defaults to \code{NULL}, which will use
the \emph{filtered} output from \code{nm_data(.boot_run, filter = TRUE)}. If provided,
must include the same column names as what's returned from \code{nm_data(.mod)}.}

\item{.bbi_args}{Named list passed to \code{model_summary(orig_mod, .bbi_args)},
where \code{orig_mod} is the model the analysis is based on. See
\code{\link[=print_bbi_args]{print_bbi_args()}} for valid options. Defaults to \code{list(no_grd_file = TRUE, no_shk_file = TRUE)} because \code{\link[=model_summary]{model_summary()}} is only called internally to
extract the number of records, so those files are irrelevant. Only used if
the based on model (the model the analysis is being performed on) has been
executed.}

\item{.overwrite}{Logical (T/F) indicating whether or not to overwrite
existing setup.}
}
\description{
This function takes a \code{bbi_nmboot_model} (created by a previous
\code{\link[=new_bootstrap_run]{new_bootstrap_run()}} call) and creates \code{n} new model objects and re-sampled
datasets in a subdirectory. The control stream found at
\code{get_model_path(.boot_run)} is used as the "template" for these new model
objects, and the new datasets are sampled from the dataset defined in its
\verb{$DATA} record (i.e. \code{get_data_path(.boot_run)}).
}
\details{
Once you have run this function, you can execute your bootstrap with
\code{\link[=submit_model]{submit_model()}}. You can use \code{\link[=get_model_status]{get_model_status()}} to check on your submitted
bootstrap run. Once all models have finished, use \code{\link[=summarize_bootstrap_run]{summarize_bootstrap_run()}}
to view the results. See examples below.
}
\examples{
\dontrun{

# Setup
.boot_run <- new_bootstrap_run(.mod)
.boot_run <- setup_bootstrap_run(
  .boot_run,
  n = 1000,
  seed = 1234,
  strat_cols = c("STUDY", "ETN")
)

# Submit
submit_model(.boot_run)

# Check status of runs during submission
get_model_status(.boot_run)

# Summarize results, once all runs have finished
if (check_nonmem_finished(.boot_run)) {
  .boot_sum <- summarize_bootstrap_run(.boot_run)
}
}
}
\seealso{
\code{\link[=new_bootstrap_run]{new_bootstrap_run()}} \code{\link[=summarize_bootstrap_run]{summarize_bootstrap_run()}} \code{\link[=submit_model]{submit_model()}}
}
