% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{setup_bootstrap_run}
\alias{setup_bootstrap_run}
\title{Set up a bootstrap model run.}
\usage{
setup_bootstrap_run(
  .boot_run,
  n = 200,
  strat_cols = NULL,
  seed = 1234,
  .overwrite = FALSE
)
}
\arguments{
\item{.boot_run}{a \code{bbi_nmboot_model} object.}

\item{n}{number of model runs.}

\item{strat_cols}{columns to maintain proportion for stratification}

\item{seed}{a seed for sampling the data. Set to \code{NULL} to avoid setting.}

\item{.overwrite}{logical (T/F) indicating whether or not to overwrite existing
setup for a bootstrap run.}
}
\description{
Creates a new model object and re-sampled dataset per model run.
}
\examples{
\dontrun{

# Setup
.boot_run <- new_bootstrap_run(.mod)
.boot_run <- setup_bootstrap_run(
  .boot_run,
  n = 1000,
  seed = 1234,
  strat_cols = c("STUDY", "ETN")
)

# Submit
submit_model(.boot_run)

# Check status of runs during submission
get_model_status(.boot_run)
}
}
\seealso{
\code{\link[=new_bootstrap_run]{new_bootstrap_run()}} \code{\link[=summarize_bootstrap_run]{summarize_bootstrap_run()}}
}
