% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{setup_bootstrap_run}
\alias{setup_bootstrap_run}
\title{Set up a bootstrap model run}
\usage{
setup_bootstrap_run(
  .boot_run,
  n = 200,
  strat_cols = NULL,
  seed = 1234,
  data = NULL,
  .join_col = "NUM",
  .overwrite = FALSE
)
}
\arguments{
\item{.boot_run}{a \code{bbi_nmboot_model} object.}

\item{n}{number of model runs.}

\item{strat_cols}{columns to maintain proportion for stratification}

\item{seed}{a seed for sampling the data. Set to \code{NULL} to avoid setting.}

\item{data}{A dataset to resample from. Defaults to \code{NULL}, which will use
the output from \code{nm_join(.mod)}. If provided, must include the same column
names as what's returned from \code{nm_data(.mod)}. If the default is used, note
that a suitable \code{.join_col} must be provided.}

\item{.join_col}{Character column name to use to join table files. Not used
if \code{data} is specified. Passed to \code{\link[=nm_join]{nm_join()}}, and used to create the
initial dataset that gets re-sampled \code{n} times. The purpose of joining the
input data to table files is to filter the population to only the subjects
that actually made it into the model. See the \code{Details} section in
\code{\link[=nm_join]{nm_join()}} for more information.}

\item{.overwrite}{logical (T/F) indicating whether or not to overwrite
existing setup for a bootstrap run.}
}
\description{
This function takes a \code{bbi_nmboot_model} (created by a previous
\code{\link[=new_bootstrap_run]{new_bootstrap_run()}} call) and creates \code{n} new model objects and re-sampled
datasets in a subdirectory. The control stream found at
\code{get_model_path(.boot_run)} is used as the "template" for these new model
objects, and the new datasets are sampled from the dataset defined in its
\verb{$DATA} record (i.e. \code{get_data_path(.boot_run)}).
}
\details{
Once you have run this function, you can execute your bootstrap with
\code{\link[=submit_model]{submit_model()}}. You can use \code{\link[=get_model_status]{get_model_status()}} to check on your submitted
bootstrap run. Once all models have finished, use \code{\link[=summarize_bootstrap_run]{summarize_bootstrap_run()}}
to view the results. See examples below.
}
\examples{
\dontrun{

# Setup
.boot_run <- new_bootstrap_run(.mod)
.boot_run <- setup_bootstrap_run(
  .boot_run,
  n = 1000,
  seed = 1234,
  strat_cols = c("STUDY", "ETN")
)

# Submit
submit_model(.boot_run)

# Check status of runs during submission
get_model_status(.boot_run)

# Summarize results, once all runs have finished
if (check_nonmem_finished(.boot_run)) {
  .boot_sum <- summarize_bootstrap_run(.boot_run)
}
}
}
\seealso{
\code{\link[=new_bootstrap_run]{new_bootstrap_run()}} \code{\link[=summarize_bootstrap_run]{summarize_bootstrap_run()}} \code{\link[=submit_model]{submit_model()}}
}
