% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inherit-param-estimates.R
\name{setup_param_records}
\alias{setup_param_records}
\title{Set up parameter records and replacement values}
\usage{
setup_param_records(
  .inherit_mod_lines,
  .new_params,
  .rec_type = BBR_ESTIMATES_INHERIT
)
}
\arguments{
\item{.inherit_mod_lines}{model lines as returned by \code{nmrec::read_ctl()}}

\item{.new_params}{Either a list for multiple replacements, or one of the following:
\describe{
\item{\code{.rec_type = 'theta'}}{a \code{vector} of replacement values}
\item{\code{.rec_type = 'sigma'} or \code{.rec_type = 'omega'}}{a \code{matrix} of replacement values}
}}

\item{.rec_type}{Record type. One of \code{c("theta", "sigma", "omega")}}
}
\description{
Extract records of a given type and filter out prior blocks. Replacement values
and extracted records are formatted to be lists of equal length
}
\keyword{internal}
