% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{setup_sim_run}
\alias{setup_sim_run}
\title{Set up \code{NONMEM} simulation}
\usage{
setup_sim_run(.mod, n = 100, seed = 1234, .join_col = c("NUM"))
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} or \code{bbi_nmsim_model} object.}

\item{n}{number of simulations/subproblems. Adds \code{SUBPROBLEMS=n} to a
\verb{$SIMULATION} record.}

\item{seed}{a seed for simulation. Appended to \verb{$SIMULATION} record.}

\item{.join_col}{Character column name(s) used to join table files post
execution. Gets appended to the generated \verb{$TABLE} record. See
\code{\link[=nm_join_sim]{nm_join_sim()}} documentation for details. Defaults to \code{'NUM'}.}
}
\description{
Creates new \verb{$SIMULATION} and \verb{$MSFI} records, and modifies the \verb{$PRED} or
\verb{$ERROR} record (whichever exists) to have the additional line: \code{REPI=IREP}.
}
\keyword{internal}
