% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc-simulation.R
\name{setup_sim_run}
\alias{setup_sim_run}
\title{Set up \code{NONMEM} simulation}
\usage{
setup_sim_run(
  .mod,
  n = 100,
  seed = 1234,
  sim_cols = c("DV", "PRED"),
  .join_col = c("NUM")
)
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} or \code{bbi_nmsim_model} object.}

\item{n}{Number of simulations/subproblems. Adds \code{SUBPROBLEMS=n} to a
\verb{$SIMULATION} record.}

\item{seed}{A seed for simulation. Appended to \verb{$SIMULATION} record.}

\item{sim_cols}{Character column name(s) defining the simulated values to
table out.}

\item{.join_col}{Character column name(s) used to join table files post
execution. Gets appended to the generated \verb{$TABLE} record. See
\code{\link[=nm_join_sim]{nm_join_sim()}} documentation for details. Defaults to \code{'NUM'}.}
}
\description{
Removes relevant record types, creates new \verb{$SIMULATION} and \verb{$MSFI} records,
and creates a new \verb{$TABLE} record to save out simulated values (\code{sim_cols}).
}
\keyword{internal}
