% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-for-tests.R
\name{skip_if_old_bbi}
\alias{skip_if_old_bbi}
\title{Skip test if current bbi version is below specified one.}
\usage{
skip_if_old_bbi(v)
}
\arguments{
\item{v}{a package version (or a string that can be converted to one) to pass
as the .min_version argument of \code{test_bbi_version()}.}
}
\description{
Skip test if current bbi version is below specified one.
}
\keyword{internal}
