% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{submit_model}
\alias{submit_model}
\alias{submit_model.bbi_nonmem_model}
\alias{submit_model.bbi_stan_model}
\title{Submit a model to be run}
\usage{
submit_model(
  .mod,
  .bbi_args = NULL,
  .mode = c("sge", "local"),
  ...,
  .overwrite = NULL,
  .config_path = NULL,
  .wait = TRUE,
  .dry_run = FALSE
)

\method{submit_model}{bbi_nonmem_model}(
  .mod,
  .bbi_args = NULL,
  .mode = c("sge", "local"),
  ...,
  .overwrite = NULL,
  .config_path = NULL,
  .wait = TRUE,
  .dry_run = FALSE
)

\method{submit_model}{bbi_stan_model}(
  .mod,
  .bbi_args = NULL,
  .mode = c("local"),
  ...,
  .overwrite = NULL,
  .config_path = NULL,
  .wait = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{.mod}{The model object to submit.}

\item{.bbi_args}{A named list specifying arguments to pass to bbi
formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}. Run \code{\link[=print_bbi_args]{print_bbi_args()}} to see valid arguments. Note that bbr does
not support changing the output directory (including through the model or
global YAML files).}

\item{.mode}{Either \code{"sge"}, the default, to submit model(s) to the grid or
\code{"local"} for local execution.}

\item{...}{args passed through. For \code{bbi_stan_model} this is how you pass
arguments through to the \verb{$sample()} method of your \code{cmdstanr} model
object. See \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}} for valid arguments and details. For
\code{bbi_nonmem_model} these are passed to \code{bbi_exec()}.}

\item{.overwrite}{Controls whether or not to overwrite existing model output from
a previous run. If \code{NULL}, the default, will defer to setting in \code{.bbi_args}.}

\item{.config_path}{Path to a bbi configuration file. If \code{NULL}, the
default, will attempt to use a \code{bbi.yaml} in the same directory as the
model.}

\item{.wait}{If \code{TRUE}, the default, wait for the bbi process to return
before this function call returns. If \code{FALSE} function will return while
bbi process runs in the background.}

\item{.dry_run}{Returns an object detailing the command that would be run,
insted of running it. This is primarily for testing but also a debugging
tool.}
}
\description{
Submits a model to be run by calling out to \code{bbi}.
}
\section{Methods (by class)}{
\itemize{
\item \code{bbi_nonmem_model}: Takes a \code{bbi_nonmem_model} object.

\item \code{bbi_stan_model}: Takes a \code{bbi_stan_model} object. All arguments
passed through \code{...} will be passed to \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}} method.
}}

\seealso{
\code{\link[=submit_models]{submit_models()}}
}
