% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{submit_model}
\alias{submit_model}
\alias{submit_model.bbi_nonmem_spec}
\alias{submit_model.character}
\alias{submit_nonmem_model}
\title{S3 generic method for submit_model}
\usage{
submit_model(.spec, ...)

\method{submit_model}{bbi_nonmem_spec}(.spec, ...)

\method{submit_model}{character}(.path, ...)

submit_nonmem_model(
  .spec,
  .mode = c("sge", "local"),
  .bbi_args = NULL,
  ...,
  .config_path = NULL,
  .wait = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{.spec}{An S3 object of class `bbi_nonmem_spec`, for example from `create_model()` or `copy_model_from()`, etc.}

\item{...}{args passed through to `bbi_exec()`}

\item{.path}{Path to YAML or model file}

\item{.mode}{Either "local" for local execution or "sge" to submit model(s) to the grid}

\item{.bbi_args}{A named list specifying arguments to pass to babylon formatted like `list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)`. Run `print_nonmem_args()` to see valid arguments.}

\item{.config_path}{Optionally specify a path to a babylon.yml config. If not specified, the config in the model directory will be used by default. Path MUST be either an absolute path or relative to the model directory.}

\item{.wait}{Boolean for whether to wait for the bbi process to return before this function call returns.}

\item{.dry_run}{Returns an object detailing the command that would be run, insted of running it. This is primarily for testing but also a debugging tool.}
}
\value{
An S3 object of class `bbi_nonmem_result` (or `bbi_nonmem_spec` if .dry_run=T)
}
\description{
S3 generic method for submit_model

S3 dispatch for submit_model from bbi_nonmem_spec

S3 dispatch for submit_model from character scaler
  Should be path to yaml (with or without .yaml extension), or a valid model file (control stream, etc.).

Submit a NONMEM model via babylon
}
