% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{submit_model}
\alias{submit_model}
\alias{submit_nonmem_model}
\title{submit model}
\usage{
submit_model(.spec, ...)

submit_nonmem_model(
  .path,
  .mode = c("sge", "local"),
  .args = NULL,
  ...,
  .config_path = NULL,
  .wait = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{.spec}{spec}

\item{...}{args passed through to `bbi_exec()`}

\item{.path}{Full path to one of the following:
1) path to a model file(s) that should be run OR
2) path to a model YAML file that must contain:
    a) a path to a model file with the key `model_path`,
    b) any NONMEM args that would like to pass through, with keys that match the arguments in `print_nonmem_args()`,
    c) any extra user data that will be passed through to `bbi_config.json` with whatever keys are specified in the yaml.
(Path must be either an absolute path or relative to the R working directory)}

\item{.mode}{Either "local" for local execution or "sge" to submit model(s) to the grid}

\item{.args}{A named list specifying arguments to pass to babylon formatted like `list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)`. Run `print_nonmem_args()` to see valid arguments.}

\item{.config_path}{Optionally specify a path to a babylon.yml config. If not specified, the config in the model directory will be used by default. Path MUST be either an absolute path or relative to the model directory.}

\item{.wait}{Boolean for whether to wait for the bbi process to return before this function call returns.}

\item{.dry_run}{Returns character string of command that would be run, insted of running it. This is primarily for testing but also a debugging tool.}
}
\value{
output from the model run (?)
}
\description{
submit model

Submit a NONMEM model via babylon
}
