% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-models.R
\name{submit_models}
\alias{submit_models}
\alias{submit_models.list}
\title{Submit models to be run in batch}
\usage{
submit_models(
  .mods,
  .bbi_args = NULL,
  .mode = c("sge", "local"),
  ...,
  .config_path = NULL,
  .wait = TRUE,
  .dry_run = FALSE
)

\method{submit_models}{list}(
  .mods,
  .bbi_args = NULL,
  .mode = c("sge", "local"),
  ...,
  .config_path = NULL,
  .wait = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{.mods}{The model objects to submit.}

\item{.bbi_args}{A named list specifying arguments to pass to babylon formatted like \code{list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)}. Run \code{\link[=print_bbi_args]{print_bbi_args()}} to see valid arguments.}

\item{.mode}{Either \code{"sge"}, the default, to submit model(s) to the grid or \code{"local"} for local execution.}

\item{...}{args passed through to \code{bbi_exec()}}

\item{.config_path}{Path to a babylon configuration file. If \code{NULL}, will
attempt to use a \code{babylon.yaml} in the same directory as the model.}

\item{.wait}{If \code{TRUE}, the default, wait for the bbi process to return before this function call returns. If \code{FALSE} function will return while bbi process runs in the background.}

\item{.dry_run}{Returns an object detailing the command that would be run, insted of running it. This is primarily for testing but also a debugging tool.}
}
\description{
Submits a group of models to be run in batch by calling out to \code{bbi} in as
few external calls as possible (see "Details").
}
\details{
The number of \code{bbi} calls to make is determined by the number of
distinct sets of \code{bbi} arguments passed to the submission calls, either
explicitly through \code{.bbi_args}, as specified in the \code{bbi_args} field of the
model YAML, or specified globally in \code{babylon.yaml}.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: Takes a list of \code{bbi_nonmem_model} objects.
}}

