% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{submit_models}
\alias{submit_models}
\alias{submit_models.list}
\alias{submit_models.character}
\alias{submit_models.numeric}
\alias{submit_nonmem_models}
\title{S3 generic method for submit_models}
\usage{
submit_models(
  .mods,
  .bbi_args = NULL,
  .mode = c("sge", "local"),
  ...,
  .config_path = file.path(get_model_directory() \%||\% ".", "babylon.yaml"),
  .wait = TRUE,
  .dry_run = FALSE,
  .directory = NULL
)

\method{submit_models}{list}(
  .mods,
  .bbi_args = NULL,
  .mode = c("sge", "local"),
  ...,
  .config_path = file.path(get_model_directory() \%||\% ".", "babylon.yaml"),
  .wait = TRUE,
  .dry_run = FALSE,
  .directory = NULL
)

\method{submit_models}{character}(
  .mods,
  .bbi_args = NULL,
  .mode = c("sge", "local"),
  ...,
  .config_path = file.path(get_model_directory() \%||\% ".", "babylon.yaml"),
  .wait = TRUE,
  .dry_run = FALSE,
  .directory = get_model_directory()
)

\method{submit_models}{numeric}(
  .mods,
  .bbi_args = NULL,
  .mode = c("sge", "local"),
  ...,
  .config_path = file.path(get_model_directory() \%||\% ".", "babylon.yaml"),
  .wait = TRUE,
  .dry_run = FALSE,
  .directory = get_model_directory()
)

submit_nonmem_models(
  .mods,
  .bbi_args = NULL,
  .mode = c("sge", "local"),
  ...,
  .config_path = file.path(get_model_directory() \%||\% ".", "babylon.yaml"),
  .wait = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{.mods}{A list of S3 objects of class `bbi_nonmem_model`}

\item{.bbi_args}{A named list specifying arguments to pass to babylon formatted like `list("nm_version" = "nm74gf_nmfe", "json" = T, "threads" = 4)`. Run `print_nonmem_args()` to see valid arguments.}

\item{.mode}{Either "local" for local execution or "sge" to submit model(s) to the grid}

\item{...}{args passed through to `bbi_exec()`}

\item{.config_path}{Optionally specify a path to a babylon.yml config. If not specified, the config in the model directory will be used by default. Path MUST be either an absolute path or relative to the model directory.}

\item{.wait}{Boolean for whether to wait for the bbi process to return before this function call returns.}

\item{.dry_run}{Returns an object detailing the command that would be run, insted of running it. This is primarily for testing but also a debugging tool.}

\item{.directory}{Model directory containing the files referenced by `.mods`. Defaults to `options('rbabylon.model_directory')`, which can be set globally with `set_model_directory()`.}
}
\value{
A list of S3 objects of class `babylon_process`
}
\description{
S3 generic method for submit_models

S3 dispatch for submit_model from bbi_nonmem_model

S3 dispatch for submit_models from character vector
  Should be a vector of paths to yaml (with or without .yaml extension), or a valid model file (control stream, etc.).

S3 dispatch for submit_models from numeric input
This will only work if you are calling from the same directory as the models, or if you have set the model directory with `set_model_directory()`

Submits multiple NONMEM models in batch via babylon
}
