% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{submit_nonmem_model}
\alias{submit_nonmem_model}
\title{Submit a NONMEM model via babylon}
\usage{
submit_nonmem_model(
  .mod,
  .bbi_args = NULL,
  .mode = c("sge", "local"),
  ...,
  .config_path = file.path(get_model_directory() \%||\% ".", "babylon.yaml"),
  .wait = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{.mod}{An S3 object of class `bbi_nonmem_model`, for example from `new_model()`, `read_model()` or `copy_model_from()`}
}
\value{
An S3 object of class `babylon_process`
}
\description{
Private implementation function called by `submit_model()` dispatches.
}
\keyword{internal}
