% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-nonmem-model.R
\name{submit_nonmem_model}
\alias{submit_nonmem_model}
\title{Submit a NONMEM model via babylon}
\usage{
submit_nonmem_model(
  .path,
  .type = c("sge", "local"),
  .args,
  ...,
  .config_path = NULL
)
}
\arguments{
\item{.path}{Full path to a model file(s) that should be run}

\item{.type}{Either "local" for local execution or "sge" to submit model(s) to the grid}

\item{.args}{A named list specifying arguments to pass to babylon. All available arguments is in `options("rbabylon.nonmen_args")`}

\item{...}{args passed through to `bbi_exec()`}

\item{.config_path}{Optionally specify a path to a babylon.yml config. If not specified, the config in the model directory will be used by default. (?)}
}
\value{
output from the model run (?)
}
\description{
Submit a NONMEM model via babylon
}
