% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit-model.R
\name{submit_stan_model_cmdstanr}
\alias{submit_stan_model_cmdstanr}
\title{Submit a Stan model via cmdstanr}
\usage{
submit_stan_model_cmdstanr(
  .mod,
  .mode = c("local"),
  ...,
  .overwrite = NULL,
  .wait = TRUE,
  .dry_run = FALSE
)
}
\arguments{
\item{.mod}{An S3 object of class \code{bbi_stan_model}, for example from \code{new_model()}, \code{read_model()} or \code{copy_model_from()}}
}
\value{
An S3 object of class \code{bbi_process} MAYBE???? Maybe a cmdstanr model. Will there be a wait=F option?
}
\description{
Private implementation function called by \code{submit_model()} dispatches.
}
\keyword{internal}
