% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{summarize_bootstrap}
\alias{summarize_bootstrap}
\alias{summarize_bootstrap_run}
\alias{bootstrap_estimates}
\title{Summarize a bootstrap run}
\usage{
summarize_bootstrap_run(.boot_run, force_resummarize = FALSE)

bootstrap_estimates(.boot_run, format_long = FALSE, force_resummarize = FALSE)
}
\arguments{
\item{.boot_run}{a \code{bbi_nmboot_model} object.}

\item{force_resummarize}{logical (T/F). If \code{TRUE}, force re-summarization.
See details.}

\item{format_long}{logical (T/F). If \code{TRUE}, format data as a long table,
making the data more portable for plotting.}
}
\description{
Summarize a bootstrap run
}
\details{
The first time \code{summarize_bootstrap_run()} is called, it will save the
results to a \code{boot_summary.RDS} data file within the bootstrap run directory.
If one already exists, that data set will be read in instead of being
re-summarized. The purpose of this is functionality two fold. For one, it
helps avoid the need of re-executing \code{model_summary()} calls for a large
number of runs. It also helps to reduce the number of files you need to
commit via version control. This can be overridden by setting
\code{force_resummarize = TRUE}.
}
\section{Functions}{
\itemize{
\item \code{summarize_bootstrap_run()}: Summarize a bootstrap run and store results

\item \code{bootstrap_estimates()}: Tabulate parameter estimates for each model submission in a bootstrap run

}}
