% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-model.R
\name{summarize_bootstrap_run}
\alias{summarize_bootstrap_run}
\title{Summarize a bootstrap run}
\usage{
summarize_bootstrap_run(
  .boot_run,
  add_summary = FALSE,
  force_resummarize = FALSE
)
}
\arguments{
\item{.boot_run}{a \code{bbi_nmboot_model} object.}

\item{add_summary}{logical (T/F). Set to \code{TRUE} to include additional
model summary parameters. If \code{FALSE}, only includes the parameter estimates
(much faster for a large number of runs).}
}
\description{
Summarize a bootstrap run
}
