% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-log.R
\name{summary_log}
\alias{summary_log}
\alias{add_summary}
\title{Create log of model summary outputs}
\usage{
summary_log(.models, ...)

add_summary(.log_df, ...)
}
\arguments{
\item{.models}{A character vector of paths to models. Can be absolute or relative to working directory, but will come out as absolute. All must exist.}

\item{...}{Arguments passed through to model_summary()}
}
\description{
This should probably be an implementation function for an S3 method that can take a list of model objects or paths etc.
Or maybe just the character (default?) dispatch? Need to think on that.
}
