% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-log.R
\name{summary_log}
\alias{summary_log}
\alias{add_summary}
\title{Create a tibble from \code{model_summaries()} data}
\usage{
summary_log(.base_dir = get_model_directory(), .recurse = TRUE, ...)

add_summary(.log_df, ...)
}
\arguments{
\item{.base_dir}{Base directory to look in for models. Defaults to \code{\link[=get_model_directory]{get_model_directory()}}.}

\item{.recurse}{If \code{TRUE}, the default, search recursively in all subdirectories. Passed through to \code{fs::dir_ls()} -- If a positive number, the number of levels to recurse.}

\item{...}{Arguments passed through to \code{\link[=model_summaries]{model_summaries()}}}

\item{.log_df}{a \code{bbi_run_log_df} tibble (the output of \code{\link[=run_log]{run_log()}})}
}
\value{
\code{summary_log()} will return a new tibble with the \code{'absolute_model_path'} column as the primary key,
and all the columns extracted from \code{\link[=model_summaries]{model_summaries()}} (but none of the other columns from the input tibble).

\code{add_summary()} takes a \code{bbi_run_log_df} tibble (the output from \code{\link[=run_log]{run_log()}}) and returns the input tibble,
with all the columns extracted from \code{\link[=model_summaries]{model_summaries()}} joined onto it.
}
\description{
Runs \code{\link[=model_summaries]{model_summaries()}} on all models in the input and returns a subset of the each resulting summary as a tibble.
}
\details{
The following fields from \code{bbi_nonmem_summary} are extracted and included by default.
If you would like more fields from the summary object, you can extract them manually from the \code{bbi_summary} list column.
\itemize{
\item \code{error_msg} -- Error message from \code{\link[=model_summary]{model_summary()}}. If \code{NULL} the call succeeded. If not \code{NULL}, the rest of the fields will be \code{NULL}.
\item \code{needed_fail_flags} -- Logical for whether the call initially failed, but passed with the inclusion of \code{.fail_flags}
\item \code{bbi_summary} -- The full \code{bbi_nonmem_summary} object for each row. This can be queried further by extracting it as a list, or by using \code{dplyr::mutate()} etc.
\item \code{ofv} -- Objective function value from last estimation method, using \verb{$ofv$ofv_no_constant} field.
\item \code{param_count} -- Count of (non-fixed) parameters estimated in final estimation method.
\item \code{estimation_method} -- Character vector of estimation method(s) used. Extracted from \verb{$run_details}
\item \code{problem_text} -- Character vector of text from \verb{$PROB}. Extracted from \verb{$run_details}
\item \code{number_of_patients} -- Extracted from \verb{$run_details}
\item \code{number_of_obs} -- Extracted from \verb{$run_details}
\item \code{any_heuristics} -- Logical indicating whether \emph{any} of the columns extracted from \verb{$run_heuristics} are \code{TRUE}. Duplicative information, but helpful for filtering.
\item \code{covariance_step_aborted} -- Extracted from \verb{$run_heuristics}
\item \code{large_condition_number} -- Extracted from \verb{$run_heuristics}
\item \code{correlations_not_ok} -- Extracted from \verb{$run_heuristics}
\item \code{parameter_near_boundary} -- Extracted from \verb{$run_heuristics}
\item \code{hessian_reset} -- Extracted from \verb{$run_heuristics}
\item \code{has_final_zero_gradient} -- Extracted from \verb{$run_heuristics}
\item \code{minimization_terminated} -- Extracted from \verb{$run_heuristics}
}
}
\seealso{
\code{\link[=run_log]{run_log()}}
}
