% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-threads.R
\name{test_threads}
\alias{test_threads}
\title{Takes a model object and runs it with various threads values}
\usage{
test_threads(
  .mod,
  threads = c(2, 4),
  .mode = getOption("bbr.bbi_exe_mode"),
  .bbi_args = list()
)
}
\arguments{
\item{.mod}{bbi_model object to copy/test.}

\item{threads}{Integer vector of threads values to test.}

\item{.mode}{Passed through to bbr::submit_models(.mode).}

\item{.bbi_args}{a named list.}
}
\value{
A list of the model objects for the submitted models.
}
\description{
Takes a model object and runs it with various threads values
}
