% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-threads.R
\name{test_threads}
\alias{test_threads}
\title{Takes a model object and runs it with various threads values}
\usage{
test_threads(
  .mod,
  .threads = c(2, 4),
  .mode = getOption("bbr.bbi_exe_mode"),
  .bbi_args = list(),
  .max_eval = 10,
  ...
)
}
\arguments{
\item{.mod}{bbi_model object to copy/test.}

\item{.threads}{Integer vector of threads values to test.}

\item{.mode}{Passed through to bbr::submit_models(.mode).}

\item{.bbi_args}{a named list.}

\item{.max_eval}{Max number of iterations for NONMEM to run.
Will update \code{MAXITER} or \code{NITER} (whichever is specified) in generated models.}

\item{...}{args passed through to submit_models()}
}
\value{
A list of the model objects for the submitted models.
}
\description{
Takes a model object and runs it with various threads values
}
\examples{
\dontrun{

mod <- read_model(file.path(MODEL_DIR, 1))


mods <- test_threads(.mod = mod, .threads = c(2, 4), .max_eval = 10,
                     .mode = "local")


check_run_times(mods, .wait = TRUE, .time_limit = 100)
check_run_times(mods, .wait = FALSE, .return_times = "all")

cleanup_mods(.mods = mods)


# Dry Run:
mods <- test_threads(.mod = mod, .threads = c(2, 4), .max_eval = 10,
                     .mode = "local", .dry_run = TRUE)

cleanup_mods(.mods = mods)
}

}
