% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-tran-data.R
\name{translate_nm_expr}
\alias{translate_nm_expr}
\title{Translate \code{IGNORE} and \code{ACCEPT} filter expressions into \code{\link[dplyr:filter]{dplyr::filter()}}
expressions.}
\usage{
translate_nm_expr(nm_expr, type = c("ignore", "accept"), data_cols = NULL)
}
\arguments{
\item{nm_expr}{a \code{NONMEM} filter expression. e.g., \code{IGNORE(ID.EQ.2, BLQ=1)}.}

\item{type}{Either \code{'ignore'} or \code{'accept'}. Denotes which type of \code{NONMEM}
filtering the expression corresponds to.}

\item{data_cols}{column names associated with the input data. Used for
\code{'ignore'} expressions}
}
\description{
Translate \code{IGNORE} and \code{ACCEPT} filter expressions into \code{\link[dplyr:filter]{dplyr::filter()}}
expressions.
}
\keyword{internal}
