% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak-initial-estimates.R
\name{tweak_initial_estimates}
\alias{tweak_initial_estimates}
\title{Tweak the initial parameter estimates}
\usage{
tweak_initial_estimates(
  .mod,
  .p = 0.1,
  tweak = c("theta", "sigma", "omega"),
  digits = 3
)
}
\arguments{
\item{.mod}{model object to update.}

\item{.p}{Percent to tweak the initial parameter estimates by. Represented as
a fraction.}

\item{tweak}{type of estimates to tweak in the model. Defaults to
updating all of THETA, SIGMA, and OMEGA records.}

\item{digits}{Number of significant digits to round estimates to.}
}
\description{
Tweak the initial parameter estimates
}
\details{
In the following cases, the initial estimate will \emph{not} be updated:
\itemize{
\item \strong{Individual} \code{FIXED} \code{THETA} parameters
\itemize{
\item e.g., \verb{$THETA 1.2 FIX 1.5 0.2} --> would only skip the first value
}
\item \strong{Individual} \code{FIXED} \code{OMEGA} & \code{SIGMA} parameters for \emph{diagonal} matrices
\itemize{
\item e.g., \verb{$OMEGA 0 FIX 1} --> would only skip the first value
}
\item \strong{Full} \code{FIXED} \code{OMEGA} & \code{SIGMA} \emph{block} matrices
\itemize{
\item e.g., \verb{$OMEGA BLOCK(2) 0.1 0.001 0.1 FIX} --> would skip the full \code{OMEGA} record
}
\item \code{THETA} parameters with no initial estimate
\itemize{
\item e.g., \verb{$THETA (0,,1)}
}
}

For bounded \code{THETA} estimates:
\itemize{
\item If an initial \code{THETA} has bounds \strong{and} an initial estimate
(e.g., \verb{(0, 0.5, 1)}, \verb{(0,1)}), the bounds will be respected when sampling
a percent to tweak by. If the tweaked value would fall below the lower bound,
the initial estimate will be set to the lower bound. The same is true for
upper bounds.
\itemize{
\item e.g., \verb{(0, 0.5, 1)} --> tweak initially, falls outside bound (\verb{(0, 1.2, 1)})
--> set to upper bound (\verb{(0, 1, 1)})
}
}
}
\examples{
\dontrun{
base_mod <- read_model(file.path(MODEL_DIR, "1"))

mod2 <- copy_model_from(base_mod, "mod2") \%>\%
  tweak_initial_estimates(.p = 0.2)

# This function may be paired with `inherit_param_estimates()`:
mod2 <- copy_model_from(base_mod, "mod2") \%>\%
  inherit_param_estimates() \%>\% tweak_initial_estimates(.p = 0.2)

# If you want to set the seed for reproducible results:
mod2 <- withr::with_seed(1234, {
   tweak_initial_estimates(mod2, .p = 0.2, digits = 3)
})

# The utilized `.Random.seed` is appended to the model object as an attribute:
head(attr(mod2, "seed"))
}
}
