% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{use_bbi}
\alias{use_bbi}
\title{Installing bbi}
\usage{
use_bbi(.dir = "/data/apps")
}
\arguments{
\item{.dir}{directory to install bbi to on linux}
}
\value{
character
}
\description{
Identifies system running and return installation instructions
}
