% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-estimates.R
\name{using_old_priors}
\alias{using_old_priors}
\title{Check if \emph{non-informative} priors are being used}
\usage{
using_old_priors(ctl)
}
\arguments{
\item{ctl}{An \code{nmrec_ctl_records} object.}
}
\description{
Check if \emph{non-informative} priors are being used
}
\keyword{internal}
