% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-bbi.R
\name{version_message}
\alias{version_message}
\title{Private helper to construct version comparison message}
\usage{
version_message(local_v, release_v)
}
\arguments{
\item{local_v}{Character scaler for version number on local installation}

\item{release_v}{Character scaler for version number of current release}
}
\description{
Private helper to construct version comparison message
}
