% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wait_for_nonmem}
\alias{wait_for_nonmem}
\alias{wait_for_nonmem.bbi_nonmem_model}
\alias{wait_for_nonmem.list}
\title{Wait for NONMEM models to finish}
\usage{
wait_for_nonmem(.mod, .time_limit = 300, .interval = 5)

\method{wait_for_nonmem}{bbi_nonmem_model}(.mod, .time_limit = 300, .interval = 5)

\method{wait_for_nonmem}{list}(.mod, .time_limit = 300, .interval = 5)
}
\arguments{
\item{.mod}{a \code{bbi_nonmem_model} object, or list of \code{bbi_nonmem_model} objects.}

\item{.time_limit}{integer for maximum number of seconds in total to wait before continuing
(will exit after this time even if the run does not appear to have finished).}

\item{.interval}{integer for number of seconds to wait between each check.}
}
\description{
Calling \code{wait_for_nonmem()} will freeze the user's console until the model(s) have finished running.
}
\section{Methods (by class)}{
\itemize{
\item \code{bbi_nonmem_model}: takes a \code{bbi_nonmem_model} object.

\item \code{list}: takes a \code{list} of \code{bbi_nonmem_model} objects.
}}

